Bug 1687433 - Remove -no-integrated-as flag from libaom and libvpx. r=firefox-build-system-reviewers,mhentges
Back when it was added, it was added to libvpx, and the clang base
version was older. Either newer version of clang fixed the issues or
the libvpx source fixed it, but we don't need the flag anymore for
libvpx.

The flag for libaom was cargo culted from libvpx, but may not have been
necessary in the first place. Or maybe it was, but things fixed
themselves in a similar manner. Either way, the flag is not necessary.

Differential Revision: https://phabricator.services.mozilla.com/D102262

UltraBlame original commit: 729bf6049b7d872a951a9460ecf4723b29c37da3