Bug 1686219 - Store chunk information at the start rather than the end of the chunk r=sfink
This moves the chunk metadata to the start of the chunk and defines the data structures in the public header.  This simplifies accessing this data and removes the need for hardcoded offsets.

Requesting review from jandem for JIT updates.

Differential Revision: https://phabricator.services.mozilla.com/D101778

UltraBlame original commit: 43978ebfc7ace2c37aba0b7644ac4e866e0730bd