Bug 1687659 - jit_test.py: repeat_copy is now aware of generators. r=jonco
When using the `options.repeat`, we consume the `job_list` generator, which
causes it to be empty on the next iteration. This patch forces the content of
the generator to be stored in a list before iterating over it multiple times.

Differential Revision: https://phabricator.services.mozilla.com/D102410

UltraBlame original commit: d38c704c37caa16562a6d66a2c08eb6304a6efee