Bug 1684821 - Fix asserts in DebugEnvironmentProxy::initSnapshot. r=jorendorff
When leaving a frame, we may need to snapshot frame values that debugger is
observing. At this time, the environment chain may still have the body
LexicalEnvironmentObject before the CallObject. This was tripping an assert
in `initSnapshot` that ensured if a snapshot already existed that the frame
was for a generator/async that may have already paused.

Differential Revision: https://phabricator.services.mozilla.com/D102636

UltraBlame original commit: 55d055106082f1174740f8d5f4197febd341c95c