Bug 1687948 - [devtools] Add test for fullpage screenshot of website with fixed element. r=jdescottes.
A fix was made in Bug 961832 to ensure fixed elements would be rendered at the
right position in the screenshot when the page was scrolled down.
Unfortunately there wasn't any test associated, so this patch adds one.
A new helper function is added into shared-head so we can assert the color of
an image at a given position.

The test does fails if we revert the changes made in Bug 961832.

Differential Revision: https://phabricator.services.mozilla.com/D102595

UltraBlame original commit: ec2c8f28b2834a6f80ced89914fcc284aaabb4be