Bug 1687657 - [devtools] Avoid attaching the thread actor in the context of the browser console. r=nchevobbe
Thread actor's options are shared across all the toolboxes,
so that when you set any type of breakpoints, they will be passed to all the thread actor.
All the thread actor, of all the toolboxes.
So that when you set a breakpoint in the content toolbox,
they may be set in the browser toolbox and also in the browser console.

Here, we avoid doing anything around the thread actor when we are in the context
of the browser console, so that we avoid setting up any type of breakpoint.
They are set when calling ThreadActor.attach().

Differential Revision: https://phabricator.services.mozilla.com/D102907

UltraBlame original commit: 92251c943b322b18f94d04f2abaa0e38dd4948b4