Bug 1687936 - Fix uminv/umaxv/sminv/smaxv; fix config. r=nbp
Fix the vector min/max reductions to not overwrite the source before writing the
destination, if src == dest.

Fix moz.configure to enable wasm simd by default even if we're running on the
simulator, this was a bug that has prevented us from testing cranelift
(and the simulator, as this bug shows) as well as we should, up until now.

Differential Revision: https://phabricator.services.mozilla.com/D102594

UltraBlame original commit: 229e636bcc052e909476d42e75d98f85d01572bc