Bug 1599917 - [devtools] Retrieve object properties for debugger inline preview r=nchevobbe
`generateInlinePreview` had a typo (`actorID` -> `actor`) that prevented object properties from being loaded,
so on a given line, we would show the whole object in the inline preview instead of only the property being used
(e.g. if line is `const y = x.a`, we want the preview only for `x.a`, not for the whole `x` object).

A test case is added to ensure we don't regress this in the future, and object inspector jest tests are updated
so they don't throw.

Differential Revision: https://phabricator.services.mozilla.com/D100064

UltraBlame original commit: d2ae2ee5024f72b2d38806003d3f1280a80bfa07