Bug 1687441 part 15 - MacroAssembler changes for loading buffer lengths/offsets. r=anba
The variants that load an int32 value now take an optional Label that's used if
the index doesn't fit in an int32.

The Label argument is optional at this point so that we don't have to touch all
the callers, but later patches will change that.

Depends on D102731

Differential Revision: https://phabricator.services.mozilla.com/D102732

UltraBlame original commit: 7a28b297920d922b8cb67ee97e05ee5d9b5c100b