Bug 1560824 - Move combobox select padding to the select rule rather than the combobox frame. r=dholbert
This doesn't change behavior by default but allows authors to remove the
padding if they wish to.

I thought this was going to be problematic because of the windows
arrowbutton, but that doesn't respect select padding so we're good.

Differential Revision: https://phabricator.services.mozilla.com/D103245

UltraBlame original commit: f2539c891f28af2421b15995c53aba46f3c2157c