Bug 1689933 - force commit_span inline. r=jrmuizel
commit_span is an extremely hot and extremely short function that seems to show
up in many profiles. For some reason the optimizer doesn't choose to inline it,
when the cost of a function call is significant relative to the small amount of
critical work the function does. Force this inline so that we don't waste perf
calling this as a function.

Differential Revision: https://phabricator.services.mozilla.com/D103606

UltraBlame original commit: ce5e2f5f521e8145c31359c3c688913e4836edb9