Bug 1688033: Allow DCE of MCreateArgumentsObject r=jandem
The goal of scalar replacement of arguments is to remove all uses of the CreateArgumentsObject instruction, so that DCE can remove it.

However, the MCreateArgumentsObject constructor calls setGuard, which stops DCE from removing it. (This has been true since it was introduced.) As far as I can tell, it doesn't have to be a guard. In FinishBailoutToBaseline, if a frame needs an args object and doesn't have one, then we call ArgumentsObject::createExpected to create it from the restored baseline frame.

Depends on D103108

Differential Revision: https://phabricator.services.mozilla.com/D103109

UltraBlame original commit: ad86d5f1d5c7b5d120d13ee38da0a9f8fa3f6a9d