Bug 1599606 - Remove an assert in non-WebRender code. r=jnicol
This assert fails on the new reftest from this bug, on Android.

I haven't dug into it too deeply, given that it's code that'll be going away at
some point soon (hopefully), but I think what happens is that we have a fixed
layer which is not annotated as fixed. That's normal for background-attachment:
fixed root backgrounds. We handle explicitly-annotated-as-fixed layers a bit
further up in this function.

Returning false here seems like a very safe thing to do.

Depends on D54855

Differential Revision: https://phabricator.services.mozilla.com/D103736

UltraBlame original commit: b56f1be360db78e260fe47b3977a27d5ff33d822