Bug 1690099 - Address some flakiness in the test.
Using the same url as the following link is a slippery slope because
the test doesn't really wait for the load to finish, so some of the
targets for iframes and such seem to never run and the test times out
very intermittently.

I've confirmed this is an issue unrelated to the changes in the patch.

MANUAL PUSH: Unreviewed one-liner test fix.

UltraBlame original commit: 5ff587e0026f2929319cd35e4bed9c76ff2db986