Bug 1690433. Update display port expiry to handle minimal display ports. r=botond
If the activate_all_scroll_frames pref is on we want to never remove a minimal display port, and when a regular display port expires we want to downgrade it to minimal.

If the pref is off it's the same as before except we want to also remove the minimal property when on expiry.

Differential Revision: https://phabricator.services.mozilla.com/D103859

UltraBlame original commit: 11bf30988746e99e05a975170679e9a070450aa3