Bug 1687183 - Fix dark background detection in nsNativeTheme when only root has a background. r=mstange
If we hit the canvas frame, and the root has a background but the body
doesn't, previously we were using the body style. Then
FindNonTransparentBackgroundFrame would return the document element
frame (with the right background), but FindBackgroundFor that failed.

Differential Revision: https://phabricator.services.mozilla.com/D103975

UltraBlame original commit: 55ca84da624c687f92209dedcb94b89f40774bd8