Bug 1686220 - Invalidate paint on image load, not on size available. r=tnikkel
Invalidating paint on size available is too soon (we can't decode the
image anyways). Bullet frames didn't use to do this and just reflowed.

Scheduling a paint causes us to do layout way too soon when there's more
important stuff to do for the load of the page.

This is a big improvement on load times and visual metrics in a variety
of websites:

  https://treeherder.mozilla.org/perfherder/compare?originalProject=try&originalRevision=3e012dad47b7f4b22741faa46153e940d079a260&newProject=try&newRevision=ce14456c07318eb3fc8d94d9cff9f41d9d122752&framework=13

Differential Revision: https://phabricator.services.mozilla.com/D104065

UltraBlame original commit: 8591dc31ccda80ce19b4865b01d8e3060ba36f54