Bug 1690686 [wpt PR 27469] - Don't send SameSite=Lax cookies on cross-site, non-top-level requests, a=testonly
Automatic update from web-platform-tests
Don't send SameSite=Lax cookies on cross-site, non-top-level requests

This fixes a bug which was sending SameSite=Lax cookies on cross-site,
*non*-top-level requests when the site-for-cookies was same-site with
the request URL. If the request is not strictly same-site, we should
only be sending Lax cookies when the site-for-cookies is same-site with
the request URL, *and* the request is a top-level navigation. Similarly
for accepting cookies set on responses.

This implements the fix behind a flag (default enabled) to allow
reverting to the old behavior if there is too much site breakage as a
result of the fix.

Bug: 1166211
Change-Id: I2cebf8011010903cd016d7d7c1a32bf84aa325ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2653663
Reviewed-by: Andrey Kosyakov <caseqchromium.org>
Reviewed-by: Maksim Orlovich <morlovichchromium.org>
Commit-Queue: Lily Chen <chlilychromium.org>
Cr-Commit-Position: refs/heads/master{#851323}

--

wpt-commits: 4cd02119fd232ecc2a57d462d509c7ca299a3591
wpt-pr: 27469

UltraBlame original commit: 25afb6825263d61439bfb4a14aec5ff989b77247