Bug 1690775 - Fix code to parse the remaining header to actually finish consuming it. r=dragana,necko-reviewers
The code enters that function with a non-empty mLineBuf (mLineBuf =
"BODY"), and we just add a single "\n", so we never really end up
finishing the header section, and fetch code asserts and reads uninit
memory afterwards. Bad!

The intention of that code is to finish parsing the headers section, so
it should make sure to handle it properly if we have a partial line
already read. For that we need two newlines.

Differential Revision: https://phabricator.services.mozilla.com/D104773

UltraBlame original commit: 89ac26b3b7926fbb6ba92cee84c5869b9c99702f