Bug 1685313 - disable menus, commands and tabswitches while window-modal dialogs are up, r=jaws
This removes observation of the 'disabled' attribute from the macOS
full screen menu items, because removing the attribute doesn't work
correctly. This is a scenario that, as far as I can tell, didn't
happen elsewhere until now. On other OSes we use a single item which
gets disabled state directly from the command attribute (ie without
an 'observes' child) which appears to work fine.

It also exempts the editing commands from being disabled, but it
appears that at least on macOS, undo history is lost anyway. It's
not clear to me why this is the case, but I don't think it needs
to block an initial landing of this work.

Differential Revision: https://phabricator.services.mozilla.com/D103389

UltraBlame original commit: 319bc4d7bc0d7bbbc395f871370f5bab269ccc29