Bug 1685061 - [devtools] Localize new perf panel onboarding UI. r=julienw,flod.
We add a few localized string in perftools.ftl to replace to hardcoded labels
in the component.
This allows us to simplify the structure of the component.
We take this opportunity to re-use the localized string we have to link to the
advanced section of the settings panel.

Depends on D104649

Differential Revision: https://phabricator.services.mozilla.com/D104155

UltraBlame original commit: 57cb40f78ccc21ea58b190e26402307168cb3e79