Bug 1688033: Refactor PushElementsAsArguments r=jandem
Pushing the contents of the `args` array in an ArgumentsData (after we have guaranteed that there are no forwarded arguments) is almost the same as pushing the contents of an elements array, but with an additional offset. In preparation for the next patch, this patch refactors PushElementsAsArguments to take an offset, and changes names/comments accordingly.

Differential Revision: https://phabricator.services.mozilla.com/D104487

UltraBlame original commit: bf23844b260bff2066b2ac4d670f213942e65d61