Bug 1669051 - Turn off async font loading in APZ tests to see if they help avoiding intermittents. r=jfkthame
Per comment 9 this seems caused by an unexpected reframe. I haven't been
able to repro this, but the only kind of thing that should cause it is
the global reflow we do when fonts change.

This patch turns these async font loading features off in APZ tests to
see whether it helps avoiding this kind of intermittent. If it doesn't,
I guess we should revert this and try to repro harder.

Differential Revision: https://phabricator.services.mozilla.com/D105166

UltraBlame original commit: bcbf99a8a2dae152888e023599f1f80387b7bcc5