Bug 1693112 - Remove ReceiverGuard and replace iterator guard array with shape array. r=iain
ReceiverGuard is now only used in the iterator code. There every ReceiverGuard is
created for a native object, in which case ReceiverGuard's group is always nullptr.

This means we can just store the shape directly instead of using ReceiverGuard.

In ExtraStringCount, this removes the |* 2| because we no longer need two words
per guard/shape. This means iterators are now a bit more compact.

Also renames guardKey to shapesHash and uses HashNumber for this instead of uint32_t.

Differential Revision: https://phabricator.services.mozilla.com/D105344

UltraBlame original commit: e9658503446ef4005519700563f8e94117bea8b9