Bug 1691878. Make test_layerization.html wait for the wheel events it sends and then paints instead of waiting for scroll events and no paints. r=botond
The waitForScroll argument to promiseMoveMouseAndScrollWheelOver determines if we wait for a scroll event (if waitForScroll is true) or a wheel event (if waitForScroll is false). That might not be clear just from looking at it (where one might assume waitForScroll = false means no waiting at all).

In all cases but one in this test we wait for the scroll event.

When activateAllScrollFrames = false I guess this is enough to guarantee that any new metrics have been sent in a transaction. But it is not with activateAllScrollFrames = true for two reasons.

The first is that apz can handle the wheel event for any scroll frame immediately, it will send a repaint request to content which will scroll and send a scroll event. Content can then handle that scroll event without even sending it's current metrics in a transaction. So we need to wait for paints to be flushed.

But even that is not enough because we can wait for a paint but in several cases we need a minimal display port to be cleared to be full display port and that happens in PrepareForSetTargetAPZCNotification, which only happens after content gets the wheel event, which can happen after apz gets the wheel event, and after content gets the scroll event, and after content gets the scroll event and paints.

So we need to wait for the wheel event to rearch content and then wait for paints.

Differential Revision: https://phabricator.services.mozilla.com/D105407

UltraBlame original commit: bd7df680fbe422a65babdf360ed653310c9beabb