Bug 1693062: Disable float32 optimization for implicitly used definitions r=jandem
When the only consumer of `Math.sqrt(2**53)` was in code that had never been executed, the float32 optimization tried to convert it to float32 (in MSqrt::trySpecializeFloat32) because we didn't see any consumers that couldn't accept float32. When we bailed out, we recovered the result as a float instead of a double, and lost precision (94906264 vs 94906265.62425156). Checking for implicit uses solves the problem.

`bug1693062-01.js` tests `CheckUsesAreFloat32Consumers`. `bug1693062-02.js` tests `TypeAnalyzer::markPhiConsumers`. (The difference is whether the instruction we want to specialize is implicitly used itself, or whether it is used by an implicitly used phi.) These are the only places that `canConsumeFloat32` is called outside of assertions.

Differential Revision: https://phabricator.services.mozilla.com/D105568

UltraBlame original commit: 5f689d230b141c7cd0aba12f50b8c2196c582e11