Bug 1691578 - part4 : allow to set different error in all platforms, but still report them to corresponding platforms, except in testing. r=bryce
If some attributes are only able to be set in certain platform, that increase the difficulty of wrting a test.

Therefore, we make them being able to set in all platforms, but doing the check in the fianl stage, which is reporting the result, to ensure that those error would still only be displayed on the right platform.

But if we're in testing, then we can test all error types in one platform.

Differential Revision: https://phabricator.services.mozilla.com/D104474

UltraBlame original commit: f8a5b58b52cd4af508efb602228a84d636606c2a