Bug 1692116 Part 4 - Add WPT reftests, and adjust test expectations. r=dholbert
Part 2 fixed table-as-item-inflexible-in-column-{1,2}.html.
table-as-item-inflexible-in-row-{1,2}.html are already pass on Nightly,
and are added for robustness and symmetry.

Part 3 fixed table-as-item-specified-height.html

Part 2 and Part 3 are both needed to not break
layout/reftests/flexbox/flexbox-table-flex-items-3.html.

Differential Revision: https://phabricator.services.mozilla.com/D106158

UltraBlame original commit: 415a2c4c73b35d5766e48249e7e38b508224ea5d