Bug 1695679 - Simplify non-native theme scrollbar sizing on Windows to match the native theme. r=mstange
The native theme doesn't enforce the size of the whole scrollbar. This is
important because if the scrollbar overflows the scroller we might suppress it.

This causes a variety of test failures with the non-native theme enabled. One
of the most glaring ones is:

  http://wpt.live/css/css-flexbox/overflow-auto-006.html

We don't show scrollbars for the inline elements because their width is 50px
and the min scrollbar ends up being 51px (17 * 3) which is pretty unfortunate.

Instead, allow the thumb / scrollbar / track to shrink like the native theme.

Differential Revision: https://phabricator.services.mozilla.com/D106781

UltraBlame original commit: 09b73f8c2145a6dd7278d91e1649d8a933c93c40