Bug 1694318 - Remove the mozjs_sys rust crate. r=jandem
The mozjs_sys crate is actually responsible for all the problems
encountered with the js crate we just removed, but the sm-mozjs-crate
task that builds and runs tests for it doesn't fail because... there is
actually no test to build and run, so all it builds is a static library,
which doesn't expose all the problems that exist.

The content of this crate in mozilla-central is also outdated compared
to servo upstream.

Differential Revision: https://phabricator.services.mozilla.com/D106390

UltraBlame original commit: 4d19c36b6b323be4c1d83ed8a687ca55f5dd958f