Bug 1559448 - Move Thunderbird primary password prompt earlier in start-up. r=kaie,xpcom-reviewers,mccr8 DONTBUILD
Triggering the password prompt after the Add-Ons Manager start-up is causing the latter to fail.

I'm not sure if something needs to be unlocked by the password, or if the existence of the prompt is causing some bad interaction, but prompting slightly earlier appears to fix it.

Differential Revision: https://phabricator.services.mozilla.com/D106382

UltraBlame original commit: ed65070d30fb3e5abce3ad51bbc262c74eca91f1