Bug 1695835 - Build geckodriver using liblowercase. r=firefox-build-system-reviewers,andi,dmajor
Bug 1694775 had to get a fixup to install the rename utility, that is
not in the base Debian image for buster, while it was there for jessie.
However, we only use rename for a hack when cross-building geckodriver
for Windows, and we don't need to.

Differential Revision: https://phabricator.services.mozilla.com/D106875

UltraBlame original commit: b162b93f8429442f9e56e3651dddd20a0d1eef70