Bug 1695528 - Activate the bundled Twemoji Mozilla font by default on all Windows versions, for better character coverage than Segoe UI Emoji offers. r=lsalzman
This essentially reverts what we did in bug 1686274, because users are not happy
with just getting Segoe UI Emoji instead. However, the pref is left in place so
that we can easily experiment with disabling the font to continue testing the
impact on startup, rendering, etc.

Differential Revision: https://phabricator.services.mozilla.com/D107056

UltraBlame original commit: 45b9f8ff281817ec3f2ce5e0e4302bb29496d877