Bug 1691486 - Generalize update config JSON API so it will be usable with other prefs r=nalexander
Currently, we only support reading and writing app.update.auto to the update config JSON. We want to be able to support multiple installation-specific prefs.

This patch also adds additional features to this pref API that will be useful for the other prefs that we plan to add:
  - A default preference branch with behavior mimicking that of the existing preference system.
  - Multiple types of prefs including boolean, string, and integer.
  - Expandable: More prefs can be added with little effort.

This also maintains previous functionality. app.update.auto migration is maintained. This shouldn't strictly be necessary, since a watershed has occurred since migration. But it is possible for someone to use a really old profile with a new installation, so I thought it would be nice to maintain it for now.

Existing functionality for app.update.auto is now also made available to new prefs, including:
  - Ability to be overridden by policy.
  - Observer notifications on pref value changes.
  - Fallback to the existing pref system on unsupported platforms.

Differential Revision: https://phabricator.services.mozilla.com/D106996

UltraBlame original commit: 17587aa3ec61b72e06963dc81edc1c058fb1048c