Bug 1694739 - FOG IPC flush after 5s of idle r=janerik
Interestingly, on my machine, it appears to be after every 5s of idle,
not just after the first 5s of idle. We may wish to tweak that.

We also instrument the size and speed of IPC as seen from the POV of the parent
process, for monitoring purposes.

Differential Revision: https://phabricator.services.mozilla.com/D106947

UltraBlame original commit: 123143e183f97172102f3064a8bb779689fd31c1