Bug 1209947 - Part 2: Fire location change events for all subframe navigations, r=smaug,Gijs
The check in this function was added in bug 82236 to avoid an issue in the
mozilla suite where the UI would update to a subframe's URI. This came up
previously in bug 1206879 when we observed we weren't sending OnLocationChange
events when we were expecting, which was causing issues for pushState location
changes. A workaround was added in that bug to avoid the issue for the specific
case of pushState.

This patch removes the redundant check, and reverts the workaround added in that
bug. Unfortunately, we aren't able to fully remove nsISHEntry::GetIsSubframe, as
it is now used by Browser{Parent,Child}::CanCancelContentJS.

Differential Revision: https://phabricator.services.mozilla.com/D105555

UltraBlame original commit: c7cfaeeb66be2e3268901c5330c2712022d75eaa