Bug 1546540 - Part 2: Override newtab colorscheme in "Light" theme, r=dao
With the changes in part 1, the newtab page would now be rendered with a dark
theme if the user has prefers-color-scheme: dark due to OS styling, but
explicitly selects the provided "Light" theme.

This patch explicitly specifies the text and background colors in the "Light"
theme, ensuring that the newtab page will always be rendered with a light
colorscheme if the "Light" theme is selected, which is more in-line with the
previous behaviour and consistent with the built-in "Dark" theme.

Differential Revision: https://phabricator.services.mozilla.com/D107772

UltraBlame original commit: 04e63023b6e852e0236b68ea0a1565f06caf58dc