Bug 1697607 - Enable widget.remote-look-and-feel everywhere. r=cmartin,geckoview-reviewers,agi
(And add a missing include while at it)

The biggest concern here is potential startup time penalty (and it
shouldn't be much anyways, if at all). In exchange, we avoid doing a lot
of this work in content processes.

Let's keep an eye on for regressions, but this sticking allows us to
simplify a lot of the lookandfeel code in follow-ups.

Differential Revision: https://phabricator.services.mozilla.com/D107926

UltraBlame original commit: d77fccbf6d2a416f7b5d4a8dd5014ce4f29c516e