Bug 1650233: Make ScalarReplacement work for arrays in Warp r=nbp
The main obstacle here is that without TI, every `LoadElement` and `StoreElement` has `needsHoleCheck` marked unless it is a `StoreElement` from an `InitElem`.

We can handle most cases by observing that `LoadElement` can only observe a hole if a magic hole value has been stored to the array. This can only be done with `StoreHoleValueElement`, which we don't support in `IsElementEscaped`. (Note that the constructor for `MStoreElement` asserts that the value being stored is not a hole.) Therefore, we can omit the hole check for loads.

This gives us support for every case in `ion/recover-arrays.js` except for `arrayCond`. That's probably good enough for now.

Differential Revision: https://phabricator.services.mozilla.com/D107955

UltraBlame original commit: 0c3efc0015ecfee5359b7dd0e41737c469ff8278