Bug 1697210 - Make disabled checkbox / radio buttons do the right thing when painted with non-webrender. r=mstange
data:text/html,<input type=checkbox checked disabled>

Has a weird double border because the DrawTarget codepath fills and
strokes the same path, which is fine if the border is opaque, but not
otherwise.

Since this is the only use for a non-opaque border that we have, I
think, dealing with it on the caller seems a bit simpler. But let me
know if you want me to fix it more generally in
PaintRoundedRectWithRadius. I think for the semi-transparent border
case we'd need to create two paths, one for the background and one for
the border, which is a bit unfortunate.

The webrender codepath does the right thing, but of course that doesn't
get used for checkboxes.

Differential Revision: https://phabricator.services.mozilla.com/D107815

UltraBlame original commit: 3466de88cadcefde1b9a7be2c700da4e1d898eea