Bug 1697963 - Handle non-animated frozen images better, avoiding WR fallback. r=aosmond
If GetAnimated fails when we're getting a static request, because e.g.
the image is not yet fully loaded, we still create a frozen image, but
in that case we can do better and avoid the WR fallback.

Depends on D108197

Differential Revision: https://phabricator.services.mozilla.com/D108198

UltraBlame original commit: bc0e95a3d1b0b9e72bebfae18e03caecf72752cf