Bug 1698302 - Paint progress/meter chunks along with the track. r=mstange
The progress chunk already has the right size, but it's easier to get
the right clipping if we just paint it as part of the progress bar.

This is much like we paint the range track and thumb. We apply the
native appearance atomically so this should be fine.

Differential Revision: https://phabricator.services.mozilla.com/D108335

UltraBlame original commit: 71549d995fb402087f5247935ca0908814631d9c