Bug 1698496 - Improve the suggestedIndex-row-reuse logic in the urlbar. r=harry
For now this only improves the earlier quick fix from D108083. We should be
checking `queryContext.maxResults` instead of `this._rows.children.length` due
to the possible presence of stale hidden rows after all the visible rows.

Ultimately we should fix this for all rows, and I'd like to do that in this bug,
but the next merge is next week and there just isn't time. Right now we should
prioritize fixing this for the problem at hand, flickering in the quick suggest
result.

Differential Revision: https://phabricator.services.mozilla.com/D108519

UltraBlame original commit: 9226448501ca4b584a82e1432282bf23cb958e0a