Bug 1698441 - Remove call to synthesizeNativeWheelAndWaitForWheelEvent. r=tnikkel
This test is not actually run in CI (disabled via mochitest.ini), so
this has zero effect. Even if it did run, the call to
synthesizeNativeWheelAndWaitForWheelEvent is not passing in a callback
function; it runs the following rAF line immediately. So calling the
function that "waits for the wheel event" is pointless, we can just
call the version that synthesizes the native wheel input and it is
functionally identical.

Depends on D108441

Differential Revision: https://phabricator.services.mozilla.com/D108442

UltraBlame original commit: 5fffc5a7ed9fb75ae16a96cc1585cc63a59f3d8b