Bug 1698989: Use GetCrossDocParentFrameInProcess() in nsTreeBodyFrame. r=mstange
This patch doesn't change behavior; GetCrossDocParentFrameInProcess() is just a
wrapper for GetCrossDocParentFrame(), which is what we were calling before.

The "InProcess" version of this API (which we're migrating to in this patch) is
used to annotate GetCrossDocParentFrame() callsites that have been vetted as
being OK with the fact that this API returns null at the boundary of a
cross-origin iframe, if fission is enabled.

In this patch, the one call that I'm migrating is inside of XUL code
(specifically for the rendering of XUL trees); and we only expect to be
rendering XUL in a single process: Firefox's parent process. So this code
should be entirely OK with the fact that this API doesn't cross process
boundaries.

Differential Revision: https://phabricator.services.mozilla.com/D108708

UltraBlame original commit: d9637b5ab3c1a58c10a6e2bfee0b75dd34734297