Bug 1696844 - Fix mask access in pixman sse2 code. r=jrmuizel
The important changes here are a handful of places where we replace

            memcpy(&m, mask++, sizeof(uint32_t));

or similar code with

            uint8_t m = *mask++;

because we're only supposed to be considering a single byte from *mask.

I've also changed a bunch of other places to use this same pattern
(a local 8-bit variable) when reading individual bytes from the mask;
the code was inconsistent about this, sometimes casting the byte to
a uint32_t instead. This makes no actual difference, it just seemed
better to use a consistent pattern throughout the file.

Differential Revision: https://phabricator.services.mozilla.com/D108614

UltraBlame original commit: 887b81d71378c7c206804784d81368478e910c24