Bug 1699551 - Simplify nsMenuX state management. r=harry
We track two pieces of state: mNeedsRebuild and mIsOpen.
mNeedsRebuild is set to true on DOM mutations that require a rebuild, and set no
false when RebuildMenu() is done.
mIsOpen is set to true in MenuOpened() and set to false in MenuClosed().

Differential Revision: https://phabricator.services.mozilla.com/D108720

UltraBlame original commit: 70d5566772ce9493c5b480b85c595f78fe96d8c0