Bug 1698834 - Fix urlbar zoom indicator style (both for Proton and not). r=harry
Fix the margins on the zoom indicator, so that it's as tall as other icons.
Style the zoom indicator as an urlbar chiclet in Proton.
Add hover and active states to the zoom indicator and the identity-box in Proton.
Fix the center alignment of the cfr label.

Differential Revision: https://phabricator.services.mozilla.com/D108920

UltraBlame original commit: 25186df80cfaa93435ac78002d1ff48c92e35b5f