Bug 1699035 [wpt PR 28115] - Fix scroll-to-text-fragment behavior for same-document navigation, a=testonly
Automatic update from web-platform-tests
Fix scroll-to-text-fragment behavior for same-document navigation

A same-document navigation to a URL with a text fragment anchor
should not scroll the page to the text fragment. The test
modified by this CL is meant to test that. However, that behavior
was broken by this CL:

https://chromium-review.googlesource.com/c/chromium/src/+/2533688

The reason the test hasn't been failing is that it has been checking
to see whether the page was scrolled too soon. When a text fragment
anchor finds a match, it doesn't scroll the page right away; it
schedules a task to scroll the page which runs at rAF timing on the
next lifecycle update. This scheduling behavior was added by:

https://chromium-review.googlesource.com/c/chromium/src/+/2182992

Just adding requestAnimationFrame to the test caused it to fail;
the change to document_loader.cc fixes it.

Bug: 1147453
Change-Id: I001a315418f5f8a61ded7d87023740cd366389f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2766287
Reviewed-by: Stefan Zager <szagerchromium.org>
Reviewed-by: Joey Arhar <jarharchromium.org>
Reviewed-by: Nick Burris <nburrischromium.org>
Commit-Queue: Stefan Zager <szagerchromium.org>
Cr-Commit-Position: refs/heads/master{#863893}

--

wpt-commits: 9d6b249df9ca502dccc33939762b3457a6cf9015
wpt-pr: 28115

UltraBlame original commit: 28ff33df386c86e2ec447cbea21d19cb79b592b3