Bug 1698791 part 2 - Use isIndex/CheckStringIsIndex for StringIsArrayIndex. r=evilpie
StringIsArrayIndex is almost identical to isIndex/CheckStringIsIndex, but it's
limited to UINT32_MAX - 2 instead of UINT32_MAX.

This patch de-duplicates this with an explicit check for the smaller max length.
This ensures we can benefit from the index-value optimization and the ATOM_IS_INDEX
flag added in the next patch.

The difference in maximum length is still confusing and should probably be cleaned
up separately.

Differential Revision: https://phabricator.services.mozilla.com/D108742

UltraBlame original commit: 467a7afcdd3d4d489833ac88803d244529b7bc0e