Bug 1699878 - Speed up and reduce fuzziness for reftest 966992-1.html. r=emilio
Delete <script> and reftest-wait because the scrollbar shouldn't be
fading in the reftest, and we can save 5 seconds on CI!

Use default monospace font which is good enough without fuzziness.

Delete style for textarea because there is no textarea in the test nor
in the reference.

Differential Revision: https://phabricator.services.mozilla.com/D109193

UltraBlame original commit: 4bb91629504a47a7870076ba9afb2bdd78d88954