Bug 1686295 - Flip the order of two time measurement calls in telemetry to provide actual guarantees to callers. r=chutten
See the comment in the code for the rationale. Having it in the other order
(before this patch) caused a but real rare intermittent on Linux and macOS, the
two OSes where this is used and the clock has a high enough resolution. This
happened when the first call was made before a millisecond mark and the second
one after, with the same time source: machines on treeherder don't suspend, the
numbers should be equal, but rarely the time excluding suspend was longer by 1ms
compared to the time including suspend.

Differential Revision: https://phabricator.services.mozilla.com/D109370

UltraBlame original commit: e8ae095cb03a960354628f2e8550b6750898fd86