Bug 1698997 - Make it possible to listen for NativeMenu opening/closing. r=harry
This will be used for native context menus. I'm planning for nsMenuPopupFrame to
be the owner of the NativeMenu, and it only wants to keep the object around
while the menu is open. So it needs to be notified when it closes.

Differential Revision: https://phabricator.services.mozilla.com/D109178

UltraBlame original commit: 39473c31bf74eae9a3a9cc05bdaf7679a3e92e19