Bug 1699910: Fix anyFormalIsAliased assertion r=jandem
When we are guarding that no argument is forwarded, ArgumentsReplacer::visitGuardArgumentsObjectFlags asserts that no formal argument is aliased. This is too general: in strict mode, arguments can be aliased by the call object, but not forwarded.

This patch fixes the check in `anyFormalIsAliased` and renames it for clarity.

(I tried adding assertions in `MaybeForwardToCallObject` that we only mark arguments as forwarded if `script->anyFormalIsAliased()`, but that runs into problems with `arguments.callee.arguments` and `ArgumentsObject::createUnexpected`.)

Differential Revision: https://phabricator.services.mozilla.com/D109413

UltraBlame original commit: fe9b5592cd3173e15bd005bd6777ccdaffa8185b