Bug 1666386 - [devtools] Find blocked request using url name fixes intermittent r=nchevobbe
The test previously depended on the position of the blocked request
in the request list view, This causes the test to fail intermittently
when the list is out of order.

This fix instead finds the request using the name in the url which is
more consistent.

Differential Revision: https://phabricator.services.mozilla.com/D109497

UltraBlame original commit: 2032f2b926ddf13702d5720983d226273a4290a6