Bug 1695451 - Handle I/O Errors when loading Normandy preference experiments r=Gijs
When there is a failure to load the experiments store, Normandy now simply
resets the store back to an empty set which is generally a safe operation.
Being resilient to errors here is especially important now that about:support
shows data from Normandy. Errors in loading data for about:support can cause
the entire page to be blank, blocking critical support information.

This is a safe operation from a clients point of view because withouot this
information Normandy will assume that no clients should be enrolled. It may
take a restart, but eventually it will reset the client back to default if
there are I/O errors.

Differential Revision: https://phabricator.services.mozilla.com/D109562

UltraBlame original commit: a571d69735d9373a98b8302f0bebe3ee71f45c54