Bug 1700900 - js shell: improve propagation of wasm-specific command line arguments to child processes.  r=lth.
In some circumstances, the JS shell may create a child process to perform wasm
compilation.  It can happen that the child process chooses a different
compiler than the parent, and/or possibly that the child allows different wasm
extensions than the parent.  This happens because command line arguments that
select such compilers or extensions are not propagated to the child, when they
should be.  This has been observed to lead to an assertion failure.

This patch ensures propagation of the following flags:
```
  --wasm-gc
  --wasm-simd-wormhole
  --wasm-exceptions
  --wasm-function-references
  --no-wasm-simd
  --no-wasm-reftypes
  --no-wasm-multi-value
  --test-wasm-await-tier2
  --wasm-force-ion
```
`--disable-wasm-huge-memory` is already propagated.

It also moves propagation of the following down, so that all of the flag
propagation logic is together, at the end of `main()`.
```
  --no-sse3
  --no-ssse3
  --no-sse41
  --no-sse42
  --enable-avx
```

Differential Revision: https://phabricator.services.mozilla.com/D109863

UltraBlame original commit: 4bf8856cbe568aff50df3e4757fa1a34f91f56e6