Bug 1699224 - Part 3. Complete shutdown of SharedSurfacesParent on the main thread. r=jrmuizel
We want to use nsExpirationTracker to decide when we should unmap shared
surfaces from our address space. This relies upon the observer service
and we need to be both initialized and shutdown on the main thread to
add/remove the observer. SharedSurfacesParent is already initialized on
the main thread, and we just need to complete shutdown on the main
thread.

Differential Revision: https://phabricator.services.mozilla.com/D109439

UltraBlame original commit: 85f3499080db2ea7e598e86750a56f982c1989ad