Bug 1701344. Change diagnostic assert in GetClipWithRespectToASR back to a debug assert. r=mstange
In bug 1700812 we made this a diagnostic assert to try to get a better handle on it, and the potential crashes it could cause later inside webrender, but we are hitting it too often. That should be enough data on this specific assert.

Differential Revision: https://phabricator.services.mozilla.com/D109984

UltraBlame original commit: 6af343f10c00b5d533b5222fe12056b9644d0402