Bug 1701608 - remove perma-busted python dependency generator. r=releng-reviewers,jmaher
I have manually pinned the requirements, and partials are working. We'll lose a way to auto-update our dependencies, but a) it's perma-busted, so realistically we're just losing a busted task, and b) mhentges is working on a way to auto-update pinned dependencies in-tree.

Differential Revision: https://phabricator.services.mozilla.com/D110148

UltraBlame original commit: 6d68a0116476bd1dbda7db65f816e580db019ec6