Bug 1699426 - Remove margin-block for resetProfile learn more r=Gijs
Chatted with Emanuela about this and she noted that this shouldn't be such a
large gap here. Note: there should be a larger gap here between the text and
the buttons, but I think that applies to all modals. That should be a separate
patch, if you agree.

Differential Revision: https://phabricator.services.mozilla.com/D110131

UltraBlame original commit: fa4e36721341b571d38543ab9edd344648bf40ec