Bug 1700987 - Invoke `post-update-processing` in `--backgroundtask backgroundupdate`. r=application-update-reviewers,bytesized
This wasn't being invoked for a number of reasons.  First, the
`profile-after-change` category notification is not registered in
background task mode.  But we don't want to register it for all
background tasks, thereby loading the update service in all tasks.  So
we manually prod the relevant code.  We do this via the
UpdateServiceStub constructor, instead of more directly, so that there
is a single entry point to this functionality in and out of background
task mode.  However, this requires not delaying post update processing
in background task mode.

The change to use the best pending state allows the backgroundtask to
try to use the Maintenance Service multiple times.

With this change, repeated `--backgroundtask backgroundupdate`
invocations with:

a) staging enabled
b) the Maintenance Service available
c) a MAR that the Maintenance Service will not accept

Leads to a loop like:
```
$ --backgroundtask backgroundupdate
  update.status="pending"
$ --backgroundtask backgroundupdate
  after updater.exe
    update.status="failed: 105"
  update.status="pending-service"
$ --backgroundtask backgroundupdate
  after MMS
    "failed: 49"
  update.status="pending"
$ --backgroundtask backgroundupdate
  after updater.exe
    update.status="failed: 105"
  update.status="pending-service"
[repeated]
```

Differential Revision: https://phabricator.services.mozilla.com/D109822

UltraBlame original commit: ca38650213070783f0e6a05a05d54a6e041ffdfe