Bug 1691907 - Part 3: Set the cookieJarSettings for Web Extension downloads.download(). r=robwu
This patch makes the downloads.download() to create the
cookieJarSettings and set it to the download source, so that the
downloading channel will get the partitionKey from the
cookieJarSettings. This will use the download url as the partitionKey,
meaning that we treat the download as a first-party loading.

Differential Revision: https://phabricator.services.mozilla.com/D109760

UltraBlame original commit: f614295240dd6394825fbb5941e293414ad31c31