Bug 1674277 - Make AUS.stopDownload asynchronous r=nalexander,preferences-reviewers
The Downloader's cleanup function really ought to be asynchronous. The observer for "quit-application" is already asynchronous, so there's no problem there. But really, we ought to be cleaning up each downloader that we use, not just the last one. Which means that the cleanup ought to happen in AUS.stopDownload. So this patch will convert AUS.stopDownload to be asynchronous so that we can properly clean up the downloader from there.

Differential Revision: https://phabricator.services.mozilla.com/D110646

UltraBlame original commit: 4fc1b109610f205393a8fb18d0b782c544cacf8e