Bug 1701597.When zooming to a rect horizontally center it. r=botond
We already vertically center the rect, but not horizontally. We generally try to zoom to the width of an element already, so the width of the zoom to rect will usually occupy the whole viewport and hence be centered. But if the width of the element is too small so that we can't zoom in so that it fills the viewport it will end up left aligned. This doesn't come up very often on mobile, but happens more so on desktop.

Differential Revision: https://phabricator.services.mozilla.com/D110063

UltraBlame original commit: 05ae63ecb3e54554b5ad4a4bf490227eebd048b7