Bug 1703732 - Catch exception if `BackgroundAppUpdate` policy has locked `app.update.background.enabled`. r=bytesized,application-update-reviewers
There are two parts to this.  The first part is simply fixing the
issue, which was an oversight.  The second part is rejecting with a
JavaScript `Error`, which provides a nicely formatted stack in the
browser console, easing debugging.

Differential Revision: https://phabricator.services.mozilla.com/D111223

UltraBlame original commit: dcd8281b5f2247f4f87f75f90b46f015e4bc9b67