Bug 1703543 - [devtools] Fix intermittent leak on browser_target_configuration_command_dppx.js. r=bomsy
It looks like the test was leaking because we were using loadUri to reload, or the fact
that we were using browserLoaded after calling loadUri.
So here we do create a promise from browserLoaded before initiating the navigation,
and we use the more straightforward gBrowser.reloadTab to reload.

Differential Revision: https://phabricator.services.mozilla.com/D111396

UltraBlame original commit: 0a1b718804639fa137739d7982a5e5a559fc91c8