Bug 1686626 - Enable Ion by default for wasm on AArch64.  r=lth.
This patch enables Ion by default as the optimising compiler for wasm on
AArch64, and disables Cranelift.  Cranelift is still available if the build is
configured with --enable-cranelift.  In that case, *only* Cranelift is
available.  There are no configuration flags to enable both Ion and Cranelift
simultaneously.

This mostly reverts the Phase 0 and Phase 1 patches that are bug 1678097
D102420 and D101867 respectively.

The command line option --wasm-force-ion has been removed.

With this patch in place, users of the shell should specify
`--wasm-compiler=optimizing` to get an optimising wasm compiler.  Which one is
provided depends on the configuration options as described above.
`--wasm-compiler=cranelift` and `--wasm-compiler=ion` are now only accepted
when the relevant compiler has been enabled, and so neither is a "safe" way to
request an optimising tier.

For that reason, test directories that previously requested
also-with-Ion-please by stating `test-also=--wasm-compiler=ion;` in their
`directives.txt` file, have been changed to use
`test-also=--wasm-compiler=optimizing;`.

In places where the JSContextOptions are set, the non-selected compiler (Ion
or CL) is explicitly set to `false` (eg, `.setWasmIon(false)`).  This may be
overly conservative, but seems wise given that it's not immediately obvious
what the previous value of that flag is, and given the recent difficulties
with incorrect option propagation/handling (eg, bug 1697560).

Differential Revision: https://phabricator.services.mozilla.com/D101695

UltraBlame original commit: 6d56998107808e18a2f495cc693894494de1ce67