Bug 1704164 - Apply darkmode theme to newtab before first paint r=mconley
This prevents a white flash in about:home in dark mode. Particularly obvious
with the skeleton UI enabled. I don't see a reason why we need to do this on
pageshow, if we start doing it on DOMContentLoaded, given that we listen for
theme change events anyway.

Differential Revision: https://phabricator.services.mozilla.com/D111493

UltraBlame original commit: dc67a43df7384b22eb61822c0b1b020ad45989cb