Bug 1704467: Add guard for constant OSR phis outside of preheader r=jandem
When we replace a redundant phi with a constant, we have to add guards in the OSR block if the phi depends on an OSRValue. With the changes to branch pruning, we can have phis that depend on OSRValues outside of the preheader.

Differential Revision: https://phabricator.services.mozilla.com/D111986

UltraBlame original commit: d0210270c7e821f637a92a50c591e5d20bb48f90