Bug 1705062 - Handle pageshow in LightweightThemeChild r=mconley
It appears for the sidebar document, this.contentWindow.top is null at the time
of DOMContentLoaded. Accordingly we need to wait for pageshow. This patch waits
for pageshow and ensures we don't update twice when pageshow shows up if we
successfully updated on DOMContentLoaded.

Differential Revision: https://phabricator.services.mozilla.com/D112124

UltraBlame original commit: 24d8661286597ff022d837d74227ab53ddf7add3