Bug 1705253 - Partially disable accessible/tests/browser/mac/browser_app.js when native context menus are used. r=morgan
This test has a part where it expects AXMenuOpened events for context menus, which are not observable for us when native context menus are used.

It also has a part where it inspects the accessibility-exposed contents of a context menu. This is not testable with native context menus either, so that part of the test needs to be disabled.

Differential Revision: https://phabricator.services.mozilla.com/D112122

UltraBlame original commit: 6ac255a315ca7a22862e54e92ce0cee1c4a7fa37