Bug 1705270 - Fix assert when compositor surface is in a pass-through picture r=gfx-reviewers,lsalzman
We don't currently handle the case where a compositor surface is inside a
picture primitive that is a pass through (has composite mode of None),
because we only count required compositor surfaces at the top level prim
list of a tile cache.

However, the code to promote a surface was only checking if it existed on
the same physical surface, not the root picture cache.

This patch prevents compositor surfaces being promoted in this case (which
only occurs if inside a backdrop-filter container, or a 3d transform root).

Differential Revision: https://phabricator.services.mozilla.com/D112153

UltraBlame original commit: e3ee5a86a98ca361d3ab34a2ae416b64287bf3c7