Bug 1705624. Change the APZCTreeManager zoom constraints map hash to be presshellid ignoring. r=botond
This is not strictly necessary (but a later patch would need to be modified slightly to get the presshellid).

But we can potentially pull zoom constraints from the apzc map, which is presshellid ignoring, so it seems better to have them in-sync, and I couldn't find a reason why zoom constraints needs to be presshellid respecting.

Depends on D112315

Differential Revision: https://phabricator.services.mozilla.com/D112316

UltraBlame original commit: 5a88fbdbaad8d7f4afbe23a356e02a8072337189