Bug 1705622. Update zoom constraints on apzc's every time we update the hit testing tree. r=botond
An apzc could have moved to inside an async zoom container.

In the failing test we did a paint while painting was suppressed, so our display list had exactly one item: a nsDisplaySolidColor. The layer tree still got tagged with metrics and so the apzc was attached. Then the zoom constraints client does it's before first paint update which is triggered when we unsuppress painting. It finds no async zoom container so it does nothing. Then we do a regular paint, with an async zoom container this time. We find the apzc from the previous paint, so we don't update the zoom contraints.

Differential Revision: https://phabricator.services.mozilla.com/D112555

UltraBlame original commit: 429579435141fa7c6de401102a16826269e31086