Bug 1705979 - Make it easier to distinguish between copying and scaling in profile. r=aosmond
With the copying and scaling code in one place, it gets a bit hard to see in the profile which one
is being used. This just moves some of the dispatching between the two up a level so they show up as
different functions in a profile.

Differential Revision: https://phabricator.services.mozilla.com/D112613

UltraBlame original commit: f00e09e4708453f0869ebfbd1919b70b7ea12696