Bug 1662559 - Part 4: Introduce ClassBodyLexicalEnvironmentObject. r=arai
It is a copy of BlockLexicalEnvironmentObject. Note that

-   ClassBodyScope is *not* a LexicalScope, BUT

-   ClassBodyLexicalEnvironmentObject *is* a LexicalEnvironmentObject

This is strictly a matter of how much implementation is shared. The word
"Lexical" in LexicalEnvironmentObject, and to some extent throughout
SpiderMonkey, is alas already pretty meaningless. In the language
specification, all environments are lexical environments.

Depends on D108286

Differential Revision: https://phabricator.services.mozilla.com/D108287

UltraBlame original commit: d3b226ecc677c0474d1f5509dc25a574947cd810