Bug 1697371 - Templatize baseline compiler part 4: Shifts and 32-bit rotates. r=rhunt
Shifts and rotates may require a variable RHS to be in a specific
register.  We can express this generically by passing an rhs-popper
function to the emitter.

(64-bit rotates are additionally complicated by requiring a temp, and
are not translated in this patch.)

Differential Revision: https://phabricator.services.mozilla.com/D109610

UltraBlame original commit: c6a95e1a6b94ad6e44360afe09f8469bd138fa0c