Bug 1678044 - Use correct profile for wdspec tests, r=webdriver-reviewers,whimboo
This ensures that we always pass a profile in the capabilitites. When
we're running with a single process it simply unpacks the profile to a
temporary folder and passes in that path as a firefox argument. When
we're running with multiple processes, that won't work. We could
invent a mechanism to get one profile per process, but for simplicity
we instead pass in the profile as a base64-encoded string.



Depends on D97665

Differential Revision: https://phabricator.services.mozilla.com/D99697

UltraBlame original commit: c5914548a0846a1e3405e09aeb71df2391f9170e