Bug 1705341 [wpt PR 28502] - CSP: Move form-action check to the browser, a=testonly
Automatic update from web-platform-tests
CSP: Move form-action check to the browser

The Content Security Policy check for 'form-action' is a navigational
check, that we perform at the moment half in Blink (for the initial
URL) and half in the Browser (for redirects). The Browser part is
currently implemented in a navigational throttle and is not using the
correct policies and context.

Thanks to the PolicyContainerHost, this CL simplifies the check in the
Browser by removing the FormSubmissionThrottle and moving the check to
the NavigationRequest, in the same place where other navigational
checks are performed. At the same time, it removes the check from
Blink relying to the Browser check also for the initial request URL.

Change-Id: I28be978be3c86f1d8ad8b41398542c4e735758e3
Bug: 694525,663512,700964,1172898,1021462,713388
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2797339
Commit-Queue: Antonio Sartori <antoniosartorichromium.org>
Reviewed-by: Mike West <mkwstchromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzognichromium.org>
Cr-Commit-Position: refs/heads/master{#875047}

--

wpt-commits: ae2b12963e524582555638df74e7a287d35aa048
wpt-pr: 28502

UltraBlame original commit: ccffb5bf191601460bcf0f9d5f082ae39f307256