Bug 1707422 - Fix non-inline string data pointer for dependent strings created by rope flattening r=jandem
There is a bug in the string flattening changes that happens when we reuse the
leftmost string as the buffer for the flattened string. In this case the
chacter data pointer for dependent strings converted from ropes is set
incorrectly until we reach the leftmost rope, since we initialize the |pos|
variable to the first character after the contents of the leftmost string. In
the original version of this code we use |str->setNonInlineChars(wholeChars)|
for these. This also results in the lengths of these dependent strings being
wrong since this is caculated based on this pointer. This leads to the
assertion failure.

Differential Revision: https://phabricator.services.mozilla.com/D113306

UltraBlame original commit: 87457b5402670d2288c4b2dee515a0f397d7b280