Bug 1706098 - [devtools] Make inspectedWindow.reload with userAgent implementation match Chrome's one and add support for Fission. r=ochameau,rpl,devtools-backward-compat-reviewers.
Our implementation was setting a custom user agent on the document
before reloading it, and then resetting it as soon as the document
was loaded.
In Chrome, once the webextension sets the user agent, it persists
across reload and navigations, and is reset only when the toolbox
is closed.
This patch is adding similar behaviour to our implementation.

We also set the currentUserAgent flag on the browsing context from
the parent process instead of content. This is re-using the
targetConfigurationCommand, which also handle for us resetting the
user agent when the toolbox closes.

The existing test is expanded so it includes a remote iframe to ensure
the feature is supported on Fission. It also check the value of the user
agent through navigator.userAgent, and not only through the request header.

Differential Revision: https://phabricator.services.mozilla.com/D112980

UltraBlame original commit: 82d59bab21dadcad3c63aa1537fc48fa3c8a4e98