Bug 1704842 - Expose gecko version at time compositor is initialized to geckoview_streaming telemetry. r=janerik
On Fenix we see many users with a value of opengl for the
gfx_status_compositor telemetry probe, despite the fact they are
running versions of gecko for which webrender should be enabled on
their device.

The theory is that although gecko is being initialized the user has
not opened a tab, and therefore initialized a compositor, since they
have been updated to a version which should be running webrender by
default. To verify that this is the case, this patch adds a new probe
gfx_status_last_compositor_gecko_version, which reports the gecko
version at the time a compositor is initialized. This will be used to
confirm that the users whose compositor unexpectedly reports opengl
have indeed not initialized a compositor in a while.

Differential Revision: https://phabricator.services.mozilla.com/D112339

UltraBlame original commit: 3d0ed017553b736cae346be66372aab66a002560