Bug 1707560. Don't scroll to the top if double tap on very tall element that is > 90% the width of the viewport. r=emilio
We have code for this, it's disabled for elements that are 90% or greater the width of the viewport. This code was added in bug 1204084 for the "double tap on tall element" problem and the 90% hasn't been touched since. I can't think of a good reason why we would want to scroll to the top of the element if its 90% the width of the viewport or more.

Differential Revision: https://phabricator.services.mozilla.com/D113335

UltraBlame original commit: 0ec725406db0a9ead80e150c24fd508dcae38d15