Bug 1707614 - Remove one unused usage of :-moz-devtools-highlighted. r=nchevobbe
This internal pseudo-class used to be to highlight nodes in the
inspector, but the inspector no longer sets it. Now the only thing that
sets it is devtools/server/actors/accessibility/audit/contrast.js to
remove the text color / shadow from the test.

We should probably find another way to implement that that doesn't
misuse this pseudo-class (and then presumably remove this pseudo-class),
but for now this code is dead.

Differential Revision: https://phabricator.services.mozilla.com/D113371

UltraBlame original commit: 1e3e0d3d04ba97ffc9c1cdbf2064d5a50506bd9f