Bug 1708179 add AudioParam::SetInitialValue() for use in AudioNode creation r=padenot
when the setter should not throw because there are no events in the timeline
and the time is known valid.

This will allow us to make SetValue() throw when it conflicts with a
setValueCurveAtTime() curve.

Depends on D113937

Differential Revision: https://phabricator.services.mozilla.com/D113938

UltraBlame original commit: 70e615eb295d5ebe006aa15e2a07f3ec8f4c1958