Bug 1706094 - [devtools] Set watchedByDevTools browsingContextflag in the parent process. r=ochameau.
We're using frame-helper#createTargets to set the flag as we're already looping
over the browsing context we need to handle.
We keep a Map of watcher/observer so we can remove the observer when the targets
get destroyed.

For the browser toolbox top-level browsing context we keep setting the flag in
the browsing context target actor as it lives in the parent process, and we
don't have watcher support enabled by default, so frame-helper wouldn't be hit.
frame-helper is not used in the legacy browser toolbox, so we won't set the flag
on tabs when devtools.browsertoolbox.fission is false, but that's okay it's not
currently set anyway in this setup.

Tests are added with multiple cases to ensure the flag is set in different situations.

Differential Revision: https://phabricator.services.mozilla.com/D113628

UltraBlame original commit: e65baad550b09cc3adffc09ca5e65a06720c13ec