Bug 1701795 [wpt PR 28287] - [GridNG] Compute the correct offset for out of flow items that start or end in the middle of a range, a=testonly
Automatic update from web-platform-tests
[GridNG] Compute the correct offset for out of flow items that start or
end in the middle of a range

Previously, if an out of flow item happened to be in the middle of a
range, its offset and size were computed incorrectly. In this change,
helpers are added to compute the extra offset needed. Two web tests were
added to validate this behaviour.

Additionally, we now store range related data instead of the set indices
for out of flow items, as in this case we need the first and last
spanned range, as well as the offsets of the tracks. With this
information, we can just compute the set indices on demand.

Bug: 1045599
Change-Id: I26c8e1027d89c601de8e81d3f31850e061423844
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2757833
Commit-Queue: Ana Sollano Kim <ansollanmicrosoft.com>
Reviewed-by: Ethan Jimenez <ethavarmicrosoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmimicrosoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Cr-Commit-Position: refs/heads/master{#877798}

--

wpt-commits: f977e557d26b2f20d58c3fbd4e1503ea553cef6b
wpt-pr: 28287

UltraBlame original commit: f498819856d46c04551e5dc1cb3d1606f092991d