Bug 1710336 - Deduplicate clip template ids when creating hit-test nodes. r=gfx-reviewers,nical
Ideally, we wouldn't need to de-duplicate clip ids here, this would
be done at a higher level. However, due to the way we currently
handle redundant stacking context clips, this is a significant
performance win on some pages.

For example, on the youtube.com home page, this drops the size of
the hit-test clip nodes array from ~3800 to ~2200 per scene. This
is still much higher than it should be, but this gives an immediate
performance win while working on the more involved changes to handle
this in a more optimal way.

Differential Revision: https://phabricator.services.mozilla.com/D114705

UltraBlame original commit: 63615880223d7ae61aaf6888b58a38d26aeee88b