Bug 1705726 - Create an interface to determine if the maintenance service registry entry exists r=nalexander
Creates nsIUpdateProcessor.serviceRegKeyExists(), which checks for existence the registry key written by the installer that indicates that the current installation can use the Maintenance Service.

The nsIUpdateProcessor interface was chosen because that is the one Update-related xpcom component that is implemented in C++. By implementing this in C++, we can use the existing implementation of CalculateRegistryPathFromFilePath() to do most of the work for us.

Unfortunately, we can't really test this feature without being able to write to HKEY_LOCAL_MACHINE, which we have no good way of doing during testing.

Differential Revision: https://phabricator.services.mozilla.com/D114803

UltraBlame original commit: 4a2629df75a41ac8f872198cebc2b13c5c63ed0b