Bug 1458339 part 1 - Add an API to set self-hosted XDR content. r=tcampbell
The JSRuntime already has an API to set the self-hosted content before the
initialization. This modification adds a proper API such that the rest of Gecko,
which does not have access to the JSRuntime implementation can use this
functions as well.

Differential Revision: https://phabricator.services.mozilla.com/D110577

UltraBlame original commit: 51ff7ca947e9c24a4cf7c99965ba6d6815725552