Bug 1708534 - Use the specified input colours for dark and light built-in themes. r=jaws
It looks like we erroneously used the placeholder colour for the dark theme
input colour. This updates the dark theme to use the correct colour. It also
updates the light theme to use the same input colours as defined in the
light variation of the system theme.

Differential Revision: https://phabricator.services.mozilla.com/D115000

UltraBlame original commit: 633b705d14eb3761c66fdea6d6c6bae63be0db54