Bug 1707832: Ensure that SurfacePoolCA doesn't store or give out null recycled surfaces. r=mstange
The general contract of a SurfacePoolEntry is that mIOSurface should be
non-null, but this isn't asserted. This patch adds release asserts to ensure
that these surfaces are non-null at the time they are stored and released. It
is one of the possible non-OOM ways that NativeLayerCA::NextSurface could
store a null surface.

Differential Revision: https://phabricator.services.mozilla.com/D113870

UltraBlame original commit: 79a02e8fc876666721aa7d31c889385ff16cf349