Bug 1711047 [wpt PR 28989] - Fixed nested RecalcAssignment in DisplayLocking case, a=testonly
Automatic update from web-platform-tests
Fixed nested RecalcAssignment in DisplayLocking case

This loosens the flat tree traversal to a simple node traversal added by
crrev.com/877657, but I think that's OK since the scenarios of interest
in both that patch and this one are where the display lock and the slot
are in the same TreeScope. We can follow up on whether or not this is OK
later, I made a TODO for it.

The relaxed check, in combination with the removal of the HeapHashMap
and additional iteration over Slots(), likely fixed the performance bug
crbug.com/1204336 as well.

Fixed: 1208573, 1204336
Change-Id: I938cff05fa3acc9f65a9b0451acb1ad0b2ce6c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891511
Reviewed-by: Chris Harrelson <chrishtrchromium.org>
Commit-Queue: Joey Arhar <jarharchromium.org>
Cr-Commit-Position: refs/heads/master{#882664}

--

wpt-commits: 508ad6b7742936774d9e86b0da19f543b40004b7
wpt-pr: 28989

UltraBlame original commit: a519411c657640933a668928f075c15fafa0b676