Bug 1707876 - Allow testrunner processes to use a full logger, r=whimboo
Previously we used a special MessageLogger class for logging in
TestRunner processes. This was only set up to allow normal logging
messages (debug/info/etc.). But for wdspec tests in Firefox we're
reading the geckodriver output in the TestRunner thread, so require
the corresponding log actions. The simplest way to do this is to
create an actual StructuredLogger in the child process and use a
handler on that logger to proxy the messages back to the parent, as
before.

Differential Revision: https://phabricator.services.mozilla.com/D113514

UltraBlame original commit: 6161d2be653b2300a1e1a1d48f0ae928fcaeb17d