Bug 1700968 - Make sure to retry if AsyncResolveNative fails r=necko-reviewers,kershaw
When TRR is used in mode 3(only in mode 3) AsyncResolveNative can fail immediately (sync) and we need to make sure that we retry it in the same way as in DnsAndConnectSocket::TransportSetup::OnLookupComplete. If RESOLVE_IP_HINT is set we already retry he lookup, but we do not retry if mRetryWithDifferentIPFamily is set.

Differential Revision: https://phabricator.services.mozilla.com/D114952

UltraBlame original commit: 1d222151411ab43832c005804fa98db4633b7529