Bug 1707959 - Set TYPE_GLOW overscroll type on Android 12. r=droeh
Android 12 introduces a new type of overscroll effect named TYPE_STRETCH which
stretches the whole surface when overscrolling.

Unforunately, given how it's implemented, SurfaceView does not support this
type of effect, similarly to many other View effects.

The OverscrollEffect is currently crashing on Android 12 builds so we restore the
type to the previous (and only) default to fix the crash.

Note that Overscroll is still not displayed, I'll open a new bug to fix that
part.

This patch also avoids using reflection for setBlendMode on API 29 and later
where an API is actually available. Using reflection to access private APIs is
not available since Android 9 so this code was a no-op before this patch.

Differential Revision: https://phabricator.services.mozilla.com/D115031

UltraBlame original commit: f50820278a98553b80954165cee19ae9f42e9f02