Bug 1700434 - Implement gl_ClipDistance support in SWGL. r=jrmuizel
This implements gl_ClipDistance so that the text shader doesn't have
to manually evaluate whether or not we're inside the untransformed
source texture. This simplification is necessary so that we can work
towards using swgl_commitTexture in the text shader.

Differential Revision: https://phabricator.services.mozilla.com/D115457

UltraBlame original commit: 6643c36ec306882c3f2a09119642b0a2cd811d8b