Bug 1567222 - [profiler, network] Change how we pass a few properties to the network markers r=valentin,necko-reviewers
This is simplifying how we pass the URI, and this is abstracting how we
pass the size.
In general the goal is to use properties and methods from
HttpBaseChannel, so that maybe in the future we could abstract marker
handling to that parent class.

Differential Revision: https://phabricator.services.mozilla.com/D112212

UltraBlame original commit: 6cb63cc7c95d7757f3e4fc334421cf50bbae258b