Bug 1712278 - Prevent arrowscrollbox from scrolling more than one page at a time when using line scrolling. r=Gijs
An alternative to this is changing the line scrolling acceleration Gecko
does to double the amount-per-line, rather than the amount-of-lines that
we scroll.

(But actually, even if we determine that that is a better course of
action, we probably want this anyways).

Differential Revision: https://phabricator.services.mozilla.com/D115692

UltraBlame original commit: 67f78f7890d9066fd1e746417984b6cc431f8513