Bug 1711218 - Don't undo user pref changes when Normandy experiments end r=Gijs
When a user changes a preference involved in a Normandy experiment, we no
longer immediately end the experiment. This caused a problem because Normandy
didn't check for user changes when resetting prefereces at the end of an
experiment. As a result it would reset all preferences back to their original
values on the specified branch, even if that would mean changing a preference
it knows the user had modified.

This change causes Normandy to never reset the user branch of an experiment
preference that the user has changed. This is true even if the user changes the
preference away from the experimental value and then back to it.

Differential Revision: https://phabricator.services.mozilla.com/D115713

UltraBlame original commit: cd30ec1b16a39579d73909c58dbdd095fe1e9ac8