Bug 1706164 - Use webrtc screen and window sharing data in EncryptedMediaChild to determine if media may be captured. r=alwu
This expands EncryptedMediaChild to be able to query and track the information
now shared by webrtcUI around screen and window sharing. The actor will use this
information to report to observers if capture is possible or not.

For now we use a coarse approach where we only track if the screen or any window
is shared, but in future I plan to have a more granular tracking of windows.

Differential Revision: https://phabricator.services.mozilla.com/D115417

UltraBlame original commit: d0e98ea7c1a906301d319d6b7f10e5a26c29db79