Bug 1710694 - [devtools] Fix custom error with message getter. r=ladybenko.
In the case a custom error object was defining the `name` and/or `message` property
in getters, we weren't displaying those information in the console.
That's because when retrieving those information, we were avoiding calling getters,
as they're deemed unsafe.
In those specific case, even if unsafe, we always want to show it to the user
so they're not missing any information when debugging.

A test case is added to ensure this works as expected and we don't regress.

Differential Revision: https://phabricator.services.mozilla.com/D115678

UltraBlame original commit: 92b0b44e2c256e4578341d3a14e5f9a6939808f2