Bug 1712182: Test aria-owning an ancestor which isn't created yet with an iframe in the subtree. r=eeejay
The actual fix for this is in bug 1387308.
However, triggering this bug with an iframe in a remote document caused a crash in the parent process because the parent process was never sent the OuterDoc.
Given the added complexity here, I thought it worth having a separate test.

Differential Revision: https://phabricator.services.mozilla.com/D115812

UltraBlame original commit: 346838dffac265527f56261bbf34fa728a8c61e5