Bug 1712815 - Fix the 'remove a file from a moz.build file' logic r=jewilde
We were raising an exception as a result of a python error
which was masked by a catch statement that thought we didn't
know how to remove the file. But we did.

Fix the python error, and add an exception at the end of the
function to be raised and then caught when we really don't know
how to remove a file.

Differential Revision: https://phabricator.services.mozilla.com/D115936

UltraBlame original commit: 9017050012ed0fede49fc8c9c9948874051c9069