Bug 1713409 - Always support a single update rect for SwCompositor. r=gfx-reviewers,jrmuizel
Prior to bug 1712871, when SwCompositor was instantiated, max_update_rects was always 1 for SwCompositor. Bug 1712871 inadvertently made it default to 0 on Linux, which is not what we want. This reverts the behavior for RenderCompositorSWGL so that max_update_rects is again 1.

Differential Revision: https://phabricator.services.mozilla.com/D116301

UltraBlame original commit: ca4db07f5cb064c108958eaabee16d41d7700d3a