Bug 1713485 - Remove unused migrationPercent pref. r=botond
With the migration completed, we don't need to keep this pref
around anymore. The profile migration code used to set the pref
value to 0 but since FF 86 the value of the pref didn't matter
as the code would always clamp it to 100. So for consistency
we eliminate that setting of the pref as well.

Depends on D116322

Differential Revision: https://phabricator.services.mozilla.com/D116323

UltraBlame original commit: 3a9cf4ad9344116ba51ef8d1d7937a9176747090