Bug 1517199 - part5 : handle the playback rate change on video track and add crash test. r=padenot
When we adjust the playback rate on the audio track, the audio clock time would be no longer align with the graph time.

Eg. playback rate=2, when the graph time passes 10s, the audio clock time actually already goes forward 20s.

After audio track ended, the video track would start to drive the clock time but the video clock time is align with the graph time, which means it would be smaller than the audio clock in that situation.

Therefore, we have to ignore the video clock time in that case. In addition, this patch also address the duration change on video frames based on the playback rate.

Differential Revision: https://phabricator.services.mozilla.com/D115035

UltraBlame original commit: b1945b56639872fa75cb1a0baf89f135ca35adb7