Bug 1713589. If the double tap default zoom in pref is 1 make sure to fully disable the functionality. r=botond
And then use this to disable the default zoom in in a couple of tests. The tablecell test specifically checks that we don't zoom in on tablecells. The hscrollable test checks that we don't scroll when double tapping and it would like to zoom out.

Differential Revision: https://phabricator.services.mozilla.com/D116457

UltraBlame original commit: 1c7cd1f3d1cb34e6c457c5a26d6e128d9a17f89f