Bug 1713679 - Split antitracking browser-chrome tests into 2 manifests; r=timhuang
When a test manifest becomes too big -- when the tests in the manifest run for
too long -- automated test scheduling becomes problematic. For instance, it may
not be possible to construct test "chunks" that run within the normally allowed
task time. This is now the case with the antitracking browser-chrome tests, at
least on some windows platforms.
This change creates a second antitracking test manifest, and moves existing
tests with "blocking" in the test name to that new manifest. Since many of these
tests are long-running, the new manifest uses about 2/3 of the test time,
perhaps leaving some room for growth in the existing, main manifest.

Differential Revision: https://phabricator.services.mozilla.com/D116508

UltraBlame original commit: d0be69629f279b4213c0f7eeed6dd2edd4a376e2