Bug 1714527 - [devtools] Remove unused serviceworker actor code in DevTools server r=nchevobbe,devtools-backward-compat-reviewers
The actor code for attaching/detaching service worker targets and to push/start workers should no longer be used when parentintercept is disabled because all the UI triggering this is disabled in that case.
Consequently we can remove the server side code supporting this. We still have to be able to list service worker targets when parentintercept is disabled, so there are still some codepaths using isParentInterceptEnabled in the server.

Differential Revision: https://phabricator.services.mozilla.com/D116850

UltraBlame original commit: ff6b83ae25d36b72421dba572ad5532a5b8ce922