Bug 1340422 - Part 8: Optimize attribute mapping by not parsing svg path string again. r=emilio
Unfortunately, we still have to convert the cpp version into rust version.
However, it's still better than parsing it again.

If we want to use the css parser to parse d attribute, we may need an
extra API to get the whole input as &str from Parser.
(expect_string() doesn't work because the SVG attribute value is not a quoted
string.) It seems Parser::slice() cannot get the whole input because we don't
know the end index of the whole input string. We should fix this in Bug 1714238.

Differential Revision: https://phabricator.services.mozilla.com/D115571

UltraBlame original commit: b6d681d6d66d426b28c74212736181f6aa350110