Bug 1708106 - [devtools] Fix testActor usage in browser_inspector_highlighter-05.js. r=ochameau.
The test opens the inspector and retrieves the testActor,
then a navigation occurs and testActor is used to check
if the highlighter works.
When server-side target switching is enabled, the testActor
gets destroyed during the navigation, and when it throws
when we try to use it in the new document.
To fix this, we only retrieve the testActor once we navigated.

Differential Revision: https://phabricator.services.mozilla.com/D117267

UltraBlame original commit: 4d1c24eeb026fb910c7dd2b690d2da9e42199473