Bug 1708760 - Build conditioned profiles during the tests. r=perftest-reviewers,AlexandruIonescu
This patch adds the ability to build conditioned profiles during a test (i.e. in-test, prebuilt, or in-task for CI). Using the `artifact:` prefix in --conditioned-profile will cause the conditioned profile to be downloaded from mozilla-central rather than being built locally.

The condprof package was also modified to prevent it from building and saving logs that are not very useful in local runs, preventing useless downloads or archiving, and to also be able to obtain the location of the test-built profile.

Differential Revision: https://phabricator.services.mozilla.com/D116520

UltraBlame original commit: 5e7be02a848720b2ba0381bf8874adc222c2058e