Bug 1681958 - Use eReplacedSizing to assert replaced elements for aspect-ratio. r=TYLin
It seems `eReplacedSizing` is much closer to the definition of replaced
elements (defined in [1]), and we override ComputeSize() for those
frames which are set eReplacedSizing or inline/table frames.

Perhaps we have to update the comments for the frame type bits later.

[1] https://html.spec.whatwg.org/multipage/rendering.html#replaced-elements

Differential Revision: https://phabricator.services.mozilla.com/D117372

UltraBlame original commit: 738bc12bd497cbb828435bbf761c5820c5cb8ffc