Bug 1715748 - Don't use NS_SAME_AS_FOREGROUND_COLOR on android for -moz-nativehyperlinktext color. r=agi a=orange
This shouldn't have a meaningful behavior change, as the default link
color right now is taken from the browser.anchor_color pref.

Returning this color from -moz-nativehyperlinktext makes no sense, and
it wasn't being handled correctly before: On GeckoView this color was
transparent.

The other patches in this bug cause NS_SAME_AS_FOREGROUND_COLOR to be
handled correctly, as currentColor, and cause a test failure on android
which asserts that -moz-nativehyperlinktext doesn't return the initial
value.

This color is supposed to be internal, but is has been historically
exposed to the web. Will try to unship these on a follow-up bug.

MANUAL PUSH: Orange fix on a CLOSED TREE

Differential Revision: https://phabricator.services.mozilla.com/D117674

UltraBlame original commit: b6ab0521071ea2945abcf988cbc619275aead05d