Bug 1692881 - part1 : let MDSM detect whether the video decoding is too slow and force to request the next key frame. r=bryce
If the decoded video MDSM received is already late, then it indicates that video decoding is too slow.

If we keep decoding but in the same speed, it's very likely all following decoded videos would still be behind the clock and need to be discarded.

The reader has its own mechanism to seek to the next keyframe, but that will only happen when the time of next keyframe is behind the given time threshold, which is the current time.

When the video has pretty long keyframe interval (eg. 5s), then that mechanism won't be triggered because the even if the decoded video frames are late, the next keyframe still far exceeds the current time, so the reader doesn't think it's necessary to skip.

Therefore, MDSM should force the reader to seek to the next keyframe directly, because it know both the decoded time of video frames and the current playback time.

Differential Revision: https://phabricator.services.mozilla.com/D117242

UltraBlame original commit: db7fd455cc6af66a965d3ee6f4b3c3a2d00f1a37