Bug 1715425 [wpt PR 29298] - URLPattern: Support relative search and hash constructor strings., a=testonly
Automatic update from web-platform-tests
URLPattern: Support relative search and hash constructor strings.

The main purpose of this CL is to support relative URLPattern strings
that begin with the search or hash component.  For example:

  new URLPattern('#baz', 'https://example.com/foo?bar');

This should result in an object with the following components:

  {
    protocol: 'https',
    hostname: 'example.com',
    pathname: '/foo',
    search: 'bar',
    hash: 'baz',
  }

To make this work, however, this CL also changes how URLPattern inherits
information from a base URL.

Previously we would inherit all base URL components *except* for search
and hash.  The idea was that the URL constructor never inherited these
from base, so we shouldn't either.

That reasoning, however, was flawed.  The URL constructor will propagate
components if they are not specified in the relative URL.  So a relative
`#baz` will get the base URL search component propagated.  This does
not happen for a relative `/foo` because this kind of relative URL is
equivalent to `/foo?#` which set the search and hash to empty string
preventing them from being taken from the base URL.

The impact of this change beyond the relative URL story above is that
a base URL in something like this:

  new URLPattern({ pathname: '/foo/bar',
                   baseURL: 'https://example.com' });

Will no longer leave the search and hash components as wildcards.
Instead the search and hash components will be propagated from the base
URL; which is the empty string value in this case.

While this may lose some flexibility, it brings search and hash into
alignment with how all the other components work with base URL.  So in
some ways its a simplification of the API.

Bug: 1141510
Change-Id: Ia3ddb19e217fa45ed1b30ea97bdd96843b7201e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2934920
Commit-Queue: Ben Kelly <wanderviewchromium.org>
Reviewed-by: Jeremy Roman <jbromanchromium.org>
Cr-Commit-Position: refs/heads/master{#890601}

--

wpt-commits: bdef08b00f2686e7db41d2bbfb40eeb7c7201881
wpt-pr: 29298

UltraBlame original commit: c4ca6d0d2b22642d38575716d0490909fab235e6