Bug 1716790 - Part 5: Switch parameter order in KeywordToString. r=dminor
Place `JSContext*` as the last argument to indicate this function is infallible.

Most functions which have `JSContext*` as their first argument are fallible, so
a SpiderMonkey convention is to move `JSContext*` as the last parameter for
infallible functions.

Depends on D118022

Differential Revision: https://phabricator.services.mozilla.com/D118023

UltraBlame original commit: 0bd5153b7662ec2638e582129028e2529e8e7d4b