Bug 1710605 - Consistenly serialize list values as unquoted where possible, r=whimboo
We were previously constructing some expected values as a list of
StringNode objects due to make_value_node incorrectly calling
make_node for list items. This is wrong as StringNode types can only
be on the lhs of an expression (e.g. if os == "linux") but never on
the right hand side (which is always a ValueNode or ListNode with ValueNode
children).

StringNodes are always serialized with quotes and this was causing
values to flip flop between quoted and unquoted forms in some cases.

Differential Revision: https://phabricator.services.mozilla.com/D115259

UltraBlame original commit: c12ba7e19c6f5bc427bb3c49639801c7200a262f