Bug 1708007 - Make sure we always reflow flex items that have a continuation.  r=dholbert
The root cause of the crash was that FlexItem::NeedsFinalReflow
returned false even though the item had a non-empty next-in-flow.
This made the flex container skip the item's reflow and consider
the item's reflow status as COMPLETE, which triggers the removal
of the container's own next-in-flow, which causes the assertions
and eventually the crash.

Differential Revision: https://phabricator.services.mozilla.com/D118420

UltraBlame original commit: a595bed7a99e7135f94a19763a8628cf7877188a