Bug 1717606 part 4 - Remove setFlag's GenerateShape argument. r=jonco
This was a footgun because the argument defaults to GENERATE_NONE, meaning it can change
a dictionary shape's object flags without allocating a new shape.

We don't rely on this anywhere for perf reasons so it's simpler and safer to always
allocate a new shape.

Depends on D118461

Differential Revision: https://phabricator.services.mozilla.com/D118462

UltraBlame original commit: a78b62f28c6a65f176059a75328f320d698e399c