Bug 1710024 - Avoid defineReuseInput for compare-and-select on ARM64.  r=yury
On ARM64 we want different register allocation, so break that case
apart from the others and specialize it.

This patch keeps limiting the compare-and-select code to the case
where the comparison has type i32 and the value operands are also of
type i32, as that fits the code generation pattern on the other
platforms.  But ARM64 is now set up for handling other comparisons and
other value types.

Differential Revision: https://phabricator.services.mozilla.com/D117915

UltraBlame original commit: ec5325d760575a47291a4d99b2ee01353d4e799b