Bug 1718166 - Ensure glyph scale quantization is >= 1 r=gfx-reviewers,jnicol
Now that picture cache tiles are raster roots, their scale is
expressed via the device-pixel scale. This means the calculated
scale can be too small when zoom factor is < 1 and a pinch-zoom
is currently active.

This is not a complete fix, it's an interim improvement until we
implement a proper fix. A complete fix will need to include the
device-pixel scale in the quantization process. However, this will
take some time to get right, since the existing code for selecting
a glyph size is quite complex due to different permutations and
how that code has evolved.

Differential Revision: https://phabricator.services.mozilla.com/D118781

UltraBlame original commit: dd947fcc5fa8c7060ec8dc25909b633789e815de