Bug 1703227 [wpt PR 28381] - Start creating a proper client for WebDriver BiDi, a=testonly
Automatic update from web-platform-tests
Start creating a proper client for WebDriver BiDi (#28381)

Add a webdriver.bidi.client module with a BidiSession class that
implements only the BiDi part of the protocol. An ordinary Session may
have an associated BidiSession if the session was created through
HTTP. To make this easier, an enable_bidi flag is added to the
top-level session which automatically sets the capabilities required
to enable BiDi.

The session allows sending commands, and registering handlers for
events.

In addition to the session class, this adds some intial work on
creating a structure for commands, with a BidiModule abstract class
that can be used as the basis for command implementations and an
command decorator that handles actually sending the command and
waiting for the result.

Co-authored-by: Maksim Sadym <69349599+sadym-chromiumusers.noreply.github.com>

--

wpt-commits: a38612f39e7752c3532080c7b482e3c8627f7f76
wpt-pr: 28381

UltraBlame original commit: afd9ed99cc5099840a348f575929f90a8380b9c7