Bug 1716410 - Regenerate the target element for OMTA tests. r=boris
nsIDOMWindowUtils.getOMTAStyle is for querying an in-progress animating value
running on the compositor thread. The key to query the value is stored in the
target nsIFrame of the animating element, and the key is basically based on
the display item (type) in question, DisplayItemType::TYPE_TRANSFORM in cases of
transform animations, etc. So if getOMTAStyle gets called for an element
where
 a) an animation is running on the compositor
 b) a new animation is just created

then, it's possible that getOMTAStyle queries a)'s animation value instead of
b)'s.

Differential Revision: https://phabricator.services.mozilla.com/D119127

UltraBlame original commit: 4b8b0a777c08d61198381676aba3c6cc45d5c182