Bug 1681692. In FindViewContaining we are only crossingZoomBoundary if we start in visual coordinates. r=botond
In the most common case the first call to FindViewContaining has aRelativeToView = aView = the root view of the root content document, and we start in visual coords. Then we convert to layout coords and call FindViewContaining recusively with aRelativeToView = the root view again, aView = a child of the root view and working in layout coords. And we still satisfy the conditions to set crossingZoomBoundary to true, which is incorrect because we already crossed. We need to check that we are also in visual coords.

Differential Revision: https://phabricator.services.mozilla.com/D105438

UltraBlame original commit: 6c675319da0d79e81856ea7f51efcfbe0110580e