Bug 1719180 - Set restrictSource to the first restriction token, dropping support for combined tokens. r=mak
This bug was introduced because UnifiedComplete was only filtering tokens when the queryContext contained a restrictToken. UrlbarProvidersManager was only setting queryContext.restrictToken when a source restriction token was typed (i.e. not including $ and #). This meant that # and $ were never filtered from the search string. This patch now sets restrictToken to whatever the first token is, including # and $. This ensures UnifiedComplete will always filter tokens when a restriction token is typed.

Differential Revision: https://phabricator.services.mozilla.com/D119197

UltraBlame original commit: 09febaa9cec98e5748e70e31c47904083c1f597e