Bug 1718102 - Extract ScriptFlags accessors to be reused. r=arai
The CompilationInput currently let the consumer reach to the lazy field, to
reach accessors implemented on the BaseScript. As we are trying to remove direct
usage of the lazy field, and later abstract between a ScriptStencilExtra or a
BaseScript, we have to make it simple to add these accessors to other
structures.

This change adds RO_IMMUTABLE_SCRIPT_FLAGS and RW_MUTABLE_SCRIPT_FLAGS macros to
define all the accessors to immutableFlags_ and mutableFlags_ fields. These
macros are parametrized with the field name or the expression used to access the
field.

Differential Revision: https://phabricator.services.mozilla.com/D119105

UltraBlame original commit: 9f48019d06ea943bfd3898255c4b1e609cfd187f