Bug 1718102 - Abstract CompilationInput lazy fields behind accessors. r=arai
This change adds accessors to reach the `BaseScript* lazy` field as well as all
information taken from it. The intent being that most information used from the
BaseScript are taken out of the ScriptStencil and ScriptStencilExtra structures,
and that we can later use these in-place where the lazy field is used.

Differential Revision: https://phabricator.services.mozilla.com/D119106

UltraBlame original commit: b4978b94e6054328409c38ba4c365c1132fef20c