Bug 1720070 - [remote] Only apply common preferences on the first call to applyPreferences r=webdriver-reviewers,whimboo
Depends on D119610.  After fixing the error with targetList.destructor, calling CDP::stop() without having called CDP::start() before will trigger another failure.
This is because it will trigger a late load of RecommendedPreferences which will try to automatically initialize preferences
As we discussed about moving the init to be only performed during a call to applyPreferences, I am doing this change here to avoid the error.
This will not yet allow to call this from BiDi, or it would have to call it with an empty `preferences` map.

Differential Revision: https://phabricator.services.mozilla.com/D119611

UltraBlame original commit: a37ae1e32765c64975e7b6a9140ac0f562b8dda4