Bug 1720772 - Upgrade sysroot libstdc++ to 7.5.0-6. r=firefox-build-system-reviewers,mhentges
7.3.0-1 was used because it was the last 7.x that provided the libstdc++6
package. Unfortunately, once libstdc++ is removed from the clang
toolchain, which makes the one from the sysroot used instead, it causes
an ASan issue which goes away when using a more recent version (which the
libstdc++ from clang was: it was 7.5).

So we take the last 7.x that exists, and tweak it so that it generates
the necessary libstdc++6 package.

Differential Revision: https://phabricator.services.mozilla.com/D120012

UltraBlame original commit: 196b9fb506b770fa9cf869284580cd62f9ed4714