Bug 1715230 part 3: Change AccessibleWrap::DispatchTextChangeToHandler to a static method accepting an Accessible*. r=morgan
In a subsequent patch, we want to call it with a RemoteAccessible instead of a RemoteAccessibleWrap.
As an instance method of AccessibleWrap, this wouldn't be possible because AccessibleWrap is a subclass of LocalAccessible.
Strictly speaking, there's no good reason for this to be part of AccessibleWrap at all any more, but I'm not dealing with that here.

Differential Revision: https://phabricator.services.mozilla.com/D117527

UltraBlame original commit: e7acb70d7295da077bfa4c4d20a7c5f36e9b4074