Bug 1720844 - Run shutdown servers concurrently, r=whimboo
Instead of requesting shutdown of each server subprocess and then
waiting for that process to terminate before moving on to the next
process, make all the shutdown requests and then wait for all the
processes to exit. This ought to reduce the typical shutdown latency
for the test harness.

Differential Revision: https://phabricator.services.mozilla.com/D120061

UltraBlame original commit: d08db81bb25084903ea630289dce674d5b05ef7c