Bug 1721375 - [devtools] Remove WebConsoleFront getString method. r=bomsy.
The method isn't related to the webconsole actor, so we can remove it, and
alongside with it, the longString method and _longStrings
property which are no longer useful.
The only customer of this switches to the shared `getLongStringFullText` method
instead.

Differential Revision: https://phabricator.services.mozilla.com/D120354

UltraBlame original commit: c0ff235df051250281ac3e2baac56ea6662f1168