Bug 1721036 - Delay buffer commits until the surface is ready,r=stransky
When the `MozContainer` surface is not yet ready in `Commit()`,
we need to queue the commit.
Similar to `nsWindow` and `WaylandVsyncSource` we can do so by
registering via a callback via `add_initial_draw_callback()`.

This required some logic changes which let to clean ups and
simplifications in the whole file (I'm sorry it got so big).

Differential Revision: https://phabricator.services.mozilla.com/D120429

UltraBlame original commit: 50735532fcdf8aef74b1a00d1e4aa53362d5fb7e