Bug 1722031 - Tweak accent-color foreground computation and let it ride the trains. r=mstange
The accent-color computation right now chooses between black and white,
which is not ideal.

I tried to make it so that authors could choose the foreground colors in
the linked CSSWG issue from the comment, but that didn't go anywhere.

I think choosing a contrasting color that is in-line and contrasting
enough with the accent-color chosen by the page when darkening is better
than just black or white.

If we want the black-or-white behavior we can just change
layout.css.accent-color.target-contrast-ratio to something large enough.

https://accent-color.glitch.me/ is a nice playground to see this patch
in action.

Differential Revision: https://phabricator.services.mozilla.com/D120723

UltraBlame original commit: 7a7d9a6cc5134befe051de531f8f04c1b76e0f7b