Bug 1693176 - Destroy frames of contents inside panels. r=Gijs
Panels don't rely so much on the popup manager code to deal with
menuitems etc, so this is trivial and fixes the issue as reported.

I think it's pretty unlikely to have animated stuff in menulists /
menupopups, so this should get us most of the way.

Differential Revision: https://phabricator.services.mozilla.com/D105550

UltraBlame original commit: b3fab819b622a2510d59378047d3fcb79b969834