Bug 1719194 part 0 - Add a new TaggedScriptThing to be used as a placeholder. r=arai
In upcoming patches, we are converting the BaseScript::gcThings to a fake
stencil vector of TaggedScriptThingIndex. As we transition progressively, some
kind are not yet represented and therefore not used. To avoid encoding bad
indexes, we create a new OpaqueThing type which is meant to represent all things
we do not yet care about.

Differential Revision: https://phabricator.services.mozilla.com/D120230

UltraBlame original commit: bd6a7fdc412e29aa76b37a2c86fb1d990d2a4a99