Bug 1719194 part 3 - Remove Rooted<BaseScript*> from the FullParseHandler. r=arai
In order to be able to parse from an existing Stencil instead of a BaseScript*,
we have to replace all references to BaseScript* data by abstract data which can
be computed from both, and without the need for allocating GC objects.

This patch is the last change to the FullParseHandler to abstract over a GC
object (BaseScript*) and a Stencil. It removes all references to the
Rooted<BaseScript*> and replaces them by a simple boolean flag which indicate
whether we are reusing data from a pre-parsed function, instead of doing a full
parse which includes resolving bindings and parsing inner functions.

This patch also include a renaming of functions named `canSkip*` into `reuse*`,
to better highlight why we are doing differently.

Differential Revision: https://phabricator.services.mozilla.com/D120233

UltraBlame original commit: 09725af9e55b09d0d672bb2e4c6f6b6233235f3c