Bug 1722397 - Add a way to use the native accent color for selection and form controls on Android. r=agi
I tied it to the use-theme-accent bit in non-native-theme just for
convenience (so that form controls just react to this).

This works nicely, but I didn't turn this on by default because the
accessiblecaret images are hardcoded-blue pngs, and they look ugly
without being the same color as the native accent.

Differential Revision: https://phabricator.services.mozilla.com/D120898

UltraBlame original commit: 7e23175225b6c0edb5e2f70d298d04cab9c315fe