Bug 1711073 - wasm: Move memory constants to WasmMemory.h. r=yury
This commit moves the constants used for memory to WasmMemory.h.
This is tricky because there's a cycle of constants that aren't
related to memory in WasmTypes.h that needed to be broken. This
is done by moving codegen related constants to a new
WasmCodegenConstants.h file. Any static assert that can be
hidden is moved to WasmMemory.cpp.

There should be no functional changes.

Depends on D120881

Differential Revision: https://phabricator.services.mozilla.com/D120882

UltraBlame original commit: 267cfdbbb88ae4e34cd49d12e16370b7068e1bcf