Bug 1722501 - [devtools] Migrate styleeditor to will-navigate. r=nchevobbe
We used to handle navigation with a subtle and complex combination of target-available + dom-loading.
Just because we were missing will-navigate event in some cases.
But will-navigate is now reliable and is fired in all types of navigations.

Differential Revision: https://phabricator.services.mozilla.com/D120961

UltraBlame original commit: e0ad3156c4a1119b66e5cfc45df116965129e03e