Bug 1688794 - Instantatiate self-hosting C++ intrinsics directly from JSFunctionSpec. r=jandem
Instead of defining these intrinsics directly on the self-hosting global and
cloning on demand, directly use the JSFunctionSpec in the target global. This
removes the need to pre-define these on the self-hosting global. Use a binary
search to find the intrinsic on first use in a realm and then cache it on the
intrinsics holder object so that the lookup only happens once per realm.

Differential Revision: https://phabricator.services.mozilla.com/D120539

UltraBlame original commit: f370efcfeb7632451debc229203a0c0f93db9eae