Bug 1698030 - Remove a redundant warning icon. r=preferences-reviewers,Gijs
The only difference between the icon that was removed and the one kept is that
the removed one has a default fill color in the SVG. This meant everywhere the
icon was replaced, we have to make sure that a fill color is defined in CSS.
In a few cases, that necessitated adding a new class. In a few others, colors
were already being defined for the icon, so there was no need to add any here.

Differential Revision: https://phabricator.services.mozilla.com/D121030

UltraBlame original commit: 5a5e93469ad90107785d294e03dca70b68da27f8