Bug 1688794 - Add mapping from atom to self-hosting ScriptIndex. r=jandem,arai
On the JSRuntime, save a mapping from (permanent) JSAtoms to corresponding
self-hosted stencil ScriptIndex. We store both the ScriptIndex and the index of
the next top-level functions. Scripts in this range are sub-scripts of the
target. This is used later to instantiate self-hosted builtins on demand.

Differential Revision: https://phabricator.services.mozilla.com/D120541

UltraBlame original commit: aa5317d3a439d937e48ddace814675c7f75dee52