Bug 1720746 - Preload bcryptPrimitives.dll in the main thread of GMP.  r=bobowen
The call to `RtlGenRandom` delay-loads bcryptPrimitives.dll.  The GMP process,
however, cannot load bcryptPrimitives.dll after process launch because its process
token is restricted.  This is not a problem normally because bcryptPrimitives.dll
is loaded in early stage when the main thread still has a non-restricted impersonation
token.

With ASan, however, the first call to `RandomUint64` happens in a non-main thread,
and it fails because the thread is not impersonated.  We have PreloadLibs to mitigate
this kind of problem, but in this case adding bcryptPrimitives.dll to the list does
not help because the call to `RandomUint64` happens before we load PreloadLibs.

The proposed fix is to explicitly call `RandomUint64` in the main thread before
any call to `RandomUint64` in the process.

Differential Revision: https://phabricator.services.mozilla.com/D121203

UltraBlame original commit: d74caf9f8941e1a4602282f573d6453155152fcb