Bug 1652560 - Remove the unused "gecko_profiler" feature from servo r=emilio
This was being used when we had special code for gecko profiler in the servo
codebase but we just removed the last one. This is safe to remove now. The
"enabled" feature in the gecko-profiler crate is being controlled by
gkrust-shared directly now.

Differential Revision: https://phabricator.services.mozilla.com/D120796

UltraBlame original commit: 99031ce526ea16d3b84d2e3a250ac4283485dd4b