Bug 1719194 part 2 - Use stencil-like data inferred from BaseScript for skipping functions. r=arai
When delazifying with a `CompileLazyFunction` call, we reused
BaseScript::gcthings to skip over information already provided by the
`SyntaxParseHandler`, such as functions and closed-over bindings.

This change adds a stencil-like scriptData and scriptExtra spans, which are
created from the JSFunctions of BaseScript::gcthings() with the intent of later
using the same API for mapping input data from a Stencil without having to
allocate GC objects for eager delazifying functions.

`CompilationState::init` is modified to call the convertion from the
BaseScript::gcthings() to the `CompilationInput::cacheScriptData_` and
`CompilationInput::cacheScriptExtra_`. The logic of the `FullParseHandler` is
updated to work with the gcthings provided by the `CompilationInput` in a
similar way.

The function `skipLazyInnerFunction` now takes its inputs from the
`ScriptStencil` and from the `ScriptStencilExtra`, thus no longer relying on
`BaseScript::gcthings()` nor on a `JSFunction*` to extract this information
from. `FunctionBox` initialization reflect these changes by taking these 2
inputs as argument as well.

Note that the `ScriptStencil` and `ScriptStencilExtra` provided by the
`CompilationInput` do not reflect the full content of each script, but the bare
minimum used by `skipLazyInnerFunction`.

Differential Revision: https://phabricator.services.mozilla.com/D120232

UltraBlame original commit: c35855c033f871d3331efeeb1eb5d59b1cdd4185