Bug 1721262 - Make sure to reflow the line again if we bail out of reflow due to break-inside: avoid. r=dholbert
The issue is that we mark the second input as dirty when it gets the
reflow request (this is as expected), but then during multicol reflow
everything seems to go smoothly, but we hit this code:

  https://searchfox.org/mozilla-central/rev/00977c4e37865a92f1c15572ae4aea90e934b25b/layout/generic/nsBlockFrame.cpp#3745

And bail out, leaving the line marked as not dirty, and as such the
stuff inside the page-break-inside: avoid dirty.

When it changes further, we think it's already dirty so don't bother
adding it to the dirty root list anymore and the text frame remains with
its original tiny width.

Differential Revision: https://phabricator.services.mozilla.com/D121869

UltraBlame original commit: 981cd4e1f9b2f3f4dbf956f3f6a18f58920d46ab