Bug 1724626 - set test expectation to unknown pseudo-element as not throwing outside Nightly. r=emilio DONTBUILD
Bug 1723921 updated the code in line with the spec to throw for invalid
pseudo-elements. This update is only enabled for Nightly, the expectation to
fail to throw for an unknown pseudo-element must be added back for Beta and
Release.

Differential Revision: https://phabricator.services.mozilla.com/D122068

UltraBlame original commit: 2743d27b31a787d79ccfbabdf20237c6aef1d9fb