Bug 1723715 part 2 - Remove the RUNTIME_CODEGEN_ENABLED slot. r=jonco
Originally this slot was used to cache the return value of the callback. However
that was changed at some point, now the slot is only set to TrueValue when there
is no callback installed. It's now better and simpler to remove the slot.

Depends on D121981

Differential Revision: https://phabricator.services.mozilla.com/D121982

UltraBlame original commit: d163b4821d8008641c8cd84c515e649b91ba900d