Bug 1341265 - Part 4: Remove unnecessary restriction for branch64 with Address operand. r=iain
ARM64 and MIPS64 already support this mode. And ARM32 and MIPS32 can simply take
the same approach as x86.

The next part will add another `branch64` method and that method supports
`Condition::Equal`. And instead of adjusting the comment for `branch64` when
`Equal` is or is not supported, it was easier to just support it everywhere.

Differential Revision: https://phabricator.services.mozilla.com/D118970

UltraBlame original commit: e5112a66b77f575f80d6af6ea2fdd85312e0cca0