Bug 1719886 - Support translated no-repeat patterns in canvas FillRect. r=jrmuizel
This is a bandaid designed to cope with a subset of the encountered transforms that
may be applied to a pattern inside FillRect. The existing code did not consider the
transform at all that might be applied to a pattern when it tried to manually clip
the geometry to a no-repeat pattern.

This manual clipping only seems to occur in FillRect, whereas no-repeat patterns
are not properly handled anywhere else in canvas entry-points. To fix this more
generally requires a clamp-to-transparent tile mode (like Skia's decal or Cairo's
none) which we can't currently rely upon with our D2D support. However, that is
much beyond the scope of this temporary workaround.

Differential Revision: https://phabricator.services.mozilla.com/D122186

UltraBlame original commit: ffc7756de0471d2a9d2a15c874f1a408fdd5b5d7