Bug 1676524 - fix intermittent failures in browser_experimental_features_filter.js due to fluent races r=emilio
The test overwrites a fluent-originating text content with its own,
but doesn't clear the fluent data-l10n-id attribute. As a result,
if it tries to write the textContent before the data-l10n-id pass
from fluent has run, fluent overwrites the textContent before we
do a search, and then the rest of the test gets bogus results
because it matches against the wrong text content.

This patch fixes this by removing the data-l10n-id attribute so
the fluent pass doesn't try to put text in these elements
anymore.

Differential Revision: https://phabricator.services.mozilla.com/D122659

UltraBlame original commit: 7088287866ca700ada7b59506a966583c8afbb27