Bug 1725291 - part 1: Make `HTMLEditor::AutoDeleteRangesHandler::DeleteUnnecessaryNodesAndCollapseSelection()` try to delete same text node twice r=m_kato
It tries to delete a text node if it's invisible both selection start and
selection end.  However, the selection range may be in a text node.  If it's
an invisible, the node is deleted at first call of
`DeleteNodeIfInvisibleAndEditableTextNode()`, but it keep trying to remove
again with same method.

Differential Revision: https://phabricator.services.mozilla.com/D122561

UltraBlame original commit: d875e0f7413e48199df7c8df5acc075c1d96cd13