Bug 1725894: Support IEnumVARIANT for RemoteAccessible when the cache is enabled. r=morgan
As part of this, I changed RemoteAccessible::RemoteChildAt (and thus RemoteAccessible::ChildAt) so it doesn't crash when passed an invalid child index.
Our EnumVariant implementation relied on this with LocalChildAt.
I think it makes sense for LocalChildAt, RemoteChildAt and ChildAt to be consistent in this regard.

Differential Revision: https://phabricator.services.mozilla.com/D122681

UltraBlame original commit: 9e3ff39009b1640d420a6a284fc85ed645a17aa1