Bug 1711912 - Move key system name constants into their own header. r=alwu
Move constants and fix up includes following doing so.

This is motivated by that future patches will want to use these constants in the
clear key CDM. If we don't move these constants and try to include the existing
VideoUtils in the clear key CDM this will break the build (because the clear key
CDM uses a restricted subset of the stuff we get in Gecko).

Differential Revision: https://phabricator.services.mozilla.com/D122631

UltraBlame original commit: 0d4ae171202c70ced523b533717ed1b5b6e8ccd8