Bug 1726463 - [devtools] Avoid error in NodeFront remoteFrame() if devtools.browsertoolbox.fission=false r=bomsy
As explained in the bug, the test test_inspector-dead-nodes.html exercises an edge case which is probably irrelevant now.
It aimed to check that you could call APIs on target scoped fronts during target reload.

Eg, get the children of a node, while you reload a page. However with SSTS, we now destroy the target, so all fronts are destroyed.
This test should probably be migrated to a commands test, and we should define what should be the behavior of such a command during reload.

However in its current shape the test is doing the following:
- get walkerFront1 and nodeFront1 for target1
- reload the page
- wait for the new target target2 and get walkerFront2
- call walkerFront2.children(nodeFront1)
- check there are no errors

This call doesn't really make sense however. walkerFront2 and nodeFront1 are fronts from different targets and nodeFront1 is already destroyed. Here it happens to crash when trying to use a getter on nodeFront1 and it would be nicer to not throw anyway, hence the proposed fix.

But the test should still be reviewed.

Differential Revision: https://phabricator.services.mozilla.com/D123138

UltraBlame original commit: ba789be08c4e596f7d615ddc6290afcebd0f7ca3