Bug 1726532 - part 2: Make `mozSpellChecker::SetupDoc()` treat succeeded state as error if valid offset/length are not available when they are required r=m_kato
`TextServicesDocument::LastSelectedBlock()` may return `NS_OK` with
`UINT32_MAX` for both offset and length of the block.  The assertions causes
the new test not runnable in debug build.  So, we should make the situation
as an error case if they are required.

Ideally, we should fix the odd behavior, but it's not safe for uplift.
Therefore, this patch just adds the wallpapers.

Depends on D123282

Differential Revision: https://phabricator.services.mozilla.com/D123283

UltraBlame original commit: 55fa746f2962617543b2db5e28629efb059283de