Bug 1724031 - Part 11: Prevent gecko profiler from observing moved function pointers r=jandem
Changing functions to use fixed slots creates a problem when the profiler
tries to get the script for on-stack functions. Unfortunately this can happen
during minor GC after the functions have been moved but before the profiler
data pointing to them has been updated.

The function data itself is mostly intact, but the object header (used for the
shape) is not valid. As far as I can see this only affects the debug assertions
around checking that slot numbers are valid. However the simplest fix is to
suppress the profiler during the intial part of minor GC, until the roots have
been traced (which will update the profiler data).

The patch moves the code to trace the roots in a minor GC to a separate method
and adds a guard to suppress the profiler.œ

What do you think?  Alternatively we would have to add a load of 'unchecked'
methods to a buch of accessors on JSFunction, and to NativeObject to get fixed
slots without touching the shape.

Differential Revision: https://phabricator.services.mozilla.com/D123642

UltraBlame original commit: b956c169063fb72de82b2529b68e10ab01c173fb