Bug 1727814 - Drop mVisibleRegion from box shadow items. r=miko
mVisibleRegion was added in bug 530686 as an optimization for when the invalid
region was vertically discontiguous. WebRender doesn't call ComputeVisibility
and so never had an updated mVisibleRegion and so never used it during drawing.

WebRender fallback drawing has limited invalidation support and currently only
passes a rect to ComputeVisibility so it also wasn't taking advantage of this
optimization. This is also true of nsDisplayList::Paint.

In summary, we're not currently using this optimization and supporting it bloats
the display items. If we were to reintroduce it, I think it would make more sense
for the paint region to show up in the Paint function and not be stored in the
display item.

Differential Revision: https://phabricator.services.mozilla.com/D123791

UltraBlame original commit: e4aff5636c7b83fe6f65de55df18f80c178cefc9