Bug 1696251: Change method names in mach commands to avoid scoping issues when de-classing. r=mhentges
The affected functions would cause an UnboundLocalError if they are moved
to module-level, because using an assignment inside a function makes a
variable local in Python, making it illegal to have a local variable name
be the same as a function name that is called within the same function.
The adopted solution is to rename those functions in preparation for the
future de-classing.

Differential Revision: https://phabricator.services.mozilla.com/D123867

UltraBlame original commit: 3602cbaf8893a15b047d2c6d011078b6cad4d3f7