Bug 1727877 - Only allocate JS frame buffer when thread is being profiled - r=canaltinova
The buffer was previously allocated as soon as a JSContext was present, meaning that some memory was used even when the profiler was not running, which is the case most of the time for most users.
This saves some memory, at the cost of having to lock the per-thread ThreadRegistration data when sampling a thread with a JSContext. This should have low contention risk, because it's mostly used when sampling on the thread, while the periodic sampler uses its own buffer so it doesn't need to lock the per-thread data.

Differential Revision: https://phabricator.services.mozilla.com/D123910

UltraBlame original commit: e6242f4c09c53bfddb41c412c3778f1b63c81e44