Bug 1715232 - Add ICU4X license sha256 to vendor_rust.py r=glandium,mhoye
This adds the sha256 for the ICU4X license as a variable to vendor_rust.py,
so it can be used as we add ICU4X crates. Since there are several projects
ongoing that want to use ICU4X, I think it makes sense to get the license
review done once, early, so that whoever is ready to land code that uses
ICU4X first is not blocked waiting on license review.

Differential Revision: https://phabricator.services.mozilla.com/D123776

UltraBlame original commit: 1ae51d10e2cb32597960a6d6fca84192407d0727