Bug 1728442 - P1: Make browser description test rely on name/desc change events. r=Jamie
By using a real image, and not a broken one, we avoid the recreation
that happens when the image gains and loses an alt. We can then rely on
name/description changes.

This is a more accurate test that doesn't rely on tree mutations.

Differential Revision: https://phabricator.services.mozilla.com/D124233

UltraBlame original commit: b8103adfd8d0e268fcf28ce1e91b3ff84df568ec