Bug 1728754 - Compute layer order during CascadeData rebuild. r=boris
For that, deal with children in add_rule recursively, to keep the
current layer name up-to-date in block layer rules.

This is not the final version of this code, as right now something like
this:

  layer A {
    ...
  }

  layer B {
    ...
  }

  layer A.A {
    ...
  }

Would give A.A more priority over B, which is not correct. There are
tests for this incoming in wpt sync and such, but that can be tweaked
later.

Differential Revision: https://phabricator.services.mozilla.com/D124335

UltraBlame original commit: 56c800f59664d6e5a811fd3f959af66e574336cd