Bug 1720958 - Fix DocumentLoadListener leak that occurs when browser.tabs.documentchannel.parent-controlled is enabled, r=nika
Without the pref enabled, DocumentLoadListener is destroyed when
DocumentChannelParent is destroyed. With the pref enabled, we create
DocumentLoadListener when we are trying to do a parent controlled navigation,
but it does not get destroyed anywhere. To do that, we can disconnect the
DLL listeners before we trigger redirection to the real channel and resume the
load.

Differential Revision: https://phabricator.services.mozilla.com/D122131

UltraBlame original commit: a55c5e73da51ab8fd9f04a434fc28e80009a7b05