Bug 1729439 - Split gray marking into separate actions r=sfink
This splits the markGrayReferencesInCurrentGroup action into
markGrayRootsInCurrentGroup and markGray, which means we don't need the extra
state to check whether it's the first time it's called. Now
markGrayRootsInCurrentGroup will run once, and then markGray will run to
process the mark stack. If we exhast the budget, we will re-enter in markGray
and won't try to mark the roots again.

Depends on D124780

Differential Revision: https://phabricator.services.mozilla.com/D124781

UltraBlame original commit: fcf04373f799402b1abe751d2415702dc5a26838