Bug 1724650 - part 1: Make `HTMLEditUtils` scans collapsible white-spaces with referring the style r=m_kato
This makes it support `white-space: pre-line` which only preserve linefeed
characters.  However, the APIs are not all of what we need to change.
Therefore, this causes hitting assertion hits due to different result from
some other APIs.  So patching only with this does not work.  The automated
test results will be updated later.

Note that it's **not** the goal of this bug that we don't support white-space
handling in text nodes whose `white-space` is `-moz-pre-space`.  It replaces
linefeed characters with white-spaces then, preserve all white-spaces.
However, it's not standardized, and it's available with `xml:space="preserve"`
in SVG if web developers don't use the vender prefixed value, and finally,
supporting it makes our white-space handling in editor more complicated
than applying the following patches.  So, we don't need to do it at least
for now.  Therefore, this and following patches does not assume that there
are no cases that only new lines are collapsible.

Differential Revision: https://phabricator.services.mozilla.com/D124555

UltraBlame original commit: a98cc0fab376d693cab126d8e41d7bd11c8a0c3b