Bug 1688888: decrease ui.*.radius.topmm on Android. r=botond
On one website, the touch input fluffing made an audio scrubber so large that it
covered other playback controls, making them inaccessible. This patch addresses
this problem by changing the top fluffing radius so it more closely matches the
bottom radius.

In our investigation, we found the top radius was added asymmetrically to the
bottom radius without an explanation. We speculate the bottom bias was added
because when you touch the screen, you tend to aim with the top part of your
finger, near your fingernail, and sometimes the bottom part of your finger will
also touch the screen, fat fingering it: if true, we would fluff the event less
below the touch point (which is what the code was). I vaguely remember this fat
fingering problem in the early 2010's but I haven't thought about it in recent
years and I didn't notice this problem in my testing for this patch – perhaps
device touch handling has just gotten better. Therefore, we feel safe to remove
the bias.

Differential Revision: https://phabricator.services.mozilla.com/D124267

UltraBlame original commit: aa1e0da893034d02fbcee39cc63428a238982b9f