Bug 1729420 - part1 : compare the output to the inputs in order to determine if it's an invalid first sample or not. r=bryce
The old way to guard the first sample is incorrect because of following reasons.

1. the demuxed samples for h264 are stored in DTS order
That mean the first sample we gave MFT decoder might not be the earliest (have latter PTS) So it would incorrectly discard some useful frames.

2. it conflicts with the seek threashold set from the format reader
We will set a threshold to guard the first sample only if MFT manager doesn't have the seek threshold. If the MFT has set that before, the old mechanism would set the guard threshold based on the timestamp of non-first input sample, which also causes discarding wrong frames.

Therefore, this patch use a new method to distinguish if the first output frame is invalid. If the FIRST output has zero timestamp and it doesn't belong to any input we gave to the decoder, then it's highly possible an invalid frame which should be discarded.

Differential Revision: https://phabricator.services.mozilla.com/D124858

UltraBlame original commit: 30a53fd882ef320f82c02aa46aab3862e6f1986c