Bug 1730188 - wasm: Check input function to WA.Function constructor. r=lth
The input function to WA.Function is currently restricted to not
be an exported wasm function. This is being debated in the proposal,
but is currently being asserted in our code and so should be checked
for explicitly.

In addition, some OOM checking, standardized naming, and general
cleanup was added to WA.Function.

Differential Revision: https://phabricator.services.mozilla.com/D125228

UltraBlame original commit: ff58a6be4b4dc68a78235087bc7a12d9a8b1f2b8