Bug 1730825 - Build NSIS as a relocatable binary. r=firefox-build-system-reviewers,andi
The reason the error mentioned in build-mingw32-nsis.sh happens is that
the default mode NSIS builds in is a fully-installed mode, where it
hardcodes the locations of its data files. This is why nsis needs to
be used from the same place it's built for. But there's another mode,
enabled with NSIS_CONFIG_CONST_DATA_PATH=no, that makes it relocatable,
and makes it find its data files relatively to the nsis binary.

However, there's a bug in the nsis build scripts, which makes the nsis
binary installed in the destination directory instead of a bin/
subdirectory, while the source code itself looks for data files relative
to the parent directory of the directory that contains the executable.
So we need to set PREFIX_BIN to force the executable to be installed in
a bin/ subdirectory.

There is also an issue in nsis itself when it's executed by anything
other than a shell, which we patch out.

Differential Revision: https://phabricator.services.mozilla.com/D125638

UltraBlame original commit: cac999538d178a282c546b4e34bd3bc81aa3e8cf