Bug 1731121 - wasm: Make TypeContext refcounted and shareable. r=lth
This commit makes TypeContext refcounted/shareable. This is so that
we can change RttValue's representation to be:
  1. RefPtr<TypeContext> typeContext;
  2. uint32_t typeIndex;
Effectively adding an explicit type context field to RttValue instead
of using an implicit global TypeContext, as it is now.

The big change here is that as TypeContext is now refcounted, it
needs to be heap-allocated, and cannot be a plain member of
ModuleEnvironment, as it is now. This adds some extra indirections.
If this is unacceptable, then I think we can find a workaround.

The next commit will implement the changes to RttValue.

Differential Revision: https://phabricator.services.mozilla.com/D125889

UltraBlame original commit: 7b2661dfc414b131344a4cea008aceb056621305