Bug 1654413 - Add a MarkerStack Rust enum for the marker API r=emilio,gerald
MarkerStack object is a bit more complex in the C++ side. But we don't need a
complex object like that in here. A simple enum is enough to determine what
type of marker stack we are capturing. C++ side will be handling the capturing
of the stack.

Differential Revision: https://phabricator.services.mozilla.com/D124021

UltraBlame original commit: 89debdc6aaaca1740b354ff00480e0254e726216