Bug 1731291 - wasm: Hide .type property of imports/exports behind compile-time flag. r=lth
The pre-existing code for printing a function type to string used a check
behind 'fuzzingSafe' and I cannot find a reason for this in any
commit history. Searchfox thinks it came from the original commit
for the imports/exports feature from luke, but it's not present
in the commits he landed, nor in the bug. I don't think it's
necessary, so I think we should remove it while we're fixing
this bug.

Differential Revision: https://phabricator.services.mozilla.com/D125979

UltraBlame original commit: 9ad1733e113bc06f36eb6321861cde33d51f8e91