Bug 1731551 - [devtools] Avoid leaking DevToolsFrame JSWindowActor on each reload. r=nchevobbe
Each time the related WindowGlobal is destroyed, the related target actor and connection will be destroyed.
But the destruction of the previous actor/connection happens *after* the creation of the new.
So that hasConnection() is always true on reload. On toolbox closing it should become false.

Regardless, we should always unregister the connectionchange.
The server may be destroyed by the very last actor/connection to be destroyed.
We might want to move this code directly into DevToolsServer, that may be clearer...
But that may break non-watcher setup, so better do that once we only support the watcher codepath.

Differential Revision: https://phabricator.services.mozilla.com/D126085

UltraBlame original commit: d406b326340e9f5bb2d5fd37a6723f8a5317a2c1