Bug 1725850 - part 11: Port editor/libeditor/crashtests/420439.html to WPT r=saschanaz
The test requires to support allowing nested `execCommand` if we want to run
completely.  From point of view of the purose of crashtests, we should allow
it for testing in more complicated situation.  Therefore, this patch allows
it while the test is running by the pref.

Differential Revision: https://phabricator.services.mozilla.com/D125914

UltraBlame original commit: 4b96fdb4c8f7172e1baedf76e9b741d7019fac81