Bug 1732328 - Remove the vestiges of LayerScope. r=gfx-reviewers,jrmuizel
`Layer::GetDisplayListLog()` also still contains the metion to
LayerScope to dump display list.

But this change does not remove it because it's a part of debugging display list.
If we remove it, I think we should open a new bug for it.

Differential Revision: https://phabricator.services.mozilla.com/D126512

UltraBlame original commit: a3f0791a87fd3d2419768eff6da663361c9a3171