Bug 1730534 - Part 1: Assert state of SegmentedVector iterator is valid r=mccr8
The next patches remove elements from a segmented vector that is being
iterated. This patch adds assertions to ensure that we don't attempt to use an
iterator that points to a removed element.

The assertions are added to Done() because all the other methods call that.

Differential Revision: https://phabricator.services.mozilla.com/D125426

UltraBlame original commit: e194c3e1470cbc4cf1637d0165e649453038d643