Bug 1733398 - Use subtler dark selection colors. r=mstange
This does the same as Android and macOS (using a semi-transparent
highlight background, and currentColor for the foreground).

This is really sort of a theming issue (where the theme
mixes system colors with theme colors).

But leaving the existing selection color is prettier
even in the dark theme, so let's do that.

Differential Revision: https://phabricator.services.mozilla.com/D127241

UltraBlame original commit: 09e0a0961e16f626d18a3250a41f5df295d53b86