Bug 1728269 [wpt PR 30252] - [wave] Use the correct part of uri for the reporturl api, a=testonly
Automatic update from web-platform-tests
[wave] Use the correct part of uri for the reporturl api

uri_parts[3] is the function "reporturl" in this case, the
api is the part before that.

This fix targets the comparison view in the wave UI. Without this
fix wave won't even try to make a comparison since it will use
"reporturl" as identifier for the test group. It should be "acid"
for the "acid" test group etc.

However, note that the comparison view depends on a specific
version of wptreport to function properly, namely
https://github.com/fraunhoferfokus/wptreport

The reason is that it uses the --tokenFileName command line
flag which is only available in that version of wptreport at the
time of preparing this patch.

--

wpt-commits: 809b09cfd3adfa9fb931445b243da9d7288c5ddb
wpt-pr: 30252

UltraBlame original commit: b08d1e7734bd6f32269cd06be32b71ce997aa003