Bug 1726764 - let release promotion work without partials for staging releases. r=releng-reviewers,jmaher,bhearsum
"This is nightly so we don't know what partials we need" is not the same
as "we're making a release and we know we don't want partials".  In one
case (when called from get_decision_parameters) partial_updates is None,
in the other (called from release_promotion_action) it's an empty dict.
The latter case still enforces non-empty partial_updates for production,
but for staging it's not really necessary and falling back to nightly
behaviour in populate_release_history is just wrong.

Differential Revision: https://phabricator.services.mozilla.com/D127228

UltraBlame original commit: 2c49181e6b6fdaf7a5e4b889305a78e9aac6e83e