Bug 1733539 - [devtools] Fix NodeActor#waitForFrameLoad. r=jdescottes.
`waitForFrameLoad` was waiting for the `DOMContentLoaded` if the document `readyState`
wasn't "complete".
But `readyState` can be "interactive", which means `DOMContentLoaded` was already
emitted. In such case, `waitForFrameLoad` would wait forever.
This would cause blank markup view in the inspector panel if a node was selected
in an iframe and the page would be passed the "loading" `readyState`, but still
in the "interactive" phase.

The fix is simple: only wait for `DOMContentLoaded` if `readyState` is "loading".

The existing test we have for checking selecting node in iframe and reloading
is modified; we include a slow loading image in the iframe, which will make it
stay in the "interactive" state a bit longer.
Without the fix, the test was failing nicely, highlighting the issue.

Differential Revision: https://phabricator.services.mozilla.com/D127282

UltraBlame original commit: 640c8f2e40b8fe7e4fa140d4953bf796593decb3