Bug 1733744 - Handle page_count in raptor-webext and missing manifestparser requirement. r=perftest-reviewers,kimberlythegeek
This patch will solve an issue in raptor-webext tests where there was a missing variable definition and a change in mozperftest to handle the missing manifestparser module. In this case, it was not needed so its usage was removed. A couple of other cosmetic changes were made to the hooks_recording.py file.

Differential Revision: https://phabricator.services.mozilla.com/D127453

UltraBlame original commit: f7b841d07e275071cfa7f3398427e8f8f92367e9