Bug 1688040 - part10.4: Test ExtensionPort instance appended to port.onMessage callback arguments is strictly same instance of the same port. r=mixedpuppy
This patch is a test-only follow up to part 8.2, both of the TODOs added in part 8.2 and removed in this patch
depends from the changes introduced in "part 10.1", "part 10.22" and "part 10.3" right before this one.

This patch adds an additional assertion to the smoke tests for the runtime API part of
test_ext_webidl_api_request_handler.js to confirm that the ExtensionPort instance received as the
last parameter of the port.onMessage listener is strictly equal to the `port` the listener was
added on.

Differential Revision: https://phabricator.services.mozilla.com/D122967

UltraBlame original commit: 36cdc1600e6669bc8cdad4d62d082d5faf0930a7