Bug 1733025. Make sure the highlight text for macOS lookup feature has the right size with fission. r=botond
If we pinch zoomed with an oop iframe the font would not take into account the pinch zoom because GetCumulativeResolution in the child process was 1.

This fix also takes into account the scale from css transforms. This is notable because in non-fission cases (fission turned off, or not in an oop iframe) we do not take into account css transforms, even after this patch. (We target the correct word, but the highlight that is drawn can be at the wrong place and have the wrong size.)

Differential Revision: https://phabricator.services.mozilla.com/D126862

UltraBlame original commit: 2f1a1c9216e86a1aa9c46f9d4416058eb3b8c881