Bug 1729729 - Intermittent sessionrestore_no_auto_restore | ABORT. r=Standard8
The asynchronous behavior of the backup initialization may be the cause of the
error, because we set the idleTime var before an await, and add the idle observer
after it.

While there, I cleaned up a bit the backups initialization path.

Differential Revision: https://phabricator.services.mozilla.com/D127526

UltraBlame original commit: 289c05ebdebbfb60d28054c2b44ab5efae3353dd