Bug 1735064 - Don't override --toolbar-non-lwt-textcolor. r=desktop-theme-reviewers,dao
If we do, we'd need to also override the background to be something light.

But the defaults (dialog/dialogtext) work fine on both light and dark mode and
in all themes, so I don't know why we couldn't just do this.

Differential Revision: https://phabricator.services.mozilla.com/D128023

UltraBlame original commit: 9a091de0aed610422a40e725004888fed3ac2caa