Bug 1733243 - P2: Recreate accessible when select size or multiple changes. r=morgan
If a <select> element with a default size or size of 1 gets the
`multiple` attribute it will change from a combobox button to an
embedded listbox. Similarly, if a select's size is set to something
larger than 1 it will convert from a combobox button to a listbox.

We should recreate the select's subtree because the
implementing classes are different for those two cases.

Differential Revision: https://phabricator.services.mozilla.com/D128276

UltraBlame original commit: 8db3140db1842126dec3352a47c6183d4c942c28