Bug 1728072 - [devtools] WindowGlobalTarget's form's BrowsingContext id and innerWindowId shouldn't be affected by the iframe dropdown. r=nchevobbe
These IDs should be fixed to the top level WindowGlobal that the target is debugging.
We somewhat fake changing to another WindowGlobal to implement the iframe dropdown
(via WindowGlobal.setWindow) but this is only transient and doesn't change
the fact that the current target actor is meant to debug its original top level WindowGlobal.

Not doing this confuses the TargetCommand which is trying to look for a target actor
for the iframe we switched to, which doesn't exists as it doesn't have any specific target actor.

Differential Revision: https://phabricator.services.mozilla.com/D127776

UltraBlame original commit: 09149d4b1e2806ed8b5331a4d3db0a7d474c18bc