Bug 1589066 - Part 1: Use TraceWeakEdge to trace the Realm's edge to the global r=sfink
This uses the new API to trace the edge to the global and clean up if necessary.

The method gets renamed as it only traces this one edge now, and the
traceGlobal gets renamed to traceGlobalData to avoid ambiguity.

Differential Revision: https://phabricator.services.mozilla.com/D128347

UltraBlame original commit: 57ff5c7c6e41ee0ad894609e178364c4439b07c5