Bug 1721601. Move "zoom in if can't zoom out" from being a flag to part of ZoomTarget for AsyncPanZoomController::ZoomToRect. r=botond
Instead of passing this flag all of the time for double taps we want to decide to do this or not in CalculateRectToZoomTo. Moving it into ZoomTarget seems like a good way to do this (which is done in the next patch).

Differential Revision: https://phabricator.services.mozilla.com/D128435

UltraBlame original commit: 9fa95cb8ad3a3f6168734a20e7cfb5fb546820b9