Bug 1731501 - Use setActivationPolicy to hide dock icon on OSX r=mstange
This could be a bit risky to enable in any widespread fashion, but the
goal here is just to land it behind an undocumented command line flag
so that we can do a bit of experimentation and dogfooding. The API
seems to work exactly as desired: firefox opens in the background
without a window or dock icon. Upon trying to launch firefox again, we
simply send a remote command to the existing instance to open a window,
and it does so and gets a dock icon.

Differential Revision: https://phabricator.services.mozilla.com/D126054

UltraBlame original commit: 5bb479b1282ca8b90904c8a9c62fb0f733857e89