Bug 1735697 - ProfilerThreadRegistry::WithOffThreadRefOr - r=florian
ProfilerThreadRegistry::WithOffThreadRefOr is to ProfilerThreadRegistry::WithOffThreadRef as
ProfilerThreadRegistration::WithOnThreadRefOr is to ProfilerThreadRegistration::WithOnThreadRef.

So it is similar to ProfilerThreadRegistry::WithOffThreadRef, but it returns whatever the callback returns if the thread id was found, otherwise it returns a fallback value.

Differential Revision: https://phabricator.services.mozilla.com/D128575

UltraBlame original commit: e82e026a40a4d100ba470889505b2ff8397d5a0d