Bug 1736559 - [lint] Fix codespell silently fixing files, r=linter-reviewers,sylvestre
We have a check in the 'CodespellProcess' class to ignore errors that are
fixing single letter camelCase errors (since these tend to be used frequently.

Unfortunately, when using '--fix', these errors are fixed regardless as the
fixing happens with the codespell process. Since we increment the 'fix'
variable after the check happens, we don't even report that anything was
'fixed'.

Ideally we would find a way to prevent these types of errors from being fixed,
but for now this at will at least ensure that the user is notified that
something was modified.

Differential Revision: https://phabricator.services.mozilla.com/D128885

UltraBlame original commit: f42752e51c9c5986e9e36d7f522298d32aa8984d