Bug 1734910 - Toggle screenshots UI for screenshots component. r=sfoster
This patch removes the disabled state from the screenshots button when the screenshots ui is activated. The button now works as a toggle to activate and deactivate the screenshots ui.
The screenshots button is still disabled on about pages.
For this patch we only need to check if there is an open dialog box for the screenshots ui and if it is open then we close the dialog box, otherwise open a new one.

Differential Revision: https://phabricator.services.mozilla.com/D128410

UltraBlame original commit: 41fbbf58ca7d2c1eedfe8da59024bddcfb0e3bcd