Bug 1654112 - Fix up AudioConduit audio thread access model, and the channel count for silence. r=padenot
This patch uses mMutex to guarantee that the audio thread does not get things
pulled away from under its feet. When on the AudioProxyThread we allow blocking
the thread during negotiation. On the real audio thread we avoid blocking on
the mutex and return an error instead. This leads to us inserting silence --
which is fine since we're going through negotiation (call setup, teardown or
re-negotiation).

This also fixes the channel count when GetAudioFrame fails and we insert
silence, to use the same channel count as when we last appended real data. Since
changing the channel count can in the worst case cause a GraphDriver switch.

Differential Revision: https://phabricator.services.mozilla.com/D103494

UltraBlame original commit: b5228352d20db6c4899ddc3de268ebc35575f0d7