Bug 1654112 - Fix remote-inbound-rtp.packetsReceived. r=bwc
There are two problems with this stat:
- The BaseSeq tracking logic in MediaPipeline doesn't account for multiple ssrcs
  (rtx)
- The stats code in PCImpl queries the BaseSeq on the wrong thread (Call vs Sts)

This patch fixes both of these, by tracking BaseSeq per ssrc, and by introducing
an extra thread hop.

If it weren't for mediapipeline_unittest we could remove all the packet counting
in MediaPipeline. Doing that proper (with gmock?) is a patch for another day,
but this patch goes in this direction by also lifting out the BaseSeq logic to
the conduits.

Differential Revision: https://phabricator.services.mozilla.com/D125709

UltraBlame original commit: 7888e4be3d7e633739cf590e73b0a3af0ef3f9fc