Bug 1730712: Do simple "pip install" when preparing venv r=ahal
`install_pip_package()` does some logic around not running pip if a
package is already installed.

However, this is redundant, because when building a venv, no packages
are installed.

Additionally, this was the constraint forcing us to populate the
virtualenv in a separate process from the one building it.

Differential Revision: https://phabricator.services.mozilla.com/D129294

UltraBlame original commit: 3326a2364d35e66ffdc2280a30563419f8c6ec29