Bug 1661127: Cache Windows IVirtualDesktopManager::GetWindowDesktopId value r=cmartin
GetWindowDesktopId is a slow function and is called on the UI thread during WM_CLOSE message handing (as part of SessionStore). This was causing Windows to assume we are stalled and it abandons some messages, namely future WM_CLOSE messages sent for other windows during a "Close All Windows" operation.

In this patch, GetWorkspaceId uses a cached value for GetWindowDesktopId (if available) but also queues an update for that cache value if one isn't already in progress. Since SessionStore already regularly calls GetWorkspaceId in response to window moves, this keeps the value fresh in plases where it might change. (The SessionStore architecture is designed that way because Windows does not notify the application if the user moves the window to another virtual desktop.)

Differential Revision: https://phabricator.services.mozilla.com/D128955

UltraBlame original commit: 542710f708e384113725be5d04b4cd03cc6c6363