Bug 1736433 - Add InitInfo::builder. r=owlish
This patch adds a builder for InitInfo and allows us to define its members as
final.

This is especially useful for File Descriptors which are all of the same type,
and could be easily swapped in method calls.

Differential Revision: https://phabricator.services.mozilla.com/D128933

UltraBlame original commit: c4da0adbf71a9aa6cbd330dfda03e2143ddeec6b