Bug 1736433 - Make sure process is not alive when running xpcshell-test. r=jmaher
Sometimes we try to start a process that already exists and that causes a very
confusing error (the test times out but there's no indicastion as to why).

This patch makes it so that we will first try to kill the existing zombie
process and then, if we fail to do so, throw an error.

Differential Revision: https://phabricator.services.mozilla.com/D128937

UltraBlame original commit: 81f20f90b6d7180ebaa5135235f59cc5b958682f