Bug 1675456 - [devtools] Enable watcher for webextension toolboxes. r=nchevobbe,devtools-backward-compat-reviewers
There is a few things complex around WebExtension debugging:
* DevTools creates a browser element and connect to it.
  The WebExtensionDescriptor instantiate its very own browser element
  and DevTools codebase communicate to the extension process via this browser element.
  We are still using the frame connector and message manager.
  We do that especially to workaround add-on not having any document running (no background page, nor any popup).
  Without that we wouldn't have any document and couldn't spawn any WindowGlobalTarget actor.
* We have a unique target, debugging many documents.
  `webExtensionTargetPrototype._searchForExtensionWindow` will iterate over all document and try to
  switch the WebExtensionTargetActor to another document.
  This is based on `WindowGlobalTarget._setWindow` and so use the same magic hack
  than the iframe dropdown to change the debugging document of a given WindowGlobal.

What is especially tricky about this is that we connect to the devtools specific browser element
in the parent process having one specific browserId, while in the content, we most likely
move the WebExtensionTargetActor to the background page, having a distinct browserId.
For a given extension, all their browser elements (background page, popups) are independant
and have distinct browserId.

Because of this we can't rely on the existing browser-element watcher context type.
We rather need to do a filtering the `addonId` which is correctly set on each document's principals.

Ideally, we would spawn one target for each of these browser elements.
To do that, we have to begin with supporting the watcher actor for webextension (purpose of this patch).
Then we would need EFT, otherwise the iframe dropdown would no longer work for webextension,
whereas that's a really key feature when debugging them.

All this is brittle, but there is a lot of chicken-egg issues here.
It would be best if we were spawning one Target for each individual extension document.
But to do that, we first need to start supporting the Watcher codepath.
Also, we would need EFT, otherwise the iframe dropdown would no longer work for extension documents.
Which is an important feature for add-ons debugging.
Note that for now, we don't watch for any additional targets when debugging a webextension.
(TargetCommand doesn't call watchTargets(FRAME))

In the meantime, we can at least start supporting the Watcher actor and server watchers,
but we will have to workaround the fact that we have a unique target, instantiated from the client,
by the descriptor. And this target may switch its focus to various documents.

Supporting server side resource watching is actually a serious motivation here.
This would help get rid of the last usage of the old storage actor (which also keeps running setupInParent and ultimately client side target switching).
But to do that we need a hack just for web extension.
As we have a unique target, with a fixed and surprising browserId/browsingContextID (related to the devtools browser element),
we have to ensure that all resources are emitted via these IDs, otherwise the ResourceCommand wouldn't be able to match the target front.

Differential Revision: https://phabricator.services.mozilla.com/D99099

UltraBlame original commit: 7950d8e5bc99411d79f5939a7b1b7cb0d01f9e9c