Bug 1735551 - [devtools] Add a timeout to findNodeFrontFromSelectors. r=ochameau.
Since the method might wait for iframe to load, it could stay unsettled for a
while.
Adding a timeout will make the callsites not wait for ever, at the expense of
correctness.
This should be okay as this function is mainly called to re-set things after
reloading the page, which is a nice to have, but not critical feature.

Differential Revision: https://phabricator.services.mozilla.com/D128854

UltraBlame original commit: 5a2367e856d2afe4e429c0c1570d8bb6b0b0c048