Bug 1739576 - Make 8GB available. r=rhunt
Tidy up MaxMemoryPages(IndexType::I64) so that it returns a value
corresponding to the max buffer byte length.  Fix two asserts that
were incorrect and that triggered when making this change.

Add a test that the max memory pages for i64 is indeed at least 8GB.

Drive-by fix to allow a test to OOM without reporting that as a
failure.

Differential Revision: https://phabricator.services.mozilla.com/D130452

UltraBlame original commit: f369c2fc3ab7160413de89167a79af31fbd3b084