Bug 1739990 - Fix perf regression in getInScopeLines. r=jdescottes.
The patch for Bug 1737993 did a very readable but not performant replacement
of lodash `without` usage in `getInScopeLines`.
This patch changes things a bit so we get a similar level of performance as
what we had before, when using `without`.
For this we :
- make `getOutOfScopeLines` returns a `Set` so checking inclusion is faster
- create the final array of "in scope lines" using a for loop and assigning
  line number when needed, and then filtering out empty slots.

Differential Revision: https://phabricator.services.mozilla.com/D130742

UltraBlame original commit: 4789d873d1cbec5371eb101f95c5e173ee7fc69e