Bug 1713030 - [puppeteer] Support for custom user data (profile) directory for Firefox. r=webdriver-reviewers,jdescottes
When using a custom Firefox profile for Puppeteer the modified
preferences as present in prefs.js need to be reset once the
profile is no longer needed by Puppeteer. If not done this could
cause side-effects when the profile is used next time outside
of Puppeteer.

Differential Revision: https://phabricator.services.mozilla.com/D128452

UltraBlame original commit: 31302975681c8978f26438b9ce3454875a264aeb