Bug 1692708 - Add cancelScrollAnimation and use it in window_bug1369072.html and browser_scroll.js. r=botond
Both window_bug1369072.html and browser_scroll.js have a pattern something like;

1) Call synthesizeKey to scroll an element
2) Wait a scroll event
3) Test something
4) (Re)set the scroll position to (0, 0) by calling Element.scrollTop and
   Element.scrollLeft or scrollTo.

With the new transmogrification the step 4) will not reset the scroll position
if there's an on-going scroll animation in APZ triggered by the synthesizeKey.

To avoid the situation where there's an on-going animation at the step 4), we
destroy the corresponding scrollable frame and reconstruct it so that it
clobbers the animation.

Differential Revision: https://phabricator.services.mozilla.com/D130852

UltraBlame original commit: b31c1668258d9d9971455acbfdf5fe8dda341cb4