Bug 1741246 - Pass JS::GCCellPtr in nsCycleCollectionTraversalCallback::NoteJSChild(). r=jonco
This is a pointer-sized value, so there's no need to pass it by
value. All of the SpiderMonkey places that pass GCCellPtr are doing
it by value, so this is aligning with that.

In addition, I removed a NoteJSChild() overload that is no longer
used.

Differential Revision: https://phabricator.services.mozilla.com/D131163

UltraBlame original commit: 31577d9f7e6931947e7df364c4dc9e0232b48e22