Bug 1737177: Add `nsSystemInfo` runtime flag for whether running application is running under Microsoft Windows S Mode. r=mhowell
This patch uses the Windows.System.Profile.WindowsIntegrityPolicy class to determine whether or not Windows is in S mode. This class is only available beginning with SDK 17763 (newer than what we use right now), so we need to fetch it at runtime with some magic.

This is a completely no-op for Windows versions older than 1810, where the WindowsIntegrityPolicy class is not available.

Differential Revision: https://phabricator.services.mozilla.com/D130397

UltraBlame original commit: 2dc54f256ad3bcded645bbe12a0c658e03494d3c