Bug 1737682 Part 1: Cache calls to NativeLayerCA::ShouldSpecializeVideo in a state bit. r=gfx-reviewers,mstange
This guarantees that any time mMutatedSpecializeVideo is set to true, the next
call to ApplyChanges will receive the value that actually triggered the
mutation. This was *probably* already true, but this change makes it explicit
and similar to other state changes.

Differential Revision: https://phabricator.services.mozilla.com/D129452

UltraBlame original commit: 117961db7b95e5e979cab956358b79c9a32caa24