Bug 1723082 - Support reading files > INT32_MAX in IOUtils r=Gijs
Previously, if you attempted to read a file greater than INT32_MAX in size, we
would either fail to read the file (on release) or crash (on debug). This is
due to the implementation of _PR_MD_READ accepting an int32_t of bytes to read
instead of a uint32_t, which nsFileStream::Read accepts and passes on.

Differential Revision: https://phabricator.services.mozilla.com/D131161

UltraBlame original commit: 492f5ca3b5923e128066c3be5c49160f9e40babf