Bug 1742010 - File descriptor leak when serializing into SurfaceDescriptorAndroidHardwareBuffer. r=sotaro,gfx-reviewers
We accidently duplicate the file descriptor when serializing because we
hit a constructor for ipc::FileDescriptor which duplicates the handle
instead of just taking it. This means we are supposed to close it
explicitly but we've already forgotten about it. We should just do a
move instead to avoid this.

Differential Revision: https://phabricator.services.mozilla.com/D131569

UltraBlame original commit: 9b6abfc76fd59d88ae961a77bb4f6692948bd0c3