Bug 1742102 - Part 3: Clean-up includes for JitScript. r=jandem
"JitScript.h" was only including "TrialInlining.h" because the destructor
needs to see the full definition of `InliningRoot`. When moving the
destructor into the cpp-file, we can avoid that include. The constructor is
already placed in the cpp-file, which gives a strong indication that it's not
necessary to inline the destructor.

Reducing the number of includes for "JitScript.h" is helpful, because that file
is included in many other files through its transitive include in "JSScript-inl.h".
And because "TrialInlining.h" includes "JSScript.h", it makes sense to avoid
including "TrialInlining.h" unless absolutely necessary.

Drive-by change:
- Use IWYU to fix the includes in JitScript.h and JitScript-inl.h

Differential Revision: https://phabricator.services.mozilla.com/D131612

UltraBlame original commit: b70e65483dfd55bad77e1dbda5448ab516fca294