Bug 1742107 - [devtools] Don't ignore same-origin iframe in EventLoop#getAllDebuggees. r=bomsy.
This was causing some issues with PausedDebuggerOverlay with EFT:
As we weren't returning the window to preEnter, we weren't calling
suppressEventHandling on it, which in the end made the call we
have in the overlay to setSuppressedEventListener inoperant.
This wasn't caught by the browser_dbg-paused-overlay-iframe.js test
we have as the method we use to simulate clicking on the button
(highlighterTestFront.clickPausedDebuggerOverlayButton), ends ups calling
the handleEvent method of the highlighter actor directly, bypassing the
regular flow of events in a paused page.

Differential Revision: https://phabricator.services.mozilla.com/D131654

UltraBlame original commit: 31fca3976f3bb1b7025b0d898d0bbbc02c72c696