Bug 1742146 - don't race localization against non-localization when updating the downloads prefs, r=jaws,preferences-reviewers
This race condition happens because some of the paths through displayDownloadDirPrefTask
are synchronous, and some are not, and we repeatedly call it in the same turn of the
event loop, both because downloads prefs are stored in 2 prefs, both of which impact
this display, and because the pref bindings like calling 'change' event listeners more
than once.

To make this reliably deterministic, instead of juggling callsites, just ensure that
the last caller to displayDownloadDirPrefTask is always the one that actually
determines the label that gets displayed.

Differential Revision: https://phabricator.services.mozilla.com/D131853

UltraBlame original commit: f03340a5c17db8d93c0d2572e403f5086c13c373