Bug 1742541 - [devtools] Don't try to show PausedDebuggerOverlay if document isn't ready. r=ochameau.
At the moment, the overlay can't be displayed if the DOM isn't loaded,
and we were waiting for that state before trying to show the overlay.
But we can't move forward in the document loading process until we resume,
which mean we could end up in weird state with the overlay, where the
page wasn't paused anymore, but we'd display it.
In this patch, we simply bail out if the document isn't ready.

Differential Revision: https://phabricator.services.mozilla.com/D131884

UltraBlame original commit: 762649eb6709733b5685608445d490fe20152e73