Bug 1742053 - Instance-level creation of indirect stubs. r=dbezhetskov
Author: dbezhetskovigalia.com
Editor: lhansenmozilla.com

Confidence: high

Instance::ensureIndirectStubs and its callees.  See comments on a
previous patch about the discrepancy in predicate meaning between
RemoveDuplicates (in this patch) and IndirectStubTarget equality (in
that patch).

I've also added rooting of the WasmInstanceObject in
ensureIndirectStubs.  This is conservative - an argument could be made
that it's not necessary at present - but it's not hard to pass in the
context from the caller, and it's good hygiene.

Differential Revision: https://phabricator.services.mozilla.com/D131630

UltraBlame original commit: 2347af2773c04c7fd7906b9c556fa8a02788cfd3