Bug 1700774 - [taskgraph] Fix circular dependency on 'get_mobile_project' in variants transform, r=gbrown
We have a 3 way circular dependency in the test transforms:

1. The 'raptor' transforms depend on 'split_variants'
2. The 'get_mobile_project' utility function depends on the 'raptor' transforms
3. The 'split_variants' transform depends on the 'get_mobile_project' utility function

The only reason things are working is that in step 3 we don't need the result
of 'get_mobile_project' to be perfect. This revision breaks the circular
dependency by checking for 'android' in the test-platform rather than relying
on 'get_mobile_project == geckoview'.

While this is admittedly not as precise, we only apply the particular variant
that needs this to geckoview tasks in the first place. In other words, this
revision does not cause any additional tasks to be added / removed.

Differential Revision: https://phabricator.services.mozilla.com/D132072

UltraBlame original commit: 60e50587018c66135367ac16f59a4b219529ec62