Bug 1741479 - Turn on Merino as a Firefox Suggest source. r=nanj,webdriver-reviewers,whimboo
We only need to default `browser.urlbar.merino.enabled` to true. If the user has
opted in (either through the modal or by toggling on the data collection pref in
the prefs UI), then `quicksuggest.dataCollection.enabled` will also be true and
we'll fetch Merino suggestions. Otherwise it will be false and we won't fetch
Merino suggestions. That logic is implemented here:
https://searchfox.org/mozilla-central/rev/9a5f36b0ddb9cb8ae556fc5b45f8ccea0f0da6f8/browser/components/urlbar/UrlbarProviderQuickSuggest.jsm#144

Note this defaults the pref to true for everyone, even users in offline. It make
senses now that we have a separate toggle for data collection in the preferences
UI. Even offline users can opt in to Merino and data collection.

I also updated the various sets of prefs for test suites so that the Merino
endpoint URL is empty when running tests so they don't hit the network. I could
have forced `merino.enabled` to false instead, but setting the endpoint URL has
a couple of benefits, although admittedly they're very small:

* It runs a little more of the Merino code path (i.e., calls
  `_fetchMerinoSuggestions`)
* It lets Merino tests set only one pref, the endpoint URL, instead of two, both
  the endpoint pref and enabled pref

Differential Revision: https://phabricator.services.mozilla.com/D131988

UltraBlame original commit: 408b89a2e02cccd63dbe5c451ac1c6b248379534