Bug 1743150 - Shorthands in computed style shouldn't be enumerable. r=layout-reviewers,jfkthame
This is as per spec, and causes failures once we enable the pref to expose
shorthands in the computed style. We need to decouple the "enabled-ness" from
the enumerable-ness, which we do by keeping two separate arrays, and leaving
ComputedStyleMap for the enumeration.

Differential Revision: https://phabricator.services.mozilla.com/D132266

UltraBlame original commit: 3c8376ad956e275d8a7463a26022e45f0cedb868