Bug 1743612 part 3 - Support AddSlot IC stubs for objects with an addProperty hook. r=iain
If the object has an addProperty hook, we call a VM function where we add the
data property and invoke the hook. This still isn't super efficient, but it
eliminates a source of IC failures and helps identify this case in profiles.

Differential Revision: https://phabricator.services.mozilla.com/D132491

UltraBlame original commit: 7fe6d2297eb5f74ed5804af89ec308b3d65af1a5