Bug 1743519 - Remove ScriptSource::finalizeGCData. r=arai
With the removal of BinAST and the addition of Stencil, the ScriptSource no
longer has any GC references in the incremental-encoder. We can now remove the
early cleanup and rely on the normal ScriptSource destructor, even if this
happens off-main-thread.

Differential Revision: https://phabricator.services.mozilla.com/D132424

UltraBlame original commit: f33f012d62e86223fba96287bfbcd3146d6e6313