Bug 1743945 - Move the `use_fetches` transform after `make_task_description`. r=taskgraph-reviewers,jmaher,ahal
This will help to avoid the duplication of the code that takes artifact
names, environment and aliases for each fetch, and will allow changes to
toolchain.py that wouldn't be possible otherwise.

The only difference it makes to the resulting taskgraph is that scopes
are not in the same order as before, but that doesn't matter.

Differential Revision: https://phabricator.services.mozilla.com/D132649

UltraBlame original commit: 21b1c88eaf7b8bd162215837053cb4a84db5d8e3