Bug 1742053 - Improve a comment.  r=dbezhetskov DONTBUILD
I confused myself for a while worrying about who keeps the stubs code alive
if an instance creates a stub for an imported function in a table that it
then exports; the stub-creating instance could subsequently die without the
table dying.  But since the stub-creating instance's tls is stored in the table
with the stub, this is fine.  Just clarifying this.

Differential Revision: https://phabricator.services.mozilla.com/D132554

UltraBlame original commit: 4f1a80f416373746696c55302cc0d8727a8623d8