Bug 1735445 - Minor editorial fixes in the CSS Gap Decorations draft spec. NPOTB DONTBUILD
A few changes to address some of the feedback in:
https://github.com/w3c/csswg-drafts/issues/6748#issuecomment-962273499

I moved the definition of the lateral/longitudinal axes earlier
in the document so that they can be used for the rule-image-*
section too.

I also tweaked the borders/rule-colors in the last example
to make it clearer.

Differential Revision: https://phabricator.services.mozilla.com/D132971

UltraBlame original commit: b999f67715e800b0f823b6a9f7b51196384e00c1