Bug 1744706 - Always show the git commit in fetch task logs. r=taskgraph-reviewers,aki
99% of the fetch tasks that pull from a git repo pull from a specific
revision, as specifically intended. There is however one exception,
which is the clang-trunk fetch, which is intended to always pull the
last revision on clang trunk. Unfortunately, the log also doesn't say
what specific commit it checked out.

Ironically, the log for all the other tasks does say which commit is
checked out, because git prints it out when it goes in detached head
state, which it does when checking out a specific sha1 (so in that
case, it's printing something we already know).

In the one instance where we checkout a branch, we don't rely on the
branch being checked out, so we just forcibly go in the same state of
detached head in that case too.

Differential Revision: https://phabricator.services.mozilla.com/D133063

UltraBlame original commit: 9d5b5b1e9944c5da8833c91053d9b2705b25be19