Bug 1743768 - let oom pass through assertErrorMessage. r=yury
assertErrorMessage should propagate plain OOM errors as these errors
can't be tested for with that predicate.  Instead, the test file will
typically be annotated with allow-oom (or not) to specify how to deal
with plain OOM.

This is hard to test, but if the intermittents stop arriving at the
present bug it probably indicates success.

Differential Revision: https://phabricator.services.mozilla.com/D132922

UltraBlame original commit: 6ec452bbf721618eae38aa9b634d7c6cb26be088