Bug 1744565 - [devtools] Fork browser_resources_sources.js to test GC-ed sources explicitely. r=bomsy
But the miss of target attaching wasn't the sole reason why it fails.
The `resurrectSource` function is still failing for javascriptURL and domTimer sources.
I'm also seeing exception about not being unable to load sources.js, but that doesn't seem to break anything in this test.

So I'm forking this test in two. One, as before, which will assert that we
catch absolutely all the sources when we reload the page *after* we started watching.
And a new test, which ensure testing GC-ed sources and better highlights
which are the one that we fail resurrecting.

(I'm also removing target related assertions as they aren't really relevant for this test)
(And migrate to CommandsFactory to simplify the test a bit)

Differential Revision: https://phabricator.services.mozilla.com/D133094

UltraBlame original commit: 66f56bc9fae343f06a7c9096dd42f13ab863ae40