Bug 1713579: Extract shape-guard folding into separate optimization r=jandem
This patch takes the code from `MGuardShape::foldsTo` and splits it out into its own optimization. Besides jitspew, the main change is adding support for MNewPlainObject (which didn't exist when this code was first written.)

We use `dependency()` to find the most recent store. It is possible for a node's dependency to be in a dead block. In `ValueNumberer::visitDefinition`, we check for this case; I used the same approach here.

Differential Revision: https://phabricator.services.mozilla.com/D133129

UltraBlame original commit: 28c38d556fb86f1a936d9f6a8627705230df895a