Bug 1743594 - [devtools] Add a readyPromise property to the RuleView panel and wait for it in test helpers. r=ochameau.
The RuleView calls `onSelected` from its constructor, which might call `CssRuleView#selectElement`,
which returns a Promise that is resolved when the view gets populated.
So far tests weren't waiting for this specific Promise, which could lead to failures
as test would query the DOM for the CssRuleView (which might not exist yet).
In order to fix this, we expose a `readyPromise` on the RuleView panel and await
on it in the `openRuleView` test helper.

Differential Revision: https://phabricator.services.mozilla.com/D132930

UltraBlame original commit: 4b509a65fef020e2f467712a92e2238cb4968e73