Bug 1707983. Refresh zoom constraints from UIResolutionChanged notification because it gets called when we swap docshells and get a new APZC tree manager. r=botond
I guess we need to do this every time we get a new APZC tree manager (because that is where the zoom constraints map/cache is stored). I'm not sure how to detect that or if there are other cases we haven't covered.

Differential Revision: https://phabricator.services.mozilla.com/D113608

UltraBlame original commit: e7b011f92cf4abb5eea83582f1f13318f6071275