Bug 1745115 - Default number of stages to build clang to 2. r=firefox-build-system-reviewers,andi
- stage 1 is building clang with whatever compiler is available
- stage 2 is building clang with the clang built during stage 1
- stage 3 is building clang with the clang built during stage 2. It's
only useful when it's actively compared against stage 2, or when
there's a stage 4.
- stage 4 is building clang with the clang built during stage 3, with the
profile generated during stage 3, when stage 2 produced a clang with
instrumentation enabled.

We're not actively comparing the output of stage 2 and 3 when not doing
PGO, so it's not useful to do 3-stage builds.

Differential Revision: https://phabricator.services.mozilla.com/D133314

UltraBlame original commit: 693ae39ae0dd8d16bfbedbd317a5c7470570481b