Bug 1743418 - Part 1: Use shared data to populate remote settings list to content processes. r=pbz
We used to use the message manager to get query stripping list from the
parent process during the startup of content processes as well as the
list gets updated from remote settings.

In this patch, we change to use the shared data to populate the list for
remote settings. This brings one benefit that we don't need to start the
list service during the startup of content processes in order to get the
list as early as possible. We can get the list lazily because the shared
data will be preloaded during the process startup.

Differential Revision: https://phabricator.services.mozilla.com/D132990

UltraBlame original commit: 1ab00f31e2ed3069bf151d08ab983daedad4cda9