Bug 1744496 - Fix titlebar corners in ElementaryOS themes. r=stransky
They style both the parent of the titlebar _and_ the titlebar, which is
a rather odd thing to do, but alas, it works on applications like Gedit
and so on (but not in gnome-tweaks, interestingly).

Relevant code:

  https://github.com/elementary/stylesheet/blob/9b7122b07b3a1878501d13d842c54e83ef9061c6/src/gtk-3.0/widgets/_headerbars.scss#L6-L25

Detect this by checking that the headerbar has no radius and the parent
does, and in that case read and render the container frame.

This shouldn't regress any other theme, ideally...

Differential Revision: https://phabricator.services.mozilla.com/D133516

UltraBlame original commit: 5ff985e01c654ce10096df52bfc2bb1b3af7b0c1