Bug 1745460 - Modify UrlbarQuickSuggest._fetchIcon() so it bails when the remote settings client is null. r=mythmon
This is a small follow-up to bug 1745297. UrlbarQuickSuggest can now be queried
even when suggestions are disabled and its `_rs` property (the remote settings
client instance) is null, which is a problem for `_fetchIcon` since it assumes
`_rs` is non-null. It can just bail in that case.

Differential Revision: https://phabricator.services.mozilla.com/D133540

UltraBlame original commit: 8ed24cf28fba332addbd6c52055638c30093bdfb