Bug 1743914 - Part 1: Allow UserChoice to set Firefox as the default handler for arbitrary file extensions. r=mhowell
This is a straight-forward extension of the existing UserChoice
mechanism to arbitrary file extensions.  This is not a broad
generalization but we don't foresee many extensions beyond the
`".pdf"` file extension at this time, so let's not over-rotate towards
generality.

Differential Revision: https://phabricator.services.mozilla.com/D132658

UltraBlame original commit: 20713ede0c0816411bf58a06a8ecbe423ae240e0