Bug 1733751. Don't consider the critical displayport in AboutToCheckerboard. r=botond
With the old layers backend we had low precision buffer code controlled via the pref layers.low-precision-buffer that was used on android. We would expand the displayport by 4x and then paint it as 1/4 the resolution, and then we would have a critical displayport without the 4x multipler that we would paint at the real resolution. The code to do the painting at the lower resolution was in the layers backend and has since been removed. (This is okay because webrender doesn't rasterize all of the content in the displayport.) So the critical displayport or displayport are no longer treated differently anywhere. Except a few pieces of code that should be fixed/changed/removed. This patch being one of them.

Using the critical display port here is incorrect: the display port and critical display port have the same status with regards to rasterized content as well as the same status with regards to webrender having the displaylist of the content to be able to rasterize.

AboutToCheckerboard is only used in AsyncPanZoomController::OnScale to potentially schedule a RequestContentRepaint sooner, so this should have a small effect in practice.

Differential Revision: https://phabricator.services.mozilla.com/D127349

UltraBlame original commit: fb9392c698c39ed95c1b0c40cbdc66720a6d2987