Bug 1737117 - Only process updates only for `--backgroundtask backgroundupdate`. r=bytesized
The aim is to avoid background tasks causing unexpected updates, as
happened when we tried to migrate `pingsender` to a Gecko background
task in Bug 1734262.  This commit makes it so that we only process
updates for the `backgroundupdate` task (and the test-only
`shouldprocessupdates` task).

Differential Revision: https://phabricator.services.mozilla.com/D133557

UltraBlame original commit: 783fd695e32bc2cbf773808eaa76e4932aed427e