Bug 1746214 - Re-cache Nimbus variables in UrlbarPrefs after awaiting feature initialization to avoid a race condition. r=nanj
The problem is a race condition: UrlbarPrefs can end up caching Nimbus variables
before `SharedDataMap` in the Nimbus implementation finishes loading its
experiment JSON from disk and `await NimbusFeatures.urlbar.ready()` is
subsequently resolved. That's why QA can only reproduce the bug on a slow
machine. If that happens, the cached variables will be out of date, which means
the `quickSuggestScenario` variable will not be cached and so the effective
scenario becomes offline. Therefore the pref migration thinks the scenario is
offline when it's actually online, and it doesn't set the right prefs.

The simple fix is to clear the Nimbus variable cache after the
`await NimbusFeatures.urlbar.ready()`. I'd like to rework initialization at some
point to make this a little nicer. I've actually started doing it in a
refactoring patch I've been working on, but I think we can save that for later
and take this simple fix for now.

While working on this I noticed there's another possible race, between two calls
to `updateFirefoxSuggestScenario`: one by BrowserGlue and one by the Nimbus
update callback in UrlbarPrefs, `_onNimbusUpdate`. The BrowserGlue call passes
true for `isStartup` but `_onNimbusUpdate` does not. Normally (always?) the
BrowserGlue call happens first so its call really is the startup call. But if
the `_onNimbusUpdate` call were to happen first and it did not finish by the
time the BrowserGlue call happens, then the BrowserGlue call would be ignored
(due to the `this._updatingFirefoxSuggestScenario` early return in
`updateFirefoxSuggestScenario`), and pref migration wouldn't happen for that
Firefox session. To fix that, we can get rid of the special `isStartup` param
altogether and just keep track internally whether `updateFirefoxSuggestScenario`
has been called before. If it has not, then that's the startup call, regardless
of who is making it. (Here again reworking initialization would make this
nicer.)

Differential Revision: https://phabricator.services.mozilla.com/D134057

UltraBlame original commit: cbc60dec8bddf30b42025db9395d6286dd1fec05