Bug 1746407 - Fix layer-import-parsing.html. r=boris
This test should pass, actually, but it was failing because we don't
allow loading import rules in a detached stylesheet because it's not
clear which document should own the load.

The sheet doesn't need to be removed for the test to be correct.

Differential Revision: https://phabricator.services.mozilla.com/D134019

UltraBlame original commit: 0ee4e4c13392a320a5119ac1bcc69b55a3bc3d27