Bug 1743760 - Clean up some names. r=rhunt
This tidies up some of the naming in Dmitry's patch.  When we create
stubs we are not interested in whether the table is imported or
exported per se, but in whether it is public (shared by multiple
instances) or not.  At the moment, that happens to be true if it were
imported or exported, but we want to move away from that.

Differential Revision: https://phabricator.services.mozilla.com/D134381

UltraBlame original commit: b16a7f303241ef528b37817e0bc8fa0d0edf6238