Bug 1745328 [wpt PR 31988] - [FlexNG] Adjust row flexbox info for item expansion, a=testonly
Automatic update from web-platform-tests
[FlexNG] Adjust row flexbox info for item expansion

When a flex-item expands due to fragmentation, we need to adjust the
intrinsic-block-size and subsequent offsets.

In CL:3290486, the original block-size of each item was stored on
NGFlexItem and reduced as the item is fragmented. We can use this same
strategy for row containers to tell when an item expanded as a
result of fragmentation (i.e. when the stored block size becomes
negative).

For rows, though, the item expansion applies to the next line rather
than items in the same line. Thus, utilize the next line's
|item_offset_adjustment| to keep track of the max item expansion
in the current row. This is then accumulated between rows.

A general expansion related bug was also fixed by this CL (see the extra
tests added for single-line column fragmentation). The bug was that
if an item breaks before, and it had a top margin, that margin
would get included in the |item_offset_adjustment|, which is incorrect.
ConsumeRemainingFragmentainerSpace() was updated to calculate the
correct amount the item expanded by.

The following wpt test is now passing as a result of this CL:
virtual/layout_ng_flex_frag/fast/multicol/flexbox/flexbox.html

Bug: 660611
Change-Id: I4f4ae8cf32a3d501e7dc293ed4c9fa45f2380cd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3328370
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: Alison Maher <almahermicrosoft.com>
Cr-Commit-Position: refs/heads/main{#950661}

--

wpt-commits: bd3d45aeffbdfce779266f6daf6dab4476012fac
wpt-pr: 31988

UltraBlame original commit: 52ecd63d39d2b0dc11a12e8c03482b0650a9baa4