Bug 1564784 - Allow drawing with HTMLCanvasElement paired with an OffscreenCanvas. r=gfx-reviewers,kvark
When CanvasRenderingContext2d.drawImage, or createImageBitmap are
called with an HTMLCanvasElement with an OffscreenCanvas gotten through
transferControlToOffscreen, we do not snapshot the underlying
OffscreenCanvas. This patch makes it so we do.

There are WPT tests for this, but they only run for canvas 2d.

Differential Revision: https://phabricator.services.mozilla.com/D134078

UltraBlame original commit: 756eda9a06dcdc400717a916bf9deb368ff5afc1