Bug 1748779 - Update midir and coremidi. r=gsvelto
The midir update reduces the differences with upstream to the coremidi
version.

And now the coremidi override is done via a patch at the top-level. The
revision we were using is gone, but it turns out the new master is
identical in content (at least, as far as vendoring is concerned).

Differential Revision: https://phabricator.services.mozilla.com/D135194

UltraBlame original commit: 466340a38933f3511d6753ad2d1f7397e55ef3b0