Bug 1742985 - Ensure ZoomConstraints get refreshed after GPU process restart. r=tnikkel
Following a GPU process restart ZoomConstraints do not currently get
set for the newly recreated APZCTreeManagers, meaning it is no longer
possible to asynchronously zoom pages.

To solve this, we make ZoomConstraintsClient observe a new
"compositor-reinitialized" topic. We send this notification in
GPUProcessManager::HandleProcessLost() to notify
ZoomConstraintsClients for parent process documents, and in
ContentChild::RecvReinitRendering() for documents in their respective
content processes. This must be performed after the compositor has
been reinitialized so that the APZCTreeManagerChild is able to send
the constraints to the APZCTreeManagerParent in the compositor
process.

Differential Revision: https://phabricator.services.mozilla.com/D135207

UltraBlame original commit: 441e0b2582cc86f15ae0a2b6d0e835240e112471