Bug 1749507 - [remote] Emit internal event when window global message handler has been initialized. r=webdriver-reviewers,jdescottes
Before an instance of a WindowGlobalMessageHandler is ready to be used the
initial session data needs to be applied. Because this happens asynchronously
an internal "window-global-handler-created" event will be used for which the
root message handler or any other module can listen for.

Differential Revision: https://phabricator.services.mozilla.com/D134265

UltraBlame original commit: 1b7c8210434b8bdeef88846c24ac77734d74a551