Bug 1745556 - Cache paths to textures to avoid falling back to Skia. r=gfx-reviewers,aosmond
This generalizes the CacheEntry infrastructure to implement path caching.
This allows us to cache recently drawn paths to a texture which can be then
drawn without falling back to Skia.

Paths using simple solid color patterns will be drawn as alpha masks that
can then have the color applied in the shader. For now, other pattern types
have the pattern baked into the texture to avoid having to complicate the
shader setup for now. In the future, other pattern types could be supported
with new shaders to allow those to be cached as alpha masks as well.

In the service of this, comparison and cloning operators were added to the
Pattern classes to make remembering and matching against them easier for
cache entries.

Differential Revision: https://phabricator.services.mozilla.com/D135261

UltraBlame original commit: ae505acc7a28bf004be0c662f722fe2907ceb15b