Bug 1749671 - Make more data bits available in LAllocation on 64-bit. r=nbp
This is a simple fix for a too-small field for the virtual register number on arm64,
where the field gets only 18 bits.  This change expands the field to 21 bits.  This
is enough to solve compilation problems on web.autocad.com.

Differential Revision: https://phabricator.services.mozilla.com/D136106

UltraBlame original commit: 1be73998d9b0eeb501411a6883e75fec8581717a