Bug 1750558 part 1 - Fix NativeObject::canReuseShapeForNewProperties to handle the HasEnumerable flag. r=iain
We were checking the old and new shapes have the same object flags, but this was
(almost) never the case after we added the HasEnumerable flag in bug 1733075,
disabling the optimization in practice.

With this patch we hit the fast path again on Reddit.

Differential Revision: https://phabricator.services.mozilla.com/D136145

UltraBlame original commit: eb404993ccfc5bfadf77d338c5ccf867aa7bf9cd