Bug 1750980 - Fix UAF for OffscreenCanvas after its associated HTMLCanvasElement is destroyed. r=gfx-reviewers,jrmuizel
OffscreenCanvasDisplayHelper needs its Destroy method called when
HTMLCanvasElement is destroyed. This was missed during the initial
implementation. It holds a weak/raw pointer to the HTMLCanvasElement it
is bound to, so if the element is destroyed, we need to clear our
reference to prevent any UAFs.

This would be a security bug if not for the fact that this feature is
still currently disabled behind a pref and has not shipped out to any
release yet.

Differential Revision: https://phabricator.services.mozilla.com/D136371

UltraBlame original commit: 81f10eb2a1b34cfa959a029b41c81dd76e837608