Bug 1750937 - Don't restart test runner activity when orientation is changed. r=geckoview-reviewers,owlish
The web-platform-tests of Screen orientation API is often failure due to
"Request for fullscreen was denied because requesting element is not in the
currently focused tab.". Because, when orientation is changed, activity is
restarted.

So we should prevent restarting activity when changing it.

Differential Revision: https://phabricator.services.mozilla.com/D136441

UltraBlame original commit: 323f3c52b376c577c6288b757cae7f17413759bc