Bug 1751162 - Part 2: Add a nursery size heuristic based on the length of collection time r=sfink
This limits the growth rate based on the length of collection time, such that
size is proportionally reduced when the time is longer than 4ms.

The existing 'timeFraction' got renamed to 'dutyFactor' to disambiguate.

This doesn't happen during pageload so as not to blow up all our page load
benchmarks.

This is pretty much a wash on octane although it improves splay latency by 3.8%
for me.

This should drastically reduce the number of long minor GCs we see. I expect
we'll see a change in the telemetry.

Depends on D136636

Differential Revision: https://phabricator.services.mozilla.com/D136637

UltraBlame original commit: 2c00bc26b97a10e0767b614e4ba89047941dcc18