Bug 1743022 - Only call AddToBlocklist once r=necko-reviewers,kershaw
On windows getaddrinfo does not return TTL info for the DNS record. so we do a
second DNS resolution that uses DNSQueryA to get the TTL. We should only call
AddToBlocklist after the first DNS resolution completes. Doing so again for
the DNSQueryA refresh is unnecessary and possibly extends the lifetime of the
blocklist past the 60s that is default.

This patch also reenables the blocklist when TRR strict mode is enabled.

Differential Revision: https://phabricator.services.mozilla.com/D136715

UltraBlame original commit: c056411bccbdd66c29b4eb017b220355b192fbd8