Bug 1751005 - Log when unable to find profile folder. r=mossop
Ordinarily, GeckoView apps are not allowed to set the profile folder, which is
always the default one. Furthermore, app's folders (which include the profile
folder) are private to the app and non-root users are not able to access them
or modify them.

For tests, however, we allow setting the profile folder through a special
config file.

If Gecko is not able to access the profile folder (e.g. because the folder was
deleted by the test but the config file is still present), it will error out,
and fire an Alert dialog to let the user know that the profile folder could not
be accessed.

On GeckoView, we cannot handle this Alert dialog as it's not originating from a
Window.

Before this patch, GeckoView would silently exit without giving a reason.

Given the test-only nature of this problem, we just log a statement and return,
so developers can troubleshoot the problem.

Differential Revision: https://phabricator.services.mozilla.com/D136545

UltraBlame original commit: 34fce4ae0c3c204cc916c118e58b212dfaeafa20