Bug 1715976 - Add test case for concurrent concurrent delazification. r=arai
This patch adds a test case which creates trees of inner functions, using
`isInStencilCache` to count the number of delazified functions at the time of
execution.

Unfortunately, there is no good way to test the stencil cache as it meant to be
a no-op in terms of execution. Thus, this test case checks that functions are
added to the stencil cache, and `waitForStencilCache` to make sure when the
cache is filled, then the executed function still works fine.

Differential Revision: https://phabricator.services.mozilla.com/D136342

UltraBlame original commit: 3adc56a7e6043470574f5aac8299ba0f9087ffdb