Bug 1751816 - Improve error handling in GLContextEGL::CreateCompatibleSurface. r=jgilbert
Rather than just crashing, if we return a nullptr in
CreateCompatibleSurface, then we will just cause a context loss for
WebGL instances. This would be better for beta/release than just
crashing the content process on the user.

So that we can get a sense of what errors are causing this, we should
record it to the critical log. This will still crash on nightly as well
so we should get the reports.

Differential Revision: https://phabricator.services.mozilla.com/D136825

UltraBlame original commit: dd2af9af327b55576ddc8d99735e87d84c041498