Bug 1717683 - Make ZIP packages and l10n repacks deterministic by using `zip.py` instead of `make_zip.py`. r=firefox-build-system-reviewers,glandium
The compression options appear to be the same: `mozjar` defaults to
level 9 compression.  The `mozjar` code is definitely slower than the
native ZIP executable, but not, I think, by enough to not do this.

This is the sole consumer of `make_zip.py` in the tree, so we remove
the file entirely.

Differential Revision: https://phabricator.services.mozilla.com/D118528

UltraBlame original commit: 3fcafb7a49465b5b3167a10511e22a307ff726df