Bug 1740304: Apply the print fragmentation-fallback "additional offset" in one case where it was being missed. r=miko
We call MaybeApplyAdditionalOffset a few lines up when setting *aOffset, but it
looks like we missed this particular codepath.  This was causing content to
render properly during print-preview (which takes the upper codepath) but not
during actual printing (which takes the lower codepath, the one where I'm
adding the new call).

Differential Revision: https://phabricator.services.mozilla.com/D137199

UltraBlame original commit: 073faa92644b692210105055ce91669b2db317ef