Bug 1673513 - Generate profile in new worker thread, and log its progress - r=florian,KrisWright
In order to keep the child process responsive to profile IPCs, the heavy work of generating the profile JSON is now done in a separate thread.
A `ProgressLogger` is used to keep track of the progress of this work, and the progress value is stored in a shared atomic `ProportionValue`.
When the JSON profile is ready, the final shmem allocation (used to send the profile to the parent process) is done on the original "ProfilerChild" IPC thread.

Differential Revision: https://phabricator.services.mozilla.com/D135484

UltraBlame original commit: 8ebc963fd9e6df91da1dbf17e7f6e526de0b5254