Bug 1749995 - Make tests in browser/base/content/test/ dFPI compatible, r=mossop,anti-tracking-reviewers,timhuang
We are fixing mochitests that fail when network.cookie.cookieBehavior = 5, i.e. when we enable Total Cookie Protection.
This is most often due to the test assuming that an origin will always have access to its storage state when embedded as
a third party.

My approach: Add third-party storage permission to the favicon's origin (http://example.com) for each test.
The feature tested here assumes third-party storage, so we have to give it to expose the code paths being tested.
In this case, those paths are to send cookies with a favicon request when `crossorigin="use-credentials"` in the favicon's link tag.

Differential Revision: https://phabricator.services.mozilla.com/D136600

UltraBlame original commit: 80ad65658ac7b97ac436d09f3eac7c4064f8ac92