Bug 1253706: Let failure to init a non-obsolete stream slide. r=mjf
This can happen in cases where we have an old ICE ctx that is infeasible, but have
kicked off an ICE restart with a new configuration that _is_ feasible. The old streams
are not obsolete yet, because we have not committed the ICE restart.

Differential Revision: https://phabricator.services.mozilla.com/D135375

UltraBlame original commit: fe16d24ab482f4575e62229fca59bdf397f0fc31