Bug 1746104 - part 3-1: Get rid of `WritingMode` argument of `IMContextWrapper::CollapseTo` r=m_kato
It's always overwritten with same value so that it is not necessary at least for
now and on Windows, we just do not update it in same situation.  Therefore,
we should drop it until we become need to use it.

Differential Revision: https://phabricator.services.mozilla.com/D137414

UltraBlame original commit: 8ce50cd6760ec123bb481ff21340dde28cd6c276