Bug 1253706: Let failure to init a non-obsolete stream slide. r=mjf
This can happen in cases where we have an old ICE ctx that is infeasible, but have
kicked off an ICE restart with a new configuration that _is_ feasible. The old streams
are not obsolete yet, because we have not committed the ICE restart.

Differential Revision: https://phabricator.services.mozilla.com/D135375

UltraBlame original commit: aaa4dfe8b3925de8430c5a41ff04ccb87f1090b5