Bug 1753025 - Using correct principal to create channel for NavigationPreload. r=dom-worker-reviewers,jesup
Currently using the pricinpal of InternalRequest to create the channel for NavigationPreload.
However, it is not a correct principal and makes NavigationPreload stop with channel security checking.
Instead of using the principal of InternalRequest, this patch uses the loadingPrincipal of InterceptedHttpChannel to create the channel.

Differential Revision: https://phabricator.services.mozilla.com/D137594

UltraBlame original commit: 7692c70b812d95fa7d68cc3be1df5430db31ab0d