Bug 1753374. Don't allow pending vv offset updates to become the vv offset purely on the content side without asking apz. r=botond
ScrollFrameHelper::GetVisualViewportOffset will return a pending vv offset if there is one. So if there is a pending vv offset this line will instantly turn it into the current vv offset. This is not how the pending vv offset is supposed to work. It should get sent over to apz, apz can then decide to accept or reject it based on whether it knows about user scrolling that overrides it.

I saw this while reading code for something else, so I don't know of anything specific this fixes.

Depends on D137720

Differential Revision: https://phabricator.services.mozilla.com/D137721

UltraBlame original commit: acd31d1d0fce78ac267179147d60900fb90d8296