Bug 1753586 - Navigator.cookieEnabled should return true when cookie is partitioned. r=smaug,anti-tracking-reviewers,pbz
We change to return true for 'navigator.cookieEnabled' when the cookie
is partitioned. It did return false because the storage access is not
allowed in this case. But, acutally, we should return true because the
cookie is available.

Also, we should still return true in the case where accessing cookies
gets a security error, e.g. a sandboxed iframe. This behavior matches to
the spec.

Differential Revision: https://phabricator.services.mozilla.com/D137865

UltraBlame original commit: 90571f6d85212bb5c63d57b49c3e72b81614de74