Bug 1752871 - Add option to specify which page to record. r=perftest-reviewers,AlexandruIonescu
This patch adds the --proxy-perftest-page option. With this option, we'll be able to specify which test pages we want to record rather than modifying the pageload_sites.json. Note that the login fields will not be taken into consideration with this option.

You can use the flag as follows (seperate multiple pages with a comma): --proxy-perftest-page microsft,linkedin,netflix

Furthermore, some changes were made to strengthen the conditions around when a login site can be tested (only when RAPTOR_LOGINS is defined locally, or if we are in CI).

Differential Revision: https://phabricator.services.mozilla.com/D137468

UltraBlame original commit: 0e476137b09d9b2832c851a7cf8e1b5cf1b448c7