Bug 1754495 - Don't use the audio clock to determine if an AudioStream is low on decoded audio. r=alwu
The old technique was slightly misleading. It was counting the audio already
sent to the device, in the decoded audio duration. But in fact, there's nothing
that can be done about this, the audio is going to be played regardless of what
happens. If the audio latency is high enough, this would return a number that's
high, but not a lot of audio would be sitting in Firefox's buffer, ready to
service the audio callbacks.

It's best to only consider the decoded audio (processed + unprocessed) here,
since the goal of this is to determine whether more audio needs to be decoded.
It also saves a `cubeb_stream_get_position` IPC call if using AudioIPC (soon to
be all platforms).

Differential Revision: https://phabricator.services.mozilla.com/D138318

UltraBlame original commit: 6dbf53f67060d45c9661457637cc8e4241688d04