Bug 1754680: Don't treat a list item bullet as an embedded object. r=eeejay
This was causing LinkAt to return a bullet instead of the first embedded object in a list item, breaking rendering with NVDA when the cache was enabled.
Rather than adding another role to RemoteAccessibleBase::IsEmbeddedObject, I changed it to use IsText().
This is consistent with how embedded objects are determined for LocalAccessible in EmbeddedObjCollector.
I think the reason this previously used roles is that when this was originally implemented, AccTypes weren't remoted, so IsText() couldn't be used.

Differential Revision: https://phabricator.services.mozilla.com/D138506

UltraBlame original commit: 1ae02b556de81865221de4c40d3ed9c2f35adb67