Bug 1755552 - Make tests in dom/workers/tests/ dFPI compatible, r=smaug
This is very similar to a few fixes we've sent your way already, like https://phabricator.services.mozilla.com/D136608. We are merging in the dFPI enabling test config now that all of the test fixes have rolled in, but this test beat us in the race condition.


Reminder:

We are fixing mochitests that fail when network.cookie.cookieBehavior = 5, i.e. when we enable Total Cookie Protection.
This is most often due to the test assuming that an origin will always have access to its storage state when embedded as
a third party.

In this case I had to request storage access to allow an embedded iframe to use a BroadcastChannel to communicate with top-level pages of the same origin.
This was only a problem in the XOrigin case where the test html file is loaded in an embedded iframe and then can't communicate with the top-level window it created.

Differential Revision: https://phabricator.services.mozilla.com/D138850

UltraBlame original commit: 9ab6b46b773172ffa4a0ca23420733fb677c3409