Bug 1755988 - Switch back to CreateSimilarDrawTarget for luminance masking. r=gfx-reviewers,aosmond
Bug 1285857 changed from using CreateSimilarDrawTarget on D2D to using a
software DrawTarget because of the need to convert the result to
luminance. However, after that bug 1417903 changed the luminance
conversion to use the IntoLuminanceSource which is supported directly by
D2D which removed the need for a software DrawTarget.

This will let us use a recording DrawTarget which will avoid serializing
the image data and is step toward switching to CreateClippedDrawTarget.

Differential Revision: https://phabricator.services.mozilla.com/D139046

UltraBlame original commit: cbc09e020dd3171545e89cce34b7d3813c19fd6e