Bug 1736202: browser_interactions_typing.js intermittently misses a single keypress. r=Standard8
Our typing interactions separate typing into sessions. If you haven't typed for longer than a
certain period of time we count the previous session as over and start afresh. Any sessions
containing just one keypress are discarded, the rest are aggregated.

In our automated tests we want to test that this detection of a new session of typing works and our
only way to do that is to actually delay the test for longer than the maximum time between
keypresses. So the test runs fast we reduce the maximum time pref to 50ms.

Unfortunately I've experimentally seen that in CI on debug and asan builds that when not testing
this aspect and sending a sentence of characters we can still occasionally see delays of over 50ms
between individual characters. If this delay happens between the first and second character or the
last and last but one character then we end up with a typing session with a single character in it
which gets discarded leading our interaction metrics to be short one keypress.

This patch removes the reduced preference value for general tests and sets a value of 150ms for the
few tests where this matters.

Differential Revision: https://phabricator.services.mozilla.com/D139039

UltraBlame original commit: 378b7f9b39a42c2d2780a36ea786aaf686bdfe02