Bug 1755343: browser_interactions_scrolling.js sometimes times out after arrow down scrolling. r=Standard8
Scrolling tests need to add an event listener for the scroll event then send the scroll action
and then wait for the scroll event to be seen in order to be sure that the interaction has been
captured. We also disable smooth scrolling which causes a single scroll to become multiple scroll
events of different distances. Sometimes the first distance is 0 leading to us detecting a scroll
distance of 0.

The test code for adding event listeners (BrowserTestUtils.waitForContentEvent) returns after
sending information about the event listener to the content processes but before the event listener
has actually been added. If we trigger scrolling using a keypress the keypress is dispatches to the
content process as a higher priority IPC message than that used to add the scroll event listener and
so in some cases the scroll can occur before the event listener has been registered. Here we do a
simple round-trip through the content process which will be sent with the same priority as the event
listener registration so we can be guaranteed that the event listener is ready before we attempt to
scroll.

Depends on D139039

Differential Revision: https://phabricator.services.mozilla.com/D139104

UltraBlame original commit: 367b27a4bd9fcddbf1ed61ff54789dd3ba043bf2