Bug 1755817 - Move LandmarkRole to Accessible. r=Jamie
Current mochitests should prove the soundness of this. The only thing
preventing this from becoming a non-virtual method is the reliance
on TagName which is only available when cached is enabled. When it
is disabled we need to fallback on the sync IPDL call.

Differential Revision: https://phabricator.services.mozilla.com/D138964

UltraBlame original commit: 441765333d9e24ca5fefa2c2cd73cd4858ef08c1