Bug 1752026 - Implement Accel + Up/Down shortcuts in findbar. r=mcheang,NeilDeakin
When the findbar input is focused, Up/Down keys scroll the page up or
down by one line. But pressing the Accel modifier doesn't modify the
behavior to scroll to the top/bottom of the document as it normally
does. It just produces a normal scroll. This patch fixes that.

Differential Revision: https://phabricator.services.mozilla.com/D137606

UltraBlame original commit: f8d7b95aef7c5b36be795c3184236963f059d702