Bug 1755736 - [devtools] Test original source being removed and replaced in the bundle by a new original source. r=bomsy
I'm using a new dedicated bundle here so that we ensure that the bundle lines don't change and stay "breakable".
Otherwise the breakpoint wouldn't be hit on reload if the breakpoint become an empty line.
My goal here is to see what happens if the new bundle still hit the breakpoint even if the original source
was removed.

We might later revisit this behavior to remove this breakpoint.
It might be challenging to avoid hitting the breakpoint once on the first load.
It would require the server to know about sourcemap in order to ignore the breakpoint on early page load sequence.

Differential Revision: https://phabricator.services.mozilla.com/D138920

UltraBlame original commit: bb31441f4fc47255fd8d03af525a1483f24d4833