Bug 1752332: Improve the blocklisting behavior r=KrisWright
We sanitize a preference (when sending from the parent
process) if it is in the blocklist, or if does not
have a Default value (i.e. it is dynamically named).
There is an exception list for dynamically named
preferences we know we need though.

In subprocesses, we know if a preference was sanitized
by checking its Sanitized bit.

Differential Revision: https://phabricator.services.mozilla.com/D138685

UltraBlame original commit: 593e5e138927c86da6ffab352e579b435716df51