Bug 1752348. Adjust transform on maskTarget instead of ctx. r=nical
Previously, if ctx->GetDrawTarget() != maskTarget we end up adjusting the
matrix on ctx and not on the maskTarget.

Also, now that we're doing the correct adjustment of the maskTarget transform
we can remove code to set maskTarget's transform to identity.

Differential Revision: https://phabricator.services.mozilla.com/D139133

UltraBlame original commit: eeabb68029e166cb9c507576741ea8e9af0f4773