Bug 1750791 - Part 2: Convert UntrustedModulesData::mEvents to a linked list. r=mhowell
This patch chnages the type of `UntrustedModulesData::mEvents` from
`Vector` to `AutoCleanLinkedList`.  The purpose of this is for
`UntrustedModulesBackupService` to splice data easier and to discard
old loading events, which will be implemented in a subsequent patch.

Differential Revision: https://phabricator.services.mozilla.com/D137245

UltraBlame original commit: 43a5f02b11607c1a102c0a6ef08408ec3fe84037