Bug 1756251 - make GeckoJavaSampler.Frame thread safe. r=geckoview-reviewers,julienw,agi
This class was not thread safe because the fields in this class were mutable
and access to mutable state must be synchronized across threads. This class was
only used in an immutable fashion, however, so it was trivial to make it
immutable and thus thread safe.

Differential Revision: https://phabricator.services.mozilla.com/D139194

UltraBlame original commit: c216f871ebea16570c78860bb19a6d5a83dea486