Bug 1757221 - Always round up in WorkerPrivate::RescheduleTimeoutTimer when calculating the delay for timers r=smaug
When the delay value gets converted from double to integer, it
would lose its precision. So the delay of the timer is going to
be smaller than the actual delay that the user expects (by a
tiny bit though).

This patch changes the behavior to always round value upwards.

Differential Revision: https://phabricator.services.mozilla.com/D139773

UltraBlame original commit: 1eb0b24a4248fcec28864419fb71f1988748025d