Bug 1756385 - Widgets to appear unchecked in customize mode. r=jaws
This patch adds a CSS rule to override rules that set background-color on
"checked" toolbar buttons, while in customize mode. Because the checked state is
no longer visible in customize mode (which mainly affects the sidebar toggle
widget), we could use the Accel+B hotkey to hide or show the sidebar without any
visual feedback, potentially causing confusion.

So this patch also adds a check for customize mode in SidebarUI.toggle(). It
might be more sophisticated to wrap the method in a command element that can be
updated by CustomizeMode.jsm, and link the key elements to that command, so it
only impacts the shortcuts. But for now I think a more minimal change is fine.

Also add a test to confirm that the above works correctly.

Differential Revision: https://phabricator.services.mozilla.com/D139625

UltraBlame original commit: 923265032cd04c7d82561f4544c2666015600c9d