Bug 1755318: Ensure that PeerConnectionImpl::Close() is called before destruction. r=mjf
In some cases, PeerConnectionImpl's refcount will go to 0 without needing the
cycle collector to call unlink on it, which is what we rely on to call Close().
So, we make sure that refcount cannot go to 0 until Close is called (either
due to an unlink, or an explicit close from JS or PeerConnectionCtx).

Differential Revision: https://phabricator.services.mozilla.com/D139216

UltraBlame original commit: cedddaa51509f851639ae5a3e0ab1f7dfd2dbc5d