Bug 1686344 - Add test for `--backgroundtask --jsdebugger --wait-for-jsdebugger`.  r=devtools-reviewers,jdescottes
This version is as simple as I can make it.  It simply expects the JS
debugger to stop on the breakpoint added automatically by the
backgroundtask debugger command line processing (using
`setBreakpointOnLoad`) and disconnects, expecting the task to continue
execution and exit with exit code 0.

In the future, we'd like to interact with the task environment, for example to:

1. stop on the automatic breakpoint
2. continue
3. stop on a `debugger;`
4. set the task's exit code from a failure code to exit code 0
5. continue
6. verifies the tasks's exit code is 0.

Sadly my attempts to do this fail intermittently in automation.

Differential Revision: https://phabricator.services.mozilla.com/D139156

UltraBlame original commit: ce693e8e4210e8c101f25ec68027ace5366e9ee9