Bug 1757573 - Use the correct zone when looking up nuked WeakRef CCWs r=sfink
WeakRefs are exposed to the user so anyone can create CCWs to WeakRefs. Don't
assume that a nuked WeakRef CCW is one that we created but get it's target to
find which zone to look it up in.

Differential Revision: https://phabricator.services.mozilla.com/D140044

UltraBlame original commit: d30323c51a42996c0cbab0597f15b7abf57b720a