Bug 1754339 - Route HandleTap message from GPU process through android's priority queue. r=botond
In response to a touch down and up event on Android,
NPZCSupport::FinishHandlingMotionEvent sends the events over the Input
Bridge to APZ, then dispatches the event to the main thread for the
widget to process them (which will in turn send them to the tab). It
uses android's priority event queue to do so, via
nsAppShell::PostEvent.

Meanwhile, the AsyncPanZoomController will detect these events as a
tap gesture, and call RemoteContentController::HandleTap. When the GPU
process is disabled, and APZ therefore lives in the parent process,
this calls BrowserParent::SendHandleTap to forward the tap event to
the tab. On Android, however, we ensure that we first dispatch through
the same priority queue used when dispatching the touch events to the
widget, otherwise the tap event could be sent to the tab before the
events from which it was synthesized.

With the GPU process enabled, however, APZ lives in the GPU process
and RemoteContentController::HandleTap must call
APZCTreeManagerParent::SendHandleTap to send the event back to the
parent process, which in turn calls BrowserParent::SendHandleTap to
send it to the tab. However, because IPDL uses a different event queue
than we used to schedule the widget's processing of the events, there
is no guarantee that APZCTreeManagerChild::RecvHandleTap will be
scheduled after the pending tasks on the android priority queue.

This means we sometimes end up calling BrowserParent::SendHandleTap
before BrowserParent::SendRealTouchEvent, and content can receive the
events in the incorrect order.

To fix this, we simply ensure that APZCTreeManagerChild::RecvHandleTap
runs through the android priority queue before calling
BrowserParent::SendHandleTap, ensuring the events are sent in order.

Differential Revision: https://phabricator.services.mozilla.com/D140326

UltraBlame original commit: 25a2fb150bd4f3235c5b703175f51356b220c6f0