Bug 1758553 - Remove fragments from URLs in Snapshot API r=mossop
When recording interactions, we strip out #hash fragments from URLs
see:
https://searchfox.org/mozilla-central/rev/9ca193b4233957439583f2eadabbd3cfb4cd9fed/browser/components/places/InteractionsChild.jsm#172-173

However the Snapshots api doesn't do this, and so we can end up with unexpected results -- for instance the SnapshotSelector suggests the same page that the user is on because the context url is the full url, while the Interaction and the Snapshot are for the #fragment-stripped url.

We decided to also strip fragments from urls provided to through Snapshots api.

Differential Revision: https://phabricator.services.mozilla.com/D139975

UltraBlame original commit: cb5597cb0e10f441cc790b6c029e3a878e409d7f