Bug 1748945 - Change isOpen to UiThread. r=calu
The fact that isOpen is AnyThread seems to be a mistake, I don't think we can
reliably tell a consumer if a session is open on a background thread.

Even with synchronized access to mWindow code like this would always be broken
on a background thread:

```
if (!session.isOpen()) {
  session.open(sGeckoRuntime);
}
```

because another thread could open the session between the isOpen() and the
session.open statements.

This patch changes isOpen to UiThread. I verified that all calls actually
happen on the UI thread, except for the calls in GeckoSessionSettings where we
are trying to handle the case where a session was closed and _then_ we would
allow modifying read-only settings. I think in that case we should ask the
embedder to use a new SessionSettings instance instead of forcing all calls to
be UiThread.

Differential Revision: https://phabricator.services.mozilla.com/D137784

UltraBlame original commit: 2149ff735175ace6d03294f39c6488d5b96590aa