Bug 1758617 - Make Firefox abort in case of content crash during PGO run. r=firefox-build-system-reviewers,mhentges
The expected behavior during a PGO run is that a browser is started and
exited once, and then it is started again on a page that runs various
workloads and once finished, exits.
When workload happens to crash the content process, the browser is left
running indefinitely, until the taskcluster task itself times out. This
leaves us with no possibility of knowing what actually went wrong during
the run, which is about the worst thing that can happen.
With the browser shutting down on its own in case of crash, the harness
can find the minidumps and report the crash, which is more useful.

Differential Revision: https://phabricator.services.mozilla.com/D140678

UltraBlame original commit: 992fb0f6343fea017f90b7cf6b96d29e5e52cea0