Bug 1756863 - Pass the external test root found in raptor to condprof. r=perftest-reviewers,AlexandruIonescu
This patch modifies condprof to accept an optional `remote_test_root` argument so that we can change where the profile gets created and pulled from. Then we pass the external test root through this parameter from raptor to condprof. This fixes an issue on the Samsung A51 device as it can't store the profiles in the default location.

Differential Revision: https://phabricator.services.mozilla.com/D139517

UltraBlame original commit: a0918b8065eb37983665949f2c3b4919a1eda89b