Bug 1754459 - Improve caret position selection for images. r=dholbert
What saves us in the non-pointer-events: none case is silly, and is
this loop that hits the draggable image:

  https://searchfox.org/mozilla-central/rev/fa69d8b248e6c1df670aa6b019e30ec37e6672be/layout/generic/nsIFrame.cpp#4752

Needs tests (if it doesn't affect existing ones), but uploading for some
initial feedback. I think this makes a lot more sense than the existing
code, and works for editing and dragging the same way, so it's nice to
have less special-cases.

Differential Revision: https://phabricator.services.mozilla.com/D141242

UltraBlame original commit: 57845918277d9f89c6878b0659fd930791e46fa0