Bug 1759792 - Avoid frequently oscillating between Skia layering and WebGL drawing. r=aosmond
In certain use-cases where we support layered drawing to Skia above a WebGL context, we can
repeatedly oscillate between clearing the Skia layer, drawing to it, then flushing it to
the WebGL context to fulfill an accelerated drawing request. For example, StrokeLine followed
by DrawSurface, repeated in a cycle, will cause each StrokeLine request to go to the layer,
and then the following DrawSurface request will go to the WebGL context, with clearing and
flushing of the Skia layer in between.

The original intent of the code was that the performance benefits of accelerated drawing
would compensate for the occasional cost of flushing. However, in practice, this turns out
not to be true and repeated clearing and flushing costs for the Skia layering can dominate
the overall performance. We really just want to continue drawing into the Skia layer as
long as possible if some fallback causes us to initially set it up and use it. While not
ideal, it prevents from bad performance cliffs from occuring where flushing happens more
than once in a frame.

Differential Revision: https://phabricator.services.mozilla.com/D141270

UltraBlame original commit: 71f586ff47826e6580155cbdc4c8e7b7f349ee48