Bug 1754017 - Make the provided stream cloneable eagerly in StreamBlobImpl, r=asuth
This patch moves the logic to ensure the stream is cloneable to during
StreamBlobImpl's constructor, allowing it to customize the properties of
the nsPipe used for non-cloneable streams, and ensuring that the blob is
storing a complete copy of the streamed data without requiring it to be
consumed.

This codepath is currently fairly infrequently used, but will be used
more frequently once DataPipe is used for IPC streams.

Differential Revision: https://phabricator.services.mozilla.com/D141033

UltraBlame original commit: 297bda3439b29163f959cd746502c42852a4d42e