Bug 1756693 [wpt PR 32945] - Fix compositor keyframe snapshot., a=testonly
Automatic update from web-platform-tests
Fix compositor keyframe snapshot.

Background color animations can run on the compositor via native paint
worklets; however, the rules for when to composite were decoupled
between the paint and animation code. It was possible to get into a bad
state. During style update, we snapshot compositor keyframes, but we
were incorrectly using the element and not the animating_element in the
call. This left the snapshot in a bad state, in which the offsets were
not populated.  This was not a problem for other pseudo-element
animations after the fix to use the correct Node in box_painter_base.cc
as long as the animations should be composited since there was as
secondary call site that updated the snapshot.  The offsets themselves
were extracted in a round about way since stored directly in the
keyframes.

Adding timing checks to the native paint worklet code, or finding a way
to streamline the checks between paint and
Animation::CheckCanStartAnimationOnCompositor are left to a followup
CL.This patch addresses the source of the crash, and not the
performance impact of compositing an infinite duration animation.

Bug: 1298148
Change-Id: I80c8c8ae5811341e447559d1951d27229a77f384
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3475165
Reviewed-by: Anders Hartvoll Ruud <andruudchromium.org>
Reviewed-by: Robert Flack <flackrchromium.org>
Commit-Queue: Kevin Ellis <keverschromium.org>
Cr-Commit-Position: refs/heads/main{#974347}

--

wpt-commits: f9fbccb2ed1c22ecead9750958e2b3efa34c487e
wpt-pr: 32945

UltraBlame original commit: 93c5d63239d9a4d3ddc9de077e5517d13f99049c