Bug 1756087 - Amend sandbox policy for libdrm r=jld
Before this change, it was assumed that	readlink operation might be
performed on /sys if the driver is AMD. However, the operation would
always be performed by Mesa via libdrm if the device is PCI. In fact,
blocking the operation breaks virtio_gpu.

The readlink operation is part of invoking
realpath("/sys/dev/char/<PCI>/device/subsystem") so the read only
permissions for the file and the ancestor directories are added.

The permissions for the resolved real directory and its files are
already set, but the directory path is modified in libdrm when the
device is virtio_gpu. The path modification is also ported to the
sandbox policy.

Differential Revision: https://phabricator.services.mozilla.com/D139095

UltraBlame original commit: 184fafcf0523a4706b3f765c3d901ba79b1b74ef