Bug 1642412 - wasm: Don't read a call site line in operations that AsmJS won't emit one for. r=lth
AsmJS will emit a call site line number for call, call indirect, and math builtins.
We currently read from the call site line number array for more operations than
these. This is normally fine, as long as the operations that try the read are
never emitted from AsmJS. If that is not the case, we will get out of sync and
can read outside of the array. This caused an issue when I added a use of
readCallSiteOrBytecode() in TeeGlobal, copying from SetGlobal. This commit
fixes this.

Differential Revision: https://phabricator.services.mozilla.com/D141030

UltraBlame original commit: 82ac83afeaeaa3665037ea587c614fddaa8b648d