Bug 1761273 - Check for multiple dictionaries when saving site preferences; r=smaug
We only want to save the site preferences when the user's dictionary
preferences do not match the preferred language for the site. This is always
the case when there is more than one dictionary in use, but the current
code does not handle this case.

Differential Revision: https://phabricator.services.mozilla.com/D141981

UltraBlame original commit: e39d2d3f51175d2914946fb82560f2fd9c3f4580