Bug 1752665 - Make the Spotlight schema less strict with respect to images r=emcminn
The previous changes to the Spotlight schema made one of the logo.imageURL or
logo.imageId required. However, this is not backwards compatible with the older
Firefoxen, so we have to support the case where `logo` is present (e.g., for
`image.size`) without either of those properties, but `logoImageURL` *is*
present.

Differential Revision: https://phabricator.services.mozilla.com/D140761

UltraBlame original commit: 37895d08ff95f2337958dff746ef5ab84330cfcb