Bug 1758710 [wpt PR 33126] - Fix absolute positioned elements inside multi-column, a=testonly
Automatic update from web-platform-tests
Fix absolute positioned elements inside multi-column

In r974634 we introduced a bug related to absolute positioned elements
when we have a mixed legacy and NG layout tree (like in the case of
multi-column).

The problem is in NGOutOfFlowLayoutPart::SweepLegacyCandidates(),
there is a check "legacy_objects->size() == placed_objects->size()"
which is no longer right, as we're not adding all the abspos elements
to PositionedObjects() list (actually that's the main change in
the patch that introduced the regression, to avoid calling
UseLegacyOutOfFlowPositioning()).

In the test case, the first abspos element which is processed
at NGOutOfFlowLayoutPart::LayoutCandidates() (but now is not added
to PositionedObjects()). Then when we're in SweepLegacyCandidates()
we have this first abspos element in |placed_objects|,
and |legacy_objects| has also one element, but not this one,
the second abspos element.
Comparing the size of |placed_objects| and |legacy_objects|
is not enough now that we're not calling
UseLegacyOutOfFlowPositioning().
So this patch checks if all legacy objects have been
already placed or not.

Fixed: 1304371
Test: web_tests/external/wpt/css/css-position/position-absolute-multicol-001.html
Change-Id: Ie2c017d277d1e2c86b0529f29e3176fb3fd9e263
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3512572
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Reviewed-by: Oriol Brufau <obrufauigalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: Manuel Rego <regoigalia.com>
Cr-Commit-Position: refs/heads/main{#979697}

--

wpt-commits: b7d099cdd3d6c01c848a6ec52a9242d8f5bd1a7f
wpt-pr: 33126

UltraBlame original commit: ecee64945e24271bd668f675e3e0c052947949a1