Bug 1758730 - Ensure no errors or warning show on search engine prefs; r=Standard8
The live language reloading triggers code that will run that assumes the view
is visible, when in fact it may not be on the dom. This fixes that. In addition
it fixes the error where the engine tries to update after it has been removed.

Differential Revision: https://phabricator.services.mozilla.com/D141685

UltraBlame original commit: 552c4c35d3318e62a34a5696d2a1ad6c217d39e4