Bug 1762100 - Part 3: Expect and use `FirefoxPDF-...`. r=bhearsum
This does two things.  First, it expects the special
FirefoxPDF-... ProgID to be available for UserChoice.  We could manage
without it for a while, but eventually we expect set-to-default to
include PDFs.  When that is the case, if it doesn't exist, something
has gone very wrong, and we'd like to find that out (via our existing
telemetry, which reports "missing" ProgIDs).

Second, it arranges to use the new FirefoxPDF-... ProgID when
setting-to-default.

Differential Revision: https://phabricator.services.mozilla.com/D142303

UltraBlame original commit: 4f5619be4f9c8ef3b7231deac012e80055a2b23b