Bug 1760979 - Fix wpt verify jobs, r=karlcow
This was regressed by the changes in PR 32403.

Also add a test to check that we don't regress in the same way
again. This mocks out actually running the tests, which of course
means we aren't testing that we get the right kind of return value
from wptrunner.

Differential Revision: https://phabricator.services.mozilla.com/D142100

UltraBlame original commit: ce95f2726565684f6ac8a491c3ddf6e182e28832