Bug 1760286 - Add metadata for all eslint-plugin-mozilla rules. r=Gijs
This metadata is based on what I thik we should include. We don't need most
of the documentation section (category, description) since they are specific
to ESLint. The url is useful, as it will show up in VSCode and other editors
for developers wanting more information.

The type of rule probably isn't important, but is kept as a general guide.

Differential Revision: https://phabricator.services.mozilla.com/D142286

UltraBlame original commit: 026a3213aeffde7b49e6fa554a5eaf1cdb644c34