Bug 1762388 - Store reference to WebRenderBridgeChild in WebrenderLayerManager even if already destroyed. r=aosmond
If the compositor is destroyed whilst in the process of being
reinitialized, it is possible for a WebRenderBridgeParent to be
created in an already-destroyed state. In this case, we currently
avoid saving a reference to the WebRenderBridgeChild in
WebRenderLayerManager. However, this means that
WebRenderLayerManager's destruction does not cause the
WebRenderBridgeChild (and therefore the corresponding
WebRenderBridgeParent) to be destroyed, meaning its layer tree ID does
not get erased from the map.

This patch makes us unconditionally store a reference to the
WebRenderBridgeChild in WebRenderLayerManager, meaning it is correctly
destroyed and the layer tree ID is correctly erased from the map.

Differential Revision: https://phabricator.services.mozilla.com/D142547

UltraBlame original commit: 0fbddbe55d8f6f5ec7c6cad6b725e09a96ce9439