Bug 1762517 - Add support for render tasks that target existing locations r=gfx-reviewers,lsalzman
This incremental patch adds the main support for render tasks that
bind to the same allocated location as an existing render task. It's
used for situations like backdrop-filter, where we write to a target,
then make that task an input to an arbitrary filter chain, and then
begin writing back to the same location again (but with the previous
existing task as an input dependency to the filter chain).

As part of this work, remove the `target_inputs` functionality which
was never used and complicates things. This required making the
assign_free pass be recursive (which now matches the assign_render_on
pass).

Differential Revision: https://phabricator.services.mozilla.com/D142643

UltraBlame original commit: 75e7df062d5a62d662fd72c22065c47c0f668ea6