Bug 1759332 - [devtools] Remove options.hideIfOffscreen for highlighters infobar elements r=Honza
This feature is currently not tested.

Summary of the issue: we have up to 2 tooltips for grid elements: cell and area.
To avoid overlapping, today they are restricted with two options: `position` (top for one, bottom for the other) and `hideIfOffscreen`.
`position` already ensures that, when there is enough room, the tooltips will not overlap.
However, when they don't have enough room to be displayed, they will start "sliding" to remain in the viewport.
This means that under extreme conditions, you could again have an overlap between the two tooltips.

To avoid this, the `hideIfOffscreen` option was added. Instead of sliding, the tooltip are purely and simply hidden.
While I understand the idea, I think this creates more problems than it solves. For the sake of avoiding rare overlaps, we just don't display the tooltips in many situations.

Ideally we would have a better logic to place the 2 tooltips in compatible positions, or maybe we should even combine them.
But this is an easy way of fixing the issue here.

Differential Revision: https://phabricator.services.mozilla.com/D142203

UltraBlame original commit: 22f78b913611372e5aabccb621edac33c67030c8