Bug 1761511 - Part 1: Set up moz.yaml vendoring for gtest, r=ahal,tjr
This doesn't actually vendor the library in, but sets up the basic
configuration to allow it to be vendored in.

The latest commit on the main branch, rather than the latest tag, is used as
recommended by the googletest documentation.

Differential Revision: https://phabricator.services.mozilla.com/D142608

UltraBlame original commit: 1afb1f32a6a475767cc87c1c75e894d89c07250e