Bug 1763351 part 2 - Always use a register for LInstanceOf* RHS. r=iain
We can simplify the code a bit now that RHS is usually a register instead of a
constant.

A constant object can still show up for inlined `isPrototypeOf`, because that uses
the same CacheIR/MIR instruction, but this is uncommon (shows up on one jit-test
with `--ion`).

Depends on D143022

Differential Revision: https://phabricator.services.mozilla.com/D143023

UltraBlame original commit: 9341ba02a8dc390673d41e59ee406b24a641ddaa