Bug 1763495 part 1: Remove some unnecessary styles in TopLevelVideoDocument.css. r=hsivonen,emilio
As indicated in the new code-comment, we do need the html element to be the
full height of the viewport. But:
- we don't need to set width to 100%. The default, 'auto', has the same effect.
- we don't need to set margin/padding to 0; they default to 0 on 'html'.
- we don't need to set anything on the body, since it doesn't render or have
  any children in the box tree. (The video descendant is abspos and hence uses
  the viewport as its containing block, not the body.)

As an extra check that this is reasonable, note that the similar
TopLevelImageDocument.css stylesheet doesn't have any CSS rules that target the
html or body elements.

Differential Revision: https://phabricator.services.mozilla.com/D143099

UltraBlame original commit: 3311357f291276b00564ce8a2a21564e09a9e3f1