Bug 1761932: Move more of the filtering and scoring related work to SnapshotScorer. r=Standard8
The work to filter and select recommendations is somewhat split between
SnapshotSelector and SnapshotScorer right now which can be a little confusing.
This simplifies things a little and makes it a bit more efficient by saying that
the selector is responsible for getting the snapshots from the sources and then
the scorer is reponsible for all the filtering and scoring. They share the
SelectionContext object which tells both what to select.

Differential Revision: https://phabricator.services.mozilla.com/D142790

UltraBlame original commit: 7cfcf640c3773d21f3ea64acf0224b569c8befed