Bug 1762813 - Add support for resolve ops on picture cache and color targets r=gfx-reviewers,lsalzman
Currently, filter primitives expect an input to be in a single
source surface texture. However, for backdrop-filter the source
is often present in multiple picture cache tile targets. For now,
this patch allows source data from picture cache tiles to be
resolved (copied) in to a single source texture for the filter
graph to operate on.

This does involve copying some extra pixels on the GPU for the
initial implementation. In future, we could add support for
filter shaders to work with multiple tiles as input sources (this
is straightforward for simple 1:1 shaders, but somewhat more
complicated for shaders such as blur filters).

Differential Revision: https://phabricator.services.mozilla.com/D142774

UltraBlame original commit: 49081419801e1614f03b3cd4a92987cbdcfe16e5