Bug 1758678 - Limit ANGLE texture upload stride-alignment workaround to windows. r=gfx-reviewers,aosmond
In bug 1686528 we added a workaround for a bug in ANGLE's D3D backend,
by ensuring that the stride of texture data being uploaded via a PBO
was not aligned. However, this appears to have uncovered a separate
bug, either in ANGLE's Vulkan backend or in the Xclipse GPU's Android
Vulkan driver.

This patch limits the workaround to be windows-only, ensuring that on
Android we avoid the latter bug.

Differential Revision: https://phabricator.services.mozilla.com/D143355

UltraBlame original commit: 3a6888ef56c58b47a2e643687875f7fedcdda798