Bug 1762552 - [devtools] Remove isGrip function. r=Honza.
Since we now have distinct list of Reps for callsites that
are handling native object and not grips, we can remove
the check to isGrip in the other reps.
This will allow us to not have to manage actors created
for getting the grip of an object inside the grip preview
(should be handled in Bug 1762553).

Differential Revision: https://phabricator.services.mozilla.com/D142659

UltraBlame original commit: a9ee5c8cf781f46c0cd448eda99290c9ccc79d7a