Bug 1763667 part 1: Merge our two copies of TopLevelImageDocument.css and TopLevelVideoDocument.css. r=emilio,jaws
For both files, we've been keeping around an extra copy in
toolkit/themes/shared/media/ with 1-2 CSS rules related to presentation.

This was probably to allow XUL-style extensions and themes to override the
presentation, back when we supported that form of extension.  But there's not a
good reason to have this separation anymore, so let's simplify and combine the
files of the same name.

This patch also removes one no-longer-esepcially-useful "N.B." comment (which
refers to a file that's now alongside the styles in question).  This patch also
adds an explanatory comment to explain why we bother to set `color` for images
(which was initially not clear to me).

Differential Revision: https://phabricator.services.mozilla.com/D143395

UltraBlame original commit: 8c4c4292e4128004a5c1bd3ab816661b18e47cae