Bug 1764366 part 3 - Shrink MachineState. r=nbp,iain
The MachineState class became pretty large, especially on ARM64 it was about 1 KB due
to the large number of (float) registers.

This patch fixes that by refactoring MachineState to store a pointer to the register
dumps (for bailouts) or the register sets + spill location (for safepoints).
This lets us recover the register address when we want to read/write it.

Differential Revision: https://phabricator.services.mozilla.com/D143588

UltraBlame original commit: 23ffbe7551339ee360f01b178947a245a495351b