Bug 1762424 - Force crash if unable to create Surface after trying all fallback configurations. r=gfx-reviewers,nical
In bug 1762025 we found ourselves in a situation where we are unable
to ever create an EGLSurface, meaning we were continually attempting
to initialize a compositor, encountering a NEW_SURFACE webrender
error, causing us to tear down the compositor and create a new one,
repeating the cycle indefinitely.

This leaves the user with an unusable browser. We'd be better off
simply crashing. This patch adds a flag to FallbackFromAcceleration(),
which forces a crash if we have already exhausted all of the fallback
options. When calling FallbackFromAcceleration due to a NEW_SURFACE
error we set the flag to true. It may also be worthwhile setting this
flag for more errors in the future.

Differential Revision: https://phabricator.services.mozilla.com/D143486

UltraBlame original commit: f6c7c0ed6ebb4950f7dce006b83411bba6f2a0af