Bug 1680243 - move call-indirect bounds check and null check traps out-of-line. r=jseward
Bounds checking and null checking can sometimes be avoided for
call_indirect, but usually the index and table contents are unknown
and there will be checks that trap if they fail.

Making the bounds check jump out-of-line to the trap speeds up two
call_indirect-heavy microbenchmarks (fib_indirect2 and fib_indirect3
in embenchen) by about 7% on x64-Xeon and ARM64-M1, and by about 12%
on x64-i7; on these systems there is no null check, as it's folded
into the call operation itself.

Making the bounds check /and/ the null check jump out-of-line to the
trap speeds up the same benchmarks about 13% on x86-Xeon.

There are minor wins on raybench, which is a more realistic
vcall-heavy program: 2% on x64-Xeon and 3.5% on x64-i7.

Differential Revision: https://phabricator.services.mozilla.com/D142670

UltraBlame original commit: efdc5519f92c609246146f911c8ced84d862e318