Bug 1764596 - remove aLoadInfosAlreadyExecuted span; r=asuth
aLoadInfosAlreadyExecuted is used to track two pieces of state, but we are doing iteration to do
this. Instead, we can track the same state by two booleans: We know all scripts are ready to execute
if we are at the end of the mLoadInfos list. We know that we have failed if we record if any script
fails on mScriptLoader.

Depends on D143621

Differential Revision: https://phabricator.services.mozilla.com/D144276

UltraBlame original commit: 06b05d708236af8fa186745cde043fe7e68f5b9b