Bug 1765091 - Make `IMContextWrapper` set focus to IME immediately if IME was blurred by disabling the input context r=m_kato
In bug 1753173, we needed to make `IMContextWrapper` wait for a focus
notification from `IMEContentObserver` because it's not ready to query
content when the input context is enabled.

However, this means that, if the input context is enabled without focus change,
`IMContextWrapper` does not have a change to set focus to IME.  Therefore,
if `IMContextWrapper` blurs IME without focus move, the state should be
remembered as IME blurred but it's caused by that the input context is disabled.
Then, `IMContextWrapper` can restore focus immediately if the input context
gets enabled and focus has not been changed.

Differential Revision: https://phabricator.services.mozilla.com/D144804

UltraBlame original commit: 9ee9675516cd9ce4d626c76105c00ed82446f6b5