Bug 1765312 - Wait for error stop before AudioInputSource::Stop() r=padenot
We need to wait for the stream stop called by the forced error
before calling `AudioInputSource::Stop` to ensure the callback order in
the test. Otherwise, the stream stop called by the forced error on the
background thread will race with the stream stop called by the
`AudioInputSource::Stop` in its task thread.

The problem of the intermittently unexpected callbacks can be reproduced
by adding a delay/thread-sleep in the forced-error's background thread
task. When MockCubeb runs into the error state by `ForceError()`, it
will create a background-thread task to stop the cubeb stream. On the
other hand, when `AuidioInputSource::Stop` is called, it will create a
task to stop the cubeb stream first and then destroy the cubeb stream,
on its own task thread. Most of time, when `AudioInputSource::Stop` is
called, the underlying cubeb stream has been stopped by the brackground
task, so the `MockCubebStream` won't fire the stopped-state callback.
`MockCubeb::StopStream` in `MockCubebStream::Stop()` called by
`AudioInputSource::Stop()` will return a `CUBEB_ERROR` in this case
since no stream needs to be stopped, and so no callback need to be
fired.

However, in rare cases, the stream stop on the AudioInputSource's task
thread finishes before the one in the background thread. Hence, the
`MockCubebStream::Stop()` called by `AudioInputSource::Stop()` can stop
the stream successfully and then fire a stopped state callback.

We don't care what thread stop the stream in
`TestAudioInputSource.ErrorCallback` test, as long as the error-state
cabllback can be forwarded correctly and the stream can be destroyed in
the end. To avoid the racy stream stop calls between the background
thread and the task thread, we don't call `AudioInputSource::Stop()`
until stream is stopped in background thread.

Differential Revision: https://phabricator.services.mozilla.com/D144637

UltraBlame original commit: 8d29346322957a0dcd4cae09221b759a35712fbb