Bug 1757604 - Make content-type on JAR channels behave the same as HTTP channels. r=Gijs
That is, treat it as a hint if called before open, and as an override if
called after. Override the hint on open.

This is a less invasive change that is green on try and also fixes the
issue.

Differential Revision: https://phabricator.services.mozilla.com/D145098

UltraBlame original commit: 6192634b7401c3d4f1ee8abfe310bd68dcd5ac1f