Bug 1766730: Refactor emitReturn r=jandem
This is in preparation for the next patch.

The current code optimistically emits `JSOp::Return` and then changes it to `JSOp::SetRval` if necessary. Given that this code is being split up, it seemed better to conservatively emit `JSOp::SetRval` and treat the conversion to `JSOp::Return` as an optimization.

Differential Revision: https://phabricator.services.mozilla.com/D145029

UltraBlame original commit: 6e40e83390a87808354c989af6f760eaf4a6f236