Bug 1766434 - Do not run nsIWindowsShellService::CheckPinCurrentAppToTaskbar on main thread in ShellService.jsm r=bhearsum
`nsIWindowsShellService::CheckPinCurrentAppToTaskbar` does IO which we want to
move off of the main thread. There is now an async version,
`CheckPinCurrentAppToTaskbarAsync`, that runs the IO on a background thread and
returns a promise, which we now use in ShellService.jsm.

Differential Revision: https://phabricator.services.mozilla.com/D145244

UltraBlame original commit: 9dd640f33e1b688d01497be94cfe1d11cf2711bf