Bug 1754004 - Part 4: Stop requiring nsMIMEInputStream to be seekable, r=asuth,necko-reviewers,dragana
This type was originally made to be required to be seekable because a
non-seekable stream would lose the nsMIMEInputStream metadata information when
it was wrapped by the HTTP channel using PartiallySeekableInputStream, leading
to potential bugs.

Changes in a later part will change HTTP channel wrapping to ensure that this
metadata is preserved in all relevant cases, so this requirement can be
relaxed. This is important, as the check was previously being bypassed for
RemoteLazyInputStream in the content process, as nsBufferedInputStream would
incorrectly report that it implemented nsISeekableStream even if the underlying
stream wasn't seekable.

Differential Revision: https://phabricator.services.mozilla.com/D141041

UltraBlame original commit: 44dfd76472aae52ac7d3bd7e0be8f321a6b0828b