Bug 1763606 - Part 2: Port TypedArray CountingSort function to C++. r=tcampbell
Direct translation of the JS counting sort implementation to C++. The stack
allocated buffer has inline storage to store the complete uint8_t range. (This
function will also be used in the next part for uint16_t, but in that case the
complete buffer will be heap allocated.)

The C++ implementation is faster than the JS implementation and therefore we
can use a smaller limit when we fallback to `std::sort` for small arrays.

Depends on D143285

Differential Revision: https://phabricator.services.mozilla.com/D143286

UltraBlame original commit: 5d55dc85cdbe888d019482f1eddc3bf28077bbfa