Bug 1768305. Remove widget.swipe.success-threshold pref. r=hiro
Since bug 1757928 we don't vary the value. If a user were to change this value then the opacity of the swipe to nav ui element would not be in sync with the decision to navigate or not. So remove it so things can't break. The pref widget.swipe.whole-page-pixel-size should be used instead now to control sensitivity.

Differential Revision: https://phabricator.services.mozilla.com/D145828

UltraBlame original commit: 1dd4ddbfe8c033d718a64f3d417c7a8fa159a1c3