Bug 1766956 - part 3: Make `SelectionState` stop using `nsRange` to compare DOM points r=m_kato
The creation and initialization cost of `nsRange` is expensive.  And just
comparing DOM points does not require the checks of ancestors which are
required by `nsRange` to observe DOM mutations.  Therefore, although the
logic is not exactly same, we can skip the additional unnecessary cost.
Note that the Editor module checks without `nsRange`'s strict initialization
in most places so that it must not be required to do the strict check only in
`SelectionState`.

Additionally, this moves some definitions of `SelectionState` methods to
its header for making them inlined since the dependency is now reduced.

Finally, some methods are renamed to align to `Selection` API.

Depends on D145226

Differential Revision: https://phabricator.services.mozilla.com/D145227

UltraBlame original commit: 377f41f6d3395cd4ce16ca72580368d1a8d1ade0