Bug 1743834 - Correctly set the start time when the audio callbacks actually are called, the first time audio audio clock starts to be in use. r=alwu,media-playback-reviewers
The base clock for the new audio stream is precisely the time when the clock
source changes from the system clock to the audio clock, accounting for the time
it takes to actually starts the stream. This allows avoiding discontinuities
when switching clocks.

Differential Revision: https://phabricator.services.mozilla.com/D136237

UltraBlame original commit: 4076e2055960a3e461f31839564bfdb5f1fd402c