Bug 1743834 - After unmuting, re-initialize the AudioSink asynchronously. r=alwu
On Windows (in particular, but also not quick on other OS, depending on the
hardware being used), opening a system-level audio stream (using cubeb)
can be very long. It's a blocking call from the MDSM thread. The MDSM can't
query the clock and update video frames while the stream is being opened,
resulting in the video freezing.

On the other hand, we don't want the clock starting to increase while first
starting the media, it's better to block the MDSM in this case.

This patch changes `AudioSinkWrapper::StartAudioSink` to have two modes:
- A synchronous mode for the first opening (and therefore for seeking, etc.)
- An asynchronous mode to use when the stream is unmuted: the AudioSinkWrapper
will continue to use a clock based on the system clock, until the system-level
audio stream is effectively opened, at which point it will switch to using it
for the clock.

This new mode does the initialization on a background thread, but starts the
system-level audio stream from the MDSM thread (it's fast enough and simplifies
the code a lot).

The promises are created synchronously, because this is what the MDSM expects,
but they are rejected in case of failure.

Finally, an edge case is handled: if the stream is stopped while it's being
opened asynchronously. This is detected, and instead of starting the
system-level audio stream, the stream is shut down.

Differential Revision: https://phabricator.services.mozilla.com/D145000

UltraBlame original commit: f09bb9e00c172dc2be0a7d970f1ef4d33301d7f0