Bug 1686654 - Track CompositorHitTestInfo accross blob groups. r=jrmuizel
Before this patch the hit test info is accumulated in the painting loop of a group, there are two issues with that:
 - We can early out before getting to the painting loop if the group does not contain any visible item (hit test info items don't count as visible) so a nsDisplayCompositorHitTestInfo can be ignored if it is between two active items.
 - Group boundaries should not affect the behavior of hit testing.

With this patch, hit test info is accumulated in the ConstructItem loops, is not reset in EndGroup and is carried over from a group to the next.
This means the hit test flags are extended to the scope of the svg container, I'm not entirely sure that it's correct but I believe it is at least less incorrect than the current behavior.

Differential Revision: https://phabricator.services.mozilla.com/D145358

UltraBlame original commit: 950227a483519fe71798772a0996c02dec2125f1