Bug 1769071 - Remove the clean up code for the search param prefs in the migration code in the BrowserGlue.jsm r=anti-tracking-reviewers,bvandersloot
The clean up code for the search param prefs is not necessary anymore.
This patch remove the code but doesn't touch the UI Version number so
that we won't mess up with the migration steps in the future.

Differential Revision: https://phabricator.services.mozilla.com/D146284

UltraBlame original commit: dd20e01cef60da1b06a9a055cada16dacc78cbb5