Bug 1754607 - Adjust GeckoViewAutoFillChild for ActorDispatcher r=owlish
The test failure in the bug is failing because the eventDispatcher in the actor
was previously only available in the top-level-frame.

In addition to adjusting for ActorDispatcher, this patch also makes every actor
handle its own frame (rather than handling all the in-process children) which
simplifies the code a bit.

We also modify the test to use a third-party iframe so that the OOP iframe case
is covered.

Differential Revision: https://phabricator.services.mozilla.com/D141647

UltraBlame original commit: 2ee50ba5679872bedc381835c722a426a2cbd022