Bug 1755094 - Use separate methods instead of onAutofill. r=owlish
This is a minor API fix wherby we call `onSessionStart`, `onNodeAdd` instead of
having a single `onAutofill` call that passes in the type of call. This aligns
the API with all our other delegates.

Note that because this API is normally implemented inside GeckoView itself,
we dont expect any action on embedders for this change.

Differential Revision: https://phabricator.services.mozilla.com/D143959

UltraBlame original commit: cce65aa3dddde1032a4cbfe339c422b26176ba5c