Bug 1499459: Remove skip-if-win7 reftest annotation for an issue that seems to have gone away. r=emilio
skip-ink-multiline-position.html's annotation was added in bug 1705826, for
what was at-the-time a known and very common category of intermittent
test-failures on Win7 (with downloadable fonts mysteriously failing to render,
in pretty much any test that uses a downloadable font).

Based on the disappearance of intermittent-failure reports on bug 1499459, it
looks like this category of intermittent-failure has gone away entirely
(possibly due to some architectural change like WebRender), so we don't need to
skip that test (or any other test that was disabled for this reason) anymore.

Explaining the other change in this patch, for 1673510-1.html: this test's
annotation didn't directly reference this bug, but its comment ("test font
doesn't load reliably on win7") sounds like it was annotated for this exact
same issue.  However: it does actually still fail on Win7, if I remove the
skip-if annotation; but it appears to fail for a new/different reason. I spun
that off as bug 1770273, which I've referenced in reftest.list for this test.

Differential Revision: https://phabricator.services.mozilla.com/D146832

UltraBlame original commit: ab22dfaba16ecc7682358cf0a859db9f7e816883