Bug 1702759 - Remove load_dump preference r=robwu
This is a breaking change. It is not possible to control the loading of dumps from a preference anymore.

Dumps are only loaded if server is pointed at PROD. The recommended way to prevent dumps loading in tests is to store a high timestamp in the local DB.

Differential Revision: https://phabricator.services.mozilla.com/D146049

UltraBlame original commit: 54cf464f80b5fdd6cb2f97c7cebfceedad2258cc