Bug 1767581 - refactor systemprincipal restrictions, disallow loads without finaluri r=ckerschb,tjr
This is a bit of a refactor.
We'll keep the spagetthi code for existing checks, to be able
to easily iterate and pref-flip if things fail later in the cycle.
This also resolves bug 1638770 and removes the "disallow all"
-pref that proved not be a useful approach anyway.

Differential Revision: https://phabricator.services.mozilla.com/D145411

UltraBlame original commit: 9e1c758297d8d34c759fca9fb2a46e4b2e0c4538