Bug 1709888. Wait one from in browser_test_swipe_gesture.js to make sure all events/input are flushed. r=hiro
The timeout in requestIdleCallback is a deadline, so we wait until the firrst time we are idle and then call the callback, or if we haven't gone idle before the timeout period then just call the callback. So it functions as a maximum, not a minimum. So if we are already idle it could callback right away.

So wait one frame to make sure any other potential events, input etc is flushed before we move on.

I haven't debugged this in detail but this seems to fix the problem and seems reasonable.

Differential Revision: https://phabricator.services.mozilla.com/D147120

UltraBlame original commit: dd7e11f8b419d3731bc2c97d089fd19b31b5f2ea