Bug 1770634 - wasm: Report TypeError when constructing global with v128. r=jseward
The JS-API for global constructor [1] has a step for throwing
a TypeError when type is v128. We previously would just throw
a TypeError if a default value was provided and type was v128.

Two unused error messages are removed, and a comment for clarity
on the error handling of ToValType is added.

[1] https://webassembly.github.io/spec/js-api/index.html#globals

Differential Revision: https://phabricator.services.mozilla.com/D147179

UltraBlame original commit: 6c44428c0eca94e9e9ca557ead58e1428db47137