Bug 1771858 - Handle interrupt in AsyncModuleExecutionFulfilled; r=jonco
In AsyncModuleExecutionFulfilled, it's possible that the call to GatherAsyncParentCompletions
will fail if we are interrupted. The current code assumes that this is only reachable as a
result of an OOM. This changes the assertion to an if statement to handle being interrupted.

Differential Revision: https://phabricator.services.mozilla.com/D147954

UltraBlame original commit: 3280722e69ed0c9f3e1f5082f82d44a33c607577