Bug 1764314 - [remote] Allow MessageHandler commands to be retried upon AbortError r=webdriver-reviewers,whimboo
Depends on D147713

This adds a "retryOnAbort" property on MessageHandler Commands.
When set to true, this will allow the FrameTransport to retry a command up to 10 times in case of an AbortError.
Difference with Marionette, this doesn't attempt to detect BrowsingContexts in bfcache, because retrying commands
does not make sense in bfcache (this is not a temporary state, retrying in a loop will most likely not help).

A browser mochitest is added to cover various retry scenarios.

Differential Revision: https://phabricator.services.mozilla.com/D147715

UltraBlame original commit: ac0842a7d090f1d2b797b0317b71bb5807bbe0fe