Bug 1354248 - Part 3: Supply LoadInfoArgs through the RemoteStreamGetter mechanism. r=necko-reviewers,kershaw
The LoadArgs are wrapped in a Maybe, since that seems to be what the conversion utilities
(LoadInfoToLoadInfoArgs and LoadInfoArgsToLoadInfo) seem to expect. It's really up to the
users of RemoteStreamGetter to determine whether or not they're willing to accept requests
without LoadArgs.

Differential Revision: https://phabricator.services.mozilla.com/D147334

UltraBlame original commit: 3bbdfa6e9b99fd4c189d7257db18b0061f77ba42