Bug 1772228: Use BehavesLikeInitialValueOnBlockAxis to broaden/optimimize a check in IntrinsicForAxis, and answer a TODO comment with an explanation in the code. r=emilio
The TODO comment (removed here) was correct that we could/should be using
BehavesLikeInitialValueOnBlockAxis instead of explicit `auto` / `none` checks
(since any value that behaves like the initial value will have the same
semantics as far as this code is concerned, in terms of not generating a
meaningful transferred constraint).

The explicit `minBSize == 0 check **is still needed** (or at least, it's still
useful), since 0 is trivially uninteresting as a transferred lower-bound, so
it's valid to exclude it from the guarded logic (and 0 is not handled by the
BehavesLikeInitialValueOnBlockAxis() check). In this patch I broaden it to
check IsDefinitelyZero(), though (to include 0%).

Hopefully the new code-comment and lambda make these checks clearer.

Differential Revision: https://phabricator.services.mozilla.com/D148079

UltraBlame original commit: 9bfc03c3ebb216b9f5211f38284a29378bce1142