Bug 1772506 part 1 - Change x64 arguments rectifier code to be a bit more like x86. r=iain
There are some subtle differences between x86 and x64, in particular on x64 we were
adding 1 to `argc` immediately and on x86 we did this later.

This patch makes the x64 code more similar to x86, but because x86 has the pushed
frame pointer already, there are still some different computations. That is addressed
more in the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D148252

UltraBlame original commit: 19f06e78b8889084688997c596dbec56895b1966