Bug 1655554 - Fix debug-mode in Raptor-Browsertime. r=perftest-reviewers,kshampur
This patch updates browsertime to add the following PR: https://github.com/sitespeedio/browsertime/pull/1798

It enables us to use a `--debug` flag in browsertime to pause test execution after each script iteration, and provides the ability for developers to pause it on their own by adding `debug.breakpoint(name)` calls to the test script. The execution can be unpaused by running `window.browsertime.pause = false` in the browser console.

The devtools panel is auto-opened when in debug-mode, and the browser is detached so it won't auto-close when you hit failures in raptor/browsertime.

This can only be used locally so we now raise a parser error when it's provided in CI.

Differential Revision: https://phabricator.services.mozilla.com/D148261

UltraBlame original commit: 5718cf939e0d24b8e7324515f8ff9fc1867aeb7b