Bug 1773156 Part 2: Don't destroy tile cache backdrop surfaces during prepare_for_new_scene. r=gw
It's not necessary to clear the tile cache backdrop surface when a new
scene arrives. Doing so has unintended consequences because clearing the
backdrop surface forces the destruction of the associated native surface.
That native surface will very likely just have to be recreated. If the new
scene doesn't have a backdrop, or has a different backdrop, the logic in
take_context is sufficient to destroy the old one.

Depends on D148704

Differential Revision: https://phabricator.services.mozilla.com/D148705

UltraBlame original commit: dfb41e7524beb3a7288e4016e4ca7af0846f7c6e