Bug 1772282 - (part 1 of 3) Replace js/src/ds/SplayTree.h with an AvlTree.h and change all of SM's uses accordingly.  r=rhunt.
This patch adds js/src/ds/AvlTree.h, which provides facilities similar to
js/src/ds/SplayTree.h, but with reduced lookup costs.  The patch merely adds
the code and test cases but does not make any use of it.

To make it easy to change use points, the public interface is identical in
naming and semantics to that of SplayTree.h, to the extent it can be.  See
comments on AvlTree::contains and AvlTree::maybeLookup.  The memory allocation
strategy is also the same.

There are new facilities for iterating over a tree, either completely or from
an arbitrary start point.

AvlTree.h is structured so as to hide the implementation from users as much as
possible.  Most of the file is an implementation class `AvlTreeImpl`, which
contains all the difficult stuff -- AVL insertion, deletion, rebalancing,
search and iteration.  The end-user interface, plus documentation, is in a
small class `AvlTree` at the bottom of the file.

The patch also contains self-test code, testAvlTree.cpp.  The public interface
class `AvlTree` is (intentionally) too restrictive for proper testing.  So
this file defines its own testing interface class, `AvlTreeTestIF`.  It then
uses that to test that insertion, deletion and iteration work correctly, and
that the tree remains balanced.

Differential Revision: https://phabricator.services.mozilla.com/D148246

UltraBlame original commit: 63cb0d0c4935a482dececb9f6ac5537cff5d0901