Bug 1754830 - Extend the login-logic to pageload tests in mozperftest. r=perftest-reviewers,sparky
This next patch in the series utilizes the same login-logic in Mozperftest and makes it available to the `pageload_test` method so that we can now automate the logging into of accounts during perftest recordings.
Additional logic is also added to account for if the site requires login, if we are running on CI or locally (and if on CI, accounting for the SCM level), and removal of the verbose flags so secrets do not leak.

Differential Revision: https://phabricator.services.mozilla.com/D147775

UltraBlame original commit: af4b59d1745b19e7d2a6a25b3cda21acbbded535