Bug 1772905 - Update DevTools DOM panel tests that fails if InstallTrigger is not listed anymore as a window global. r=nchevobbe
These two DevTools DOM panel tests are expecting the elements being inspected to be at a certain row index
which depends (likely due to alphabetic order) that InstallTrigger is listed as a global in the window
being inspected (it doesn't matter what the value is, the failure is triggered because when InstallTrigger
is completely hidden the elements expected by the tests are shift by one row position).

And so these two tests would be currently permafailing when "extensions.InstallTrigger.enabled" is set to `false`
(but they are completely uneffected when "extensions.InstallTriggerImpl.enabled" is set to `false`, because the
InstallTrigger global is still defined, just set to null instead of providing the API defined by the
InstallTriggerImpl interface).

This patch changes name of the global properties to which they are set to "A" and "B", to make sure those properties
 are going to be in rows listed in the DOM panel before the two non-standard globals `InstallTrigger` and `SpecialPowers`
 and so that the visibility of the InstallTrigger global (either disabled or still visible) doesn't affect the test case.

Differential Revision: https://phabricator.services.mozilla.com/D148975

UltraBlame original commit: 786400ce2e54b7c791e0d21c40b93b1431ab91be