Bug 1768583 - Update preload-error.sub.html.ini based on recent changes on PerformanceResourceTiming r=necko-reviewers,kershaw
We used to not generating a PerformanceResourceTiming entry for
cross-origin redirects. And this makes some of the tests in
preload-error.sub.html to fail because the tests check the number
of entries.

Since we now start to generate these entries, the tests start to pass.

Depends on D146581

Differential Revision: https://phabricator.services.mozilla.com/D146582

UltraBlame original commit: b78f7a65d677b396e1544f43210109df58a8d498