Bug 1770363 - [devtools] Fix iframe dropdown on target destruction from the MBT r=nchevobbe
This only fail on MBT because it still uses non-EFT
and the top level document is displayed in the iframe dropdown
via the old "listFrames/frameUpdate" codepath.
So that its frameData's id isn't related to a target actor
but rather to a frame of the parent-process/window-global target actor...

I'm adding test coverage for the regular web toolbox,
but the fix only reproduce in the context of the browser toolbox.

Depends on D149257

Differential Revision: https://phabricator.services.mozilla.com/D148793

UltraBlame original commit: 0c6d39a1c5d119771af7cb45dfc03f01442bb476