Bug 1773621 - Remove panda.tv from list of sites tested on WebPageTest. r=perftest-reviewers,AlexandruIonescu
A week ago we received a notification that we had a test that the WPT chrome tests were perma failing on [[ https://bugzilla.mozilla.org/show_bug.cgi?id=1773621 | bugzilla ]]
After going through the fail logs I realized it was because of website "panda.tv" directing to a unable to connect page message, after some digging it was not returning proper data because panda.tv has not been a company since March 2019(bankruptcy filing), why this only is causing an issue now I believe is because of some kind of update from WPT as I can see a noticeable UI difference on the test results page from before and after the failures started.

My resolution was to remove Panda.tv from our test list and that seems to have resolved the issue.

I also updated the error message to display which website is causing the issue so that if this happens again I don't need to go through each and every webpagetest result to know which of the 40 websites are having an issue.

Differential Revision: https://phabricator.services.mozilla.com/D149642

UltraBlame original commit: 35e55232d5605fa71001673f0e98492230323bb1