Bug 1698184 - Move submetrics to 2^25 to make room for dynamic metrics r=janerik
Previously we used the 27th bit to signal if a metric instance was actually a
labeled_* metric's "submetric". This bit is actually part of the metric type
part of the lower (upper?) 32 bits of the metric_entry_t, but that was okay
because builtin metrics all have their top 5 bits scrubbed.

itsFreeRealEstate.jpg

This can no longer work, though, since JOG will need to be able to have its
own submetrics that don't conflict _and_ will need to keep the type bits
around so we know what instances to build.

(( Yes, yes, I could've designed the system so that we didn't need to,
but this was the less complicated way. We can go down this route when we
provision the 34 Millionth metric and run out of space. ))

So that means taking two bits away from the id portion: one for runtime (we'll
give it 2^26) and one for submetric (2^25).

Differential Revision: https://phabricator.services.mozilla.com/D143044

UltraBlame original commit: 70a4a8e1c94e331608acd08d10c028c8012b8ee7