Bug 1698184 - Set out a skeleton of the JOG system r=janerik
We're gonna try our best to keep most of the JOG stuff in bindings/jog.

We won't always be able to. For example:
In order for the created metrics instances to be available to e.g. FFI,
we'll have to codegen them into the `fog` crate.

Differential Revision: https://phabricator.services.mozilla.com/D143047

UltraBlame original commit: 1b95ed4b63a99bbfb069a0bd50fc359a8ccbca3a