Bug 1775119 - Make sure NSS is initialized before calling NS_NewCryptoHash from nsCSPHashSrc::allows. r=ckerschb
The intermittent failure tracked by Bug 1775119 is being triggered when nsCSPHashSrc::allows is called while NSS isn't
initialized yet (which is what makes NS_NewCryptoHash to return a NS_ERROR_ILLEGAL_VALUE nsresult).

After recording the test passing successfully with rr and looked into what did actually initialize NSS when the test pass,
it seems that the NSS initialization (at least in the xpcshell test environment) happened just as a side effect of a call
to `Performance::Now()` triggered by a call to `Document::HasRecentlyStartedForegroundLoads()`.
This call is originated by the TaskController and so it may not be always triggered before the call to ns CSPHashSrc::allows
and when that happens the inline script triggers an unexpected CSP violation and it is not executed.

This also explains why the test was never failing in the in-process-webextensions, when the extensions are running on the
parent browser process NSS is likely already initialized by some other Gecko internals.

This patch adds an explicit call to EnsureNSSInitializedChromeOrContent right before the call to NS_NewCryptoHash, which is
making sure NSS is always initialized before trying to create a new crypto hash instance (which then makes the test to
pass consistently, included when executed with --verify).

Depends on D150241

Differential Revision: https://phabricator.services.mozilla.com/D150242

UltraBlame original commit: b14241477254436c8f58894e30d154ce1747596f