Bug 1682621, remove keypress listener for main toolbar buttons, and use the inputSource to distinguish keyboard events, r=Gijs
The virtual click event is also removed. It seems this event is only used to check whether this was a key or mouse event in PanelMultiView.jsm

This also removes support for pressing the enter key to trigger these toolbarbuttons on Mac which shouldn't be happening anyway.

Differential Revision: https://phabricator.services.mozilla.com/D149285

UltraBlame original commit: 41c2f75017bfaed1925264aaeb2ff4bd95eb71eb