Bug 1754932 - LUL initialisation: replace `CallFrameInfo::Rule` and 7 children with a fixed-sized structure.  r=mstange.
At profiler startup, LUL reads and preprocesses Dwarf unwind info for all
shared objects mapped into the process. That includes libxul.so and around 90
other shared objects on an x86_64-linux build. This takes too long.

Profiling just this reading phase shows a huge number of mallocs and frees
resulting from LulDwarf.cpp. The largest cause of these is the implementation
of a sum-of-products type with 7 variants (CallFrameInfo::*Rule) using C++
inheritance. The variants have different sizes and so have to be boxed (stored
on the heap and referred to by pointers). They are very short lived and this
causes a lot of heap turnover.

For an m-c build on x86_64-linux on Fedora 35 using clang-13.0.0 -g -O2, when
measuring a MOZ_PROFILER_STARTUP=1 run that is hacked so as to exit
immediately after the unwind info for libxul.so has been read, this patch has
the following perf effects:

* run time reduced from 0.50user to 0.41user (Core i5 1135G7, circa 4 GHz)

* heap allocation reduced from 8,087,404 allocs/397,177,435 bytes to
  3,176,936 allocs, 328,204,042 bytes.

* instruction count reduced from 6,848,730,307 to 5,572,028,393.

Main changes are:

* `class CallFrameInfo::Rule` has been completely rewritten, so as to merge
  its 7 children into it, and those children have been removed.  The new
  representation uses just 3 machine words to represent all 7 alternatives.
  The various virtual methods of `class CallFrameInfo::Rule` (`Handle`,
  `operator==`, etc) have been rewritten to use a simple switch-based scheme.

* The code that uses `class CallFrameInfo::Rule` has been changed (but not
  majorly rewritten) so as to pass around instances of it by value, rather
  than pointers to it.  This removes the need to allocate them on the heap.
  To simulate the previous use case where a NULL value had a meaning, the
  revised class in fact has an 8th alternative, `INVALID`, and a routine
  `isVALID`.  INVALID rules are now used in place of NULL pointers to rules as
  previously.

* Accessors and constructors for the revised `class CallFrameInfo::Rule` hide
  the underlying 3-word representation, ensure that the correct
  representational invariants are maintained, and make it impossible to access
  fields that are meaningless for the given variant.  So it's not any less
  safe than the original.

* Additionally, Dwarf expression strings are no longer represented using
  `std::string`.  Instead a new two-word type `ImageSlice` has been provided.
  This avoids all unnecessary (and unknown) overheads with `std::string` and
  provides a significant part of the measured speedup.

Differential Revision: https://phabricator.services.mozilla.com/D139288

UltraBlame original commit: 27a3c69f1ef63bf5220c82d6bbaa8bf0ecab289b