Bug 1773641 - Enable browser_navigator_clipboard_readText.js test on Windows; r=mbrodesser
There are two issues that cause the test stuck,

1). The test clicks (0,0) to dismiss paste button, which is also the position of
browser menu, so the dismissing action also opens the browser menu which is overlap
with button in web content, so the subsequent synthesized native mouse action clicks
the menu item, instead of the web content.

2). Sometimes synthesized native mouse event is dispatched to menupopup
incorrectly, instead of web content, even if we have received the popuphide event.
It presumably due to the layout isn't updated yet for menupopup, so the result
of hit testing is unexpected.

Differential Revision: https://phabricator.services.mozilla.com/D151030

UltraBlame original commit: 2db4f03d1afefb54d17a9799a0c6c584e8c52302