Bug 1778182: put private browsing taskbar pinning & window separation behind a distinct pref r=nalexander
`browser.privacySegmentation.enabled` is going to be set when users opt into Felt Privacy. When I originally wrote some of this code I naively used it to pref things on, on the assumption that there would never be a situation where it would be `true` before 106 (when we remove the pref-off for this work).

However, we want to run an experiment with pinning private browsing to the taskbar ahead of 106. To avoid any possibility of accidentally opting users into Felt Privacy without their consent, let's use a distinct pref for turning these features off for now. This pref will disappear in the 106 cycle, as none of this code needs to be behind one at that point - so I don't _think_ this needs to go in `firefox.js`.

Differential Revision: https://phabricator.services.mozilla.com/D151065

UltraBlame original commit: 8e00db0440f3e48161d68daa756ac816178d0fe0