Bug 1778651 - Add gfxCriticalNote for negative ANativeWindow size r=gfx-reviewers,geckoview-reviewers,aosmond,owlish
One possible cause of the failures to create an EGLSurface we are
seeing in bug 1772839 is that the Surface's underlying BufferQueue has
already been abandoned. In such circumstances,
ANativeWindow_getWidth/Height will return a negative value. If this
occurs, ensure we annotate crash reports via gfxCriticalNote. This
will allow us to determine whether the crashes we are seeing are
indeed due to an abandoned BufferQueue, or some other reason.

Differential Revision: https://phabricator.services.mozilla.com/D151360

UltraBlame original commit: 39fecab5c852d2ed3bd4ef452b3c063863766327