Bug 1543337 - Allow multiple pointermove events in touch infrastructure tests, r=webdriver-reviewers,whimboo
At least in gecko, we are generating a pointermove event for each
pointer even when there was no actual change. This corresponds to the
behaviour of touch events, where the touchmove event includes all
active touch points.

Although this looks like it doesn't match the intent of the Pointer
Events spec, for an infrastructure test that's just checking if the
testdriver feature is supported, it seems better to allow the
discrepancy from the ideal behaviour rather than imply that the whole
feature is broken.

Differential Revision: https://phabricator.services.mozilla.com/D151276

UltraBlame original commit: fcb9effe543df4065896151dc8a3a468be6db561