Bug 1779680 - Allow "www" to be omitted when searching with a prefix for adaptive history autofill. r=daisuke
Please see the bug for details. We recently improved the handling of "www" and
prefixes for adaptive history autofill and made it more like origin autofill,
and this revision makes it even more like origin autofill. Origin autofill
computes a `host_fixed` value that is the matching host except it doesn't
contain "www" if the user didn't type it. This revision does something similar
for adaptive history autofill by computing `url_fixed`. This also allows
`_processRow()` to be a little simpler for the adaptive history case and it's
almost exactly the same as the origin case now.

I changed the `:searchString` param so that it's `this._searchString` instead of
`queryContext.searchString`. That's consistent with the two other autofill
queries and it made fixing this bug easier.

I also added some comments that should be helpful.

Differential Revision: https://phabricator.services.mozilla.com/D151897

UltraBlame original commit: 3f2fd9e5c3570fe0c45b597b390338a7211da686