Bug 1679556 - Toggle browser.fixup.alternate.enabled and add telemetry, r=mak
The suggested solution for this bug is to disable browser.fixup.alternate.enabled
unless toggled by the user, and to add telemetry to the preference so that if we
find users rarely use the option, we can safely remove the feature in a future release.

Many tests rely on expecting URI's to be modified, so I modified them so that they
should not expect the prefix or suffix to be modified.

Differential Revision: https://phabricator.services.mozilla.com/D147795

UltraBlame original commit: 7bd4cbb9bdd0fbb67ef79ba9e40495df68cd78e7