Bug 1737829: Change the Navigator::UserAgent function for RFP r=asuth
We change it from taking bool aIsCallerChrome to
Maybe<bool> aShouldResistFingerprinting.  If there is
a value in there, it is used for determining RFP status.
If not, we call ShouldRFP with the provided Document.

This is needed because on a Worker, we wind up in UserAgent
with no Document, we need to pass the RFP decision in
from WorkerPrivate.

Differential Revision: https://phabricator.services.mozilla.com/D144581

UltraBlame original commit: 56f6d2cdc706b1d8b437bff224a432bfb00213ed