Bug 1779180 - [devtools] Replace usages of InspectorUtils.getRuleLine with InspectorUtils.getRelativeRuleLine. r=jdescottes.
For <style> elements, `getRelativeRuleLine` gives us the position of the rule
inside a stylesheet, when `getRuleLine` gives the position of the rule relatively
to the whole HTML source.
The former is more accurate for us when we want to display such line number in
DevTools UI (for example in the StyleEditor media query sidebar).

Test will be added in next patch of this queue.

Differential Revision: https://phabricator.services.mozilla.com/D151596

UltraBlame original commit: fc236aff6e5fa7dad54a60b19b15e8b06a9e5ae8