Bug 1777249 - [devtools] Create CSS variable for alternate element background instead of abusing --theme-selection-background-hover. r=bomsy.
In a few places we were using the variable to have a different background than
the ones we have everywhere (blueish on light theme, dark-grey blue in dark mode).
In order to avoid any confusion, this patch adds a dedicated variable with a name
that should be more explicit.

Differential Revision: https://phabricator.services.mozilla.com/D150574

UltraBlame original commit: 6040e53370dce5879930402475b818d0589bb53a