Bug 1776254 - Install ProfilingCategoryList.h (r=arai)
When installing embedded SpiderMonkey, we need to install the generated
header ProfilingCategoryList.h into /usr/include/mozjs-___/js/ because
that's where the rest of the SpiderMonkey code expects to find it.
It's not automatically installed already, because it's generated.

Thanks to Arai for the suggestion of where to look.

Differential Revision: https://phabricator.services.mozilla.com/D152549

UltraBlame original commit: eb2b21f037a5ec7213ebcb3ea7690c7ced2ce368