Bug 1767792 - [devtools] Ignore errors in Firefox data provider if clear was called mid-flight r=nchevobbe
depends on D152726

https://searchfox.org/mozilla-central/rev/d5edb4a4538657b7d691a41c00e6796a19ade6e7/devtools/client/netmonitor/src/har/test/browser_net_har_copy_all_as_har.js#79-104.

This method will perform 2 reloads immediately. This means that one of the requests will immediately cleared by the next navigation and if there was a pending call to FirefoxDataProvider.requestData, it will throw because the actor was already destroyed on the server.

Differential Revision: https://phabricator.services.mozilla.com/D152727

UltraBlame original commit: cd97145a6a9300aa2fb99b7cdc433042c9113025