Bug 1755700: Serialize IAudioSessionControl destruction/re-creation r=cmartin
Destroying the IAudioSessionControl has concurrency issues with audio playback that require it to be done on the main thread.  We have been trying to create a new AudioSessionControl on a background thread while this happens but some AudioSessionControl internals/dependencies are not thread safe.  In order to avoid concurrency crashes, we destroy the old IAudioSessionControl on the main thread, then dispatch the restart to a background (MTA) thread asynchronously.

Differential Revision: https://phabricator.services.mozilla.com/D152302

UltraBlame original commit: 62e28f4dc1af9249bc7543d77e4af3bce4e431f0