Bug 1781786 - Optimize some of the clip-set building process r=gfx-reviewers,nical
Pre-cache the clip set -> node information as entries are pushed
on to the clip-tree builder stack.

This page is still not fast (it's a worst case for our current
code with box-shadow clips) but it's better than without this
patch (and we have plans to optimize the box-shadow case).

Differential Revision: https://phabricator.services.mozilla.com/D152956

UltraBlame original commit: badab985521a7e656f3334a08d0eb2a0c7ab68e4