Bug 1782584 [wpt PR 35293] - ContainingFragmentationContextRoot() for column spanners, a=testonly
Automatic update from web-platform-tests
ContainingFragmentationContextRoot() for column spanners

In CL:3343283, when an OOF was contained within a column spanner, we
would lay out the OOF in the fragmentation context root above, if there
was one (given that spanners don't participate in the fragmentation
context root of the nearest context).

If an OOF was nested in an abpos multicol inside a column spanner, we
would hit NOTREACHED() in NGOutOfFlowLayoutPart::ReplaceFragment(),
because the fragments of the abspos multicol were not found in the
nearest fragmentation context root ancestor, but rather in the next
above.

To fix this, update ContainingFragmentationContextRoot() to retrieve
the correct fragmentation context root in the case of a column
spanner.

Bug: 1348128
Change-Id: Ica05a41b23cade26c93a18cfd80eaa167ed2863e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3795714
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Commit-Queue: Alison Maher <almahermicrosoft.com>
Cr-Commit-Position: refs/heads/main{#1030634}

--

wpt-commits: 0d6d9dc80bf23f94fab8137920b73228690b6de7
wpt-pr: 35293

UltraBlame original commit: a2fac14f3774aa4a03133332f515038be1fd2ee6