Bug 1769152 - Make IOUtils Shutdown blockers depend on previous phase completion r=nika
Now we create barriers for all shutdown phases (xpcom-will-shutdown included)
and have them register the previous phase's completion as a blocker. This
allows for a simplified implementation of IOUtilsShutdownBlocker, especially in
the next patch where we add support for an additional shutdown phase.

In normal shutdown, the previous phase will already be completed by the time we
hit the blocker. However, in xpcshell tests, we are not going to hit the
profile-before-change etc. shutdown phases, so we still need to ensure that IO
finished, so we can recursively wait on previous phases to complete.

Differential Revision: https://phabricator.services.mozilla.com/D152922

UltraBlame original commit: 486eea40c98d28f60fb943c01fe5028224ccc10c