Bug 1779094 - Improve internal handling of user search settings Part II. r=Standard8
This patch creates a cache of the user search settings file that is read from
disk. We use the cached settings to determine if the settings has been updated
or changed. If there has been a change, we allow the settings to be written
and saved, otherwise if there's no change, there is no need to write and save
the file.

Public setter and getter methods have also been created to modify or retrieve
the metadata of the engines.

Differential Revision: https://phabricator.services.mozilla.com/D151551

UltraBlame original commit: 54f9c347df9a44ea7ff1c2f00edaf571c27943b2