Bug 1783642 - P2: Don't use weak ptr in TLSTransportLayer, r=necko-reviewers,dragana
In the case of a WebSocket connection through a https proxy, TLSTransportLayer can be accessed on main thread.
So, we can't use weak ptr in this case.
This patch uses the same technique as nsSocketTransport to better manage the lifetime of TLSTransportLayer.

Depends on D154093

Differential Revision: https://phabricator.services.mozilla.com/D154094

UltraBlame original commit: caca601f2f5e87dd660434f3db2156e950151adb