Bug 1780792 - use Buffer::GetDevice instead of mParent. r=jimb
Another cosmetic change. I've dabbled with IPDL actors too much to not think about WebGPUParent when reading "mParent". Also the parent-child relationship between Device and Buffer is not very obvious to me (nor is it part of the specification).
So I find that wrapping mParent in a GetDevice method to make the code easier to understand. It also makes it explicit that the parent pointer cannot be null.

Depends on D151630

Differential Revision: https://phabricator.services.mozilla.com/D151631

UltraBlame original commit: f611bfcfc02d3a01fb1193699a2da3a43b02b0a9