Bug 1782770 - wasm: Transfer all type definitions to Metadata. r=yury
Metadata stores a vector of TypeDef that comes from the type section of the module.
We currently only transfer a type if it's a struct type, array type, or else a
function type that cannot fit in an immediate for call_indirect signature checks.

Because we filter out some types, this means we need to renumber the type index
space to match. This is a bit hacky, and is difficult to do consistently.

The original reason for only bringing function types that cannot fit in an
immediate was to reduce the size of metadata as most function types can fit in an
immediate. However, I discovered that we make a copies of function types for
FuncExport and FuncImport metadata. This can create multiple copies of the original
function type, e.g. when multiple exported functions reference the same type
definition.

If we instead transfer all types and store the index of the function type in
FuncExport/FuncImport we can reduce total metadata size, reduce allocations,
and remove type renumbering.

This commit does this by:
  1. Transferring all type definitions to Metadata (WasmGenerator.cpp)
  2. Removing the typeRenumbering vector
  3. Replacing FuncType with typeIndex on FuncImport/FuncExport
  4. Updating all users of FuncImport/FuncExport to get the FuncType using
     the typeIndex and Metadata. This the bulk of the changes.

Differential Revision: https://phabricator.services.mozilla.com/D153498

UltraBlame original commit: a881b6ab47954d0d0521f91932d520f36835b9e2