Bug 1783991 - ensure newly opened tabs are synced. r=Gijs
This patch takes a bit of a nuclear option: it removes all
dependencies on SessionStore, and instead just inspects the
tab itself. This means that we no longer store the full
"url history", but no sync clients actually leverage that.

A nice side-effect of this is that we now can store more tabs
in the payload, so heavy tab users are likely to see even more
of their tabs synced.

Differential Revision: https://phabricator.services.mozilla.com/D154192

UltraBlame original commit: 4703b15279080dfbf5844390dd87b374f9ed4efa