Bug 1775627 - Caching the Email WebApp domain pref. r=dimi
This patch make the pref value of the Email WebApp domain pref to be
cached in the memory. So, we won't get the pref everytime we access the
pref value. The cached pref value will be updated when the pref changes.

Differential Revision: https://phabricator.services.mozilla.com/D154085

UltraBlame original commit: 2fba7ebaf4698db729a9fbf427924e0275a71caa