Bug 1784514 - Avoid initialization and commit deadlocks for OffscreenCanvas. r=emilio
When creating an ImageContainer for OffscreenCanvas, we were holding a
lock for the OffscreenCanvas. The constructor for an ImageContainer may
do a sync call back to the ImageBridgeChild thread in the content
process, unlike its typical async notifications for other operations
once it has been created.

There is no need to hold the OffscreenCanvas lock while we create the
ImageContainer itself, so it is best to not to avoid any potential
deadlocks.

Similarly, we should avoid holding the lock when performing the
readback, if any, for the canvas during the display commit.

Differential Revision: https://phabricator.services.mozilla.com/D154556

UltraBlame original commit: 9cbc684894159a79a1bc437a2e02096540746395