Bug 1783299 - Add support for threadsafe mirrored StaticPref strings. r=KrisWright
Prior to this patch, one would need to manually instantiate a copy of a
string from a preference on the main thread in order to access it in a
threadsafe manner on another thread.

This patch adds support for a `DataMutexString` threadsafe type for
mirror: always type StaticPrefs, and works similarly to the existing
atomic types.

Differential Revision: https://phabricator.services.mozilla.com/D153829

UltraBlame original commit: c71190090952eed8fc1912cf81d2e0703ef39fc3