Bug 1769791, part 2 - Add allowXULXBL for file URIs for reftests even if not http. r=jmaher
This ensures we have the permission to load XUL files from file URIs when we are
not using HTTP, which we seem to need for remote reftests, like Android.

This also removes the setting of dom.allow_XUL_XBL_for_file in the reftest pref
file, which doesn't seem to be enough to make Android remote reftests work for
some reason, so I don't think it is doing anything.

Differential Revision: https://phabricator.services.mozilla.com/D154690

UltraBlame original commit: cfe4296e57103886997e573b5592d3d7ca33185b