Bug 1787411 - Migrate layout debugger localization to Fluent. r=fluent-reviewers,flod,emilio
The localization file is moved to `content/` to make it more clear that it's not intended to be localised.

For the same reason, no Fluent migration script is included as there are no localised strings to migrate.

Differential Revision: https://phabricator.services.mozilla.com/D155702

UltraBlame original commit: a013e4932de1d6a8e59ebdb053fde84831c30a39