Bug 1787085 - Fix undefined behavior from cast from float to char. r=gfx-reviewers,lsalzman
Floating-point -> integral conversions rules are such that if the value
cannot fit into the destination type, the behavior is undefined. So when
casting a value between 0.0 and 255.0 to a char, it is not defined what
happens to values above 127 on platforms where char is signed. So far,
we've been lucky that the compiler made it work, but that's not true
anymore in clang 15, which ends up doing the equivalent of
`std::min((unsigned char)(255.0f * x), (unsigned char)127)`, which is
about as valid as anything else considering undefined behavior.

Differential Revision: https://phabricator.services.mozilla.com/D156256

UltraBlame original commit: a0f2f959a9ae4b515418190313a19a5cd9aecd8d