Bug 1726462 - Fix No Window Auto Restart feature to rely on more dependable observer r=bhearsum
It seems that the "browser-lastwindow-close-granted" observer is not entirely reliable about notifying when the last window was closed. Instead we will rely on "domwindowclosed" and checking if that was the last window, which seems to be more reliable.

Differential Revision: https://phabricator.services.mozilla.com/D151859

UltraBlame original commit: 44e6f14f69e1a118b3bdd96c24419c12a9c6bc04