Bug 1775674 - Synchronize _checkOverflow properly. r=mak
My guess is that this didn't use to fail so much because
_getOverflowInfo was effectively synchronous before, but after my patch
the flush is lazy, and the promiseDocumentFlushed call is actually
async.

With this change I can't reproduce the failure locally anymore, while
it's trivial to repro this failure with --run-until-failure before this
change.

Differential Revision: https://phabricator.services.mozilla.com/D156418

UltraBlame original commit: a09c1001ede09e3ce8a5b2e06d457d39defb03ed