Bug 1788977 - Consolidate the finishOffThread*Stencil shell functions. r=arai
We use the same VM functions now for all the variations, so it not longer makes
sense to have distinct testing functions that must be carefully matched. As a
result we can simplify how off-thread jobs are handled in the shell.

Depends on D156336

Differential Revision: https://phabricator.services.mozilla.com/D156337

UltraBlame original commit: 66067684af0aca4853f69413a39467911d7bd73e