Bug 1790127 - Hide the secondary CTA on welcome screen if the primary action will have no effect. r=Mardak
If Firefox is the default app and pinned to the dock/taskbar, the
primary CTA on the welcome screen will have the same effect as the
secondary CTA. So we can hide one of them. The non-MR page already has
this feature and elects to hide the secondary CTA, so do the same here.

Differential Revision: https://phabricator.services.mozilla.com/D157019

UltraBlame original commit: 42915f752ad4abc68d384ed6ced84fd0e871cfe9