Bug 1791154 - Check for null root content controller in UpdateOverscrollVelocity/Offset. r=hiro
In bug 1785786 we added a GPU process implementation for
RemoteContentController::UpdateOverscrollVelocity and
RemoteContentController::UpdateOverscrollOffset. These work by looking
up the root content controller and using it to forward the message to
the parent process. However, we neglected to add a null check on the
result of the lookup. This patch adds the null check.

Differential Revision: https://phabricator.services.mozilla.com/D157553

UltraBlame original commit: 473798bd11916dc1d69f8f3486377dfaa4e30610