Bug 1790920 - Propagate modifiers properly in PanelMultiView's key event handling. r=Gijs
This is needed to fix browser_ext_browserAction_click_types.js.

Before the first patch in this bug, the order of the built-in command
event and the synthetic event here was slightly different, so it was
papered over. Also, the click event will trigger a command event if
needed, so there's no need to do it explicitly.

Differential Revision: https://phabricator.services.mozilla.com/D157591

UltraBlame original commit: 1bb0cca45dadec6937b09eaf206714452698ac68