Bug 1776077 - Fix delay phase for scroll animations in the compositor thread. r=hiro
Basically, we need to detect a case if there is only scroll animations
which are not sampled because it goes back to before phase. If this
happens, we set ScrollToDelay to |mReason| in the returned value.

Without this patch, the visual result is intermittent because it
is possible that there are new animations arrived at the moment when
scrolling from active phase to before phase.

We don't have a test for after phase because we don't support end delay
for scroll animations now. However, although we set the adjusted fill, we may
still need this in the future so let's keep the check of after phase for now.

Differential Revision: https://phabricator.services.mozilla.com/D155108

UltraBlame original commit: 81bfa19dee8352b27d3fa52253cb7b9a7e7db054