Bug 1549678 - Add pref whether to retry failed proxies r=necko-reviewers,kershaw
This is a regression from:
Bug 1494364 - don't prune proxy if all non-direct proxies are disabled

This pref preserves existing behaviour (and regression) but allows users to opt out of this behaviour which might cause slow browsing when the proxies are not responsive.

We might implement a proper fix for this problem in Bug 1791655.

Differential Revision: https://phabricator.services.mozilla.com/D157825

UltraBlame original commit: 7cdc782094723465d05bd429da2d5eee4822b126