Bug 1791972 - Simplify dialog stack and rdm styles. r=Gijs,devtools-reviewers,jdescottes
.browserContainer is relatively positioned, so we can position the
tab-modal dialogs absolutely inside it instead of making them part of
the browser stack.

While at it, make the rdm toolbar part of the regular browserContainer,
just like the regular devtools toolbox is. That way there's no need to
do ResizeObserver shenanigans to be able to let it grow. Keep it also
absolutely positioned tho, because we need to overlay the whole
container when the device modal is opened. That's somewhat gross.

This should in general be simpler to understand than the current set-up,
and more performant to since it avoids the dialog stack from forming
part of the browser element's flow.

Differential Revision: https://phabricator.services.mozilla.com/D157912

UltraBlame original commit: fcfa744811794944c06c9e3890a4e17edf9d04bb