Bug 1782911 - part 8: Make `HTMLEditor::MoveOneHardLineContentsWithTransaction` clean up line break before right block if its first line is moved there r=m_kato
When the first line of right block element is moved before the block element,
unnecessary line break may be moved to immediately before the right block
element.  In the case, it needs to clean it up instead of trying to find
unnecessary line break at end of the left block which is a container of the
right block element.

Differential Revision: https://phabricator.services.mozilla.com/D157417

UltraBlame original commit: f3ed57cc26349f95938def0c4330a0730c893333