Bug 1791224 - Use `CreateNodeResultBase` as the ok type of `mozilla::Result` r=m_kato
Similar to the previous patch, this changes a lot of lines.  However, I think
that it's not so hard to investigate regression point in this patch because
`CreateNodeResultBase` is not used so many times at handling on edit action.

Differential Revision: https://phabricator.services.mozilla.com/D157575

UltraBlame original commit: 865c5d2929582c30a6808b6c95d9b3b36f637c5d