Bug 1739611 - Cleanups for glxtest, r=lsalzman
We already run the EGL test before the GLX one for a while now. Some
reordering and ignoring the case of multi-GPU systems with outdated
Mesa, combined with the fact that the only non-Mesa driver where we
enable HW-WR is the Nvidia one, which again we only support on driver
versions with EGL support, allows us to do a bunch of cleanups.

 - Stop requiring `EGL_MESA_query_driver` support for EGL on multi-GPU
   systems.
 - Make use of the fact that we always run the EGL test first, stop
   doing it after the GLX one.
 - Lots of cleanups that become possible as the result.

Potential issues to have an eye on:
 - EGL on Nvidia-Prime should now get HW-WR on EGL (including dmabuf
   etc.). This was previously blocked and thus needs testing.
 - Multi-GPU system with old Mesa version between 17.0 and 19.0 may
   loose HW-WR.
 - Mesa users on Xorg using 30bit color depth now run the EGL GL test
   fully (no issues expected here).

Differential Revision: https://phabricator.services.mozilla.com/D157865

UltraBlame original commit: d254e2dd277b574fb7299afd7a44d84159a54188