Bug 1792546: add exe metadata to private_browsing.exe r=nalexander
This is, I think, an improvement to the hover text when a user directly pins `private_browsing.exe` to the Taskbar through Explorer. When pinned like this, we have no control over the text of the shortcut. Windows uses the EXE Description -- so we end up with `private_browsing.lnk`, which causes `private_browsing` as the hover text.

With this patch, it ends up being MOZ_APP_DISPLAYNAME, which in the real world will typically come out as something like "Firefox Nightly" or "Firefox Nightly (2)" (the latter will happen if a "Firefox Nightly" shortcut already exists). I _think_ this is an improvement over the current situation, as it least it's a more human readable string, and different per branding - but I'm open to opinions here.

In an ideal world this would be a localized string, but due to it ultimately coming from `configure.sh`, we do not have an existing method to localize it.

Even if we want to use a different string here we ought to take some form of this patch so that we get exe metadata for `private_browsing.exe`.

Differential Revision: https://phabricator.services.mozilla.com/D158216

UltraBlame original commit: 68247a087e2b65de5206caf5df02ca2eff3dd0eb