Bug 1783242 - Part 1: Serialize large JSStructuredCloneData with SharedMemory, r=ipc-reviewers,handyman
This changes the serialization strategy for JSStructuredCloneData to use shared
memory if the total size exceeds 64k. This doesn't change how it is represented
when it is within source or destination processes, just how it is represented
when in transit.

Differential Revision: https://phabricator.services.mozilla.com/D153804

UltraBlame original commit: 2ce8a67f4d702ca80b3b6b8bf5f0a3c63454c031