Bug 1784218 - Add origin controls information to the unified extensions panel. r=zombie,Itiel,desktop-theme-reviewers
In addition to adding messages for origin controls, this patch slightly
changes the UI to account for long l10n strings. As a result, the name
of the extension is no longer truncated with CSS (ellipsis). This isn't
a big problem as these names cannot exceed 45 characters (max 2 lines of
text). This allows us to better handle long permission messages and
(Windows) scrollbars. UX is OK with that at the moment.

We also retain the "default" height of the message below the name of an
extension (using the `min-height` property) to avoid flickering as per
UX guidelines.

Differential Revision: https://phabricator.services.mozilla.com/D156808

UltraBlame original commit: 9862aaa671f16dbe7e73ad07db6956226ce78cc7