Bug 1755044 - Make every pixel of movement past the touch-start tolerance threshold cause scrolling. r=botond
Previously, the entire delta of the touch-move event that crossed
the threshold would be dropped.

While at it, fix the transition to a TOUCHING state for the
PINCHGESTURE_FINGERLIFTED case (mStartTouch should be updated to
the position of the remaining finger).

Differential Revision: https://phabricator.services.mozilla.com/D156221

UltraBlame original commit: c4bdea458a08b975ffd70faed4a2f6fbe1e563bc