Bug 1791892 - part2 : use mutex to prevent crashes. r=jolin
Using mutex can solve the following issues.

1. Racing between `MFMediaSource::Shutdown()` and other media source's functions running on the mf thread pool.

I originally thought the media engine would call media source's shutdown
on the mf thread pool, but I was wrong. The media engine would call that
method on the thread where we called the media engine's function.

Therefore, when we call `mediaEngine->shutdown()` on the manager thread,
it would also trigger the source's shutdown on the same thread. There
are also some other sources' methods would be executed on the manager
thread by this kind of calls as well.

This situation causes a race, we're shutting down the media stream on
the manager thread, and are still using them on the mf thread pool.

2. Media stream doesn't rely pending media engine's sample request properly

In [1], after a stream starts, it would check if there is any pending
request we need to reply. And that will check this condition [2] to
determine if we should do that.

That is a racing as well. Because at the time the stream checks the
source's state, the source might be still running media source's start
function and haven't changed its status [3] yet. That causes stream not
to reply on pending requests and we also don't have other chances to
reply that, resulting in the playback stalled.

[1] https://searchfox.org/mozilla-central/rev/52da19becaa3805e7f64088e91e9dade7dec43c8/dom/media/platforms/wmf/MFMediaEngineStream.cpp#183-185
[2] https://searchfox.org/mozilla-central/rev/52da19becaa3805e7f64088e91e9dade7dec43c8/dom/media/platforms/wmf/MFMediaEngineStream.cpp#314-317
[3] https://searchfox.org/mozilla-central/rev/52da19becaa3805e7f64088e91e9dade7dec43c8/dom/media/platforms/wmf/MFMediaSource.cpp#194

Depends on D157882

Differential Revision: https://phabricator.services.mozilla.com/D157883

UltraBlame original commit: 677cfd18e9feb5e47d3ed501f935004fedf04b46