Bug 1766497: Mozharness: use `venv` instead of `virtualenv` r=ahal
The 'pip' and 'setuptools' that are bundled with 'venv' are not
consistent across python version and platforms. As such, we must
explicitly install a version of each that works with our codebase.

To achieve this, we simply create the venv using '--without-pip'
then 'pip install' both 'pip' and 'setuptools' using the wheels we have
'vendored' inside 'virtualenv'. There's no way to install a specific
version of 'pip' on venv creation, so it's faster to skip installing the
old version of 'pip' then, and install specific version we want right
afterwards.

We don't want to use the venv '--upgrade` feature to upgrade 'pip',
since that will tie us to pypi, and will fail if pypi is down for
whatever reason.

Depends on D144872

Differential Revision: https://phabricator.services.mozilla.com/D145786

UltraBlame original commit: 394555265c86f5c2b77251ce6aefb337ca10ecf4