Bug 1793769 - Avoid trying to create DrawTargetWebgl if shader compilation failed. r=aosmond
If we fail to compile DrawTargetWebgl's shaders, we bail out to a normal software canvas.
However, it will still try to create a DrawTargetWebgl every time we need to create a canvas.
To avoid this, remember if shader compilation failed in the process, and don't try to create
an accelerated canvas again in that case.

Differential Revision: https://phabricator.services.mozilla.com/D158903

UltraBlame original commit: 9ee98b3301d05ae35ac0d16c41ee54ab09fd2b94