Bug 1794841 - wasm: Remove unneeded asm.js bytecode offset handling from baseline. r=jseward
Calls in original asm.js source code that result in function calls at the wasm level
require preservation of the original JS line number. This metadata is through a
callSiteLineNums vector which needs to be iterated through exactly in sync between
the compiler and bytecode generator. We have this logic in ion, but for some reason
it also exists in baseline, even though baseline is never used (and won't be) for
asm.js code.

This commit removes logic related to line numbers from baseline, simplifying the
logic. Previously, we needed to consume the potential line number at the beginning
of the instruction, then propagate it to wherever the function call would be. Now
we can just read the bytecode offset of the current instruction whenever we need
it.

Differential Revision: https://phabricator.services.mozilla.com/D159201

UltraBlame original commit: b3643eec6798d7929e8d563c95c34be749889bbd