Bug 1796013 - Clean up defunct protocols in sync-messages.ini. r=nika
The PTest protocols are only built for the C++ unit tests, which we
never want to build as part of a regular build.

The PPlugin protocols were removed along with NPAPI.

PLayerTransaction was removed at some point.

For the PDocAccessible messages I removed, I added logging to
checkFixedSyncMessages to make it print out everything in "fixed",
and I manually checked the dozen or so messages to see if there
were any messages with that name in .ipdl files in the tree. The
ones I removed aren't present.

Differential Revision: https://phabricator.services.mozilla.com/D159653

UltraBlame original commit: 2ab3151d89f939da36e8d8b8db75e6901eec0847