Bug 1795972 - Prevent offscreen view container from growing the popup. r=Gijs
This depends on bug 1795892 and restores the sizing behavior that XUL had for
this case (making the overflow: hidden element not contribute to the intrinsic
size of the container).

If this were to be written today, we couldd probably use absolute positioning
or transforms, and maybe this could be simplified a bit. This is the less risky
change tho.

Differential Revision: https://phabricator.services.mozilla.com/D159639

UltraBlame original commit: 63d94513f7ee9016d7816adc2f4b5b807910d048