Bug 1796231 - Fix _collapsed map calculations. r=Gijs
The bug fix is in the line:

    this._collapsed.set(child.id, targetContentWidth);

we used to set this._target.clientWidth, which doesn't account for the
potentially-overflowing children of this._target. Propagate that from
the _checkOverflow function instead.

I'm not sure what's the best way to write a test for this. Any ideas?

Differential Revision: https://phabricator.services.mozilla.com/D159773

UltraBlame original commit: ce0145e9f630d17556b384c8b21c7f4d38c141e8