Bug 1797272 - Fix flex container dynamic overflow updates to avoid triggering extra reflows. r=TYLin
Move the custom overflow code to UnionChildOverflow, so that it's
computed correctly from UpdateOverflow() when a transformed descendant
changes.

This can cause extra reflows in some cases. In the case of the tab bar
it was here:

  https://searchfox.org/mozilla-central/rev/12a18f7e112a4dcf88d8441d439b84144bfbe9a3/layout/generic/nsGfxScrollFrame.cpp#6916

The subsequent reflow ended up fixing up the overflow areas, so I don't
think it's a correctness issue otherwise.

It seems grid has the same issue, but it seems a bit harder to fix
because of subgrid and masonry. Filed filed bug 1797305 for that.

Differential Revision: https://phabricator.services.mozilla.com/D160221

UltraBlame original commit: d48aaf20cd4fa73b8407cd0066e2cdb04833017f