Bug 1727820 - Add logging to all UpdateService state transitions r=bhearsum
The Update Service stores information about the current update state in the status file and in the state properties of both downloadingUpdate and readyUpdate. It's important that we log changes to this information in order to better be able to debug tests and user issues.

Differential Revision: https://phabricator.services.mozilla.com/D159304

UltraBlame original commit: 1e63d512b4a0c4b03a4745b87af1abb5572ba378