Bug 1795510 - Explicitly check for perspective in clip tile shaders for SWGL. r=aosmond
Clip tile shaders don't output the vertex W, so that SWGL can't detect if we're trying
to render something with perspective or not. swgl_drawSpan functions then get invoked
as if no perspective was used. In this case, we need to explicitly check to see if
there is perspective, or rather, that the W value varies along the span.

Differential Revision: https://phabricator.services.mozilla.com/D160464

UltraBlame original commit: d945b8d9c2e4bcc9e5014b1c98af48f3dd75698f