Bug 1794438 - Add ability to test push_computed_frame() to wrench. r=gfx-reviewers,bradwerth
Gecko uses the webrender API DisplayListBuilder.push_computed_frame()
to create a reference frame with a transform specified by a scale_from
size, an optional vertical_flip, and a rotation. This is used for
video.

This patch adds the ability to test this API from wrench yaml. An
example of a test using this capability can be found in the next patch
in this series.

Depends on D160816

Differential Revision: https://phabricator.services.mozilla.com/D160817

UltraBlame original commit: ae1bcfe2f7a1b1979df7c0c255b121fc112925ab