Bug 1797806: don't do private browsing specific things when private browsing is disabled r=mhowell
At the moment, launching with `-private-window` (which we now do more easily with `private_browsing.exe` and the associated shortcut) when private browsing is disabled will open a private-looking window that is not private...which is very misleading.

This includes:
* Shortcut creation (which will end up with private shortcuts that open non-private windows)
* Setting AUMID and window/taskbar icons

The latter is handled with a guard in the two places we do it (nsWindow.cpp & BrowserContentHandler.jsm). Ideally, no caller would ever request a private window when private browsing is disabled...but I don't think that's a realistic possibility.

I also noticed we're doing some unnecessary work in the jump list builder when private browsing is disabled, so I fixed that.

I don't think this is reasonably testable in automation (ultimately, AUMID and icons are attributes of the Windows API objects, and we don't have the scaffolding to retrieve those in tests) -- but I'd be happy to be wrong about this.

Differential Revision: https://phabricator.services.mozilla.com/D160568

UltraBlame original commit: bfc380f167f3be93b4cab989343afeeddc0a0ec2