Bug 1798388 - add Accessibility profiler category; convert existing code. r=canaltinova,eeejay
We use profiler categories to separate out the profiler markers for major
components, such as DOM or Layout, to quickly understand what is happening in
each of them: this will also be helpful for accessibility. One might be
concerned that we're adding too many categories but I think it's fine: the
accessibility category shouldn't appear if accessibility services aren't in use.
Furthermore, if accessibility code is running when it's unexpected, having the
accessibility category appear will only highlight that fact.

Differential Revision: https://phabricator.services.mozilla.com/D160846

UltraBlame original commit: 142678682797013ee9b87a1d2f347ff219258500