Bug 1262276 - part10 : bypass skip to the next frame until media loops back. r=padenot
We don't want to trigger skip-to-next-keyframe after reaching video
EOS. Because now we've reset the demuxer to 0, and are going to request
data from start.

If playback hasn't looped back, the media time would still be too large,
which makes the reader think the playback is way behind and performs
nonecessary skipping which causes stutter.

Eg. Media is 10s long, reaching EOS at 8s, requesting data at 9s.

Assume media's keyframe interval is 3s, which means keyframes will
appear on 0s, 3s, 6s and 9s.

If we uses current media time as a threshold, the reader sees the next
key frame is 3s but the threashold is 9s, which usually happens when
the decoding is too slow.

But that is not the case for us, we should by pass the
skip-to-next-keyframe logic until the media loops back. Otherwise, we
will seek to EOS directly because the requested time is too large.

Differential Revision: https://phabricator.services.mozilla.com/D160160

UltraBlame original commit: abd21243cfda241d4eae6022ebb4d8105772487d