Bug 1797836 - Part 1: browser.js should only call registerToolbarNode on CUI areas of TYPE_TOOLBAR. r=cmkm
It looks like we had some kind of hard-coded assumption here that every area
in CustomizableUI was of TYPE_TOOLBAR except for AREA_FIXED_OVERFLOW_PANEL, which
was manually spliced out of the area array when initting the toolbars after
DOMContentLoaded.

A later patch in this series is adding a new area of TYPE_MENU_PANEL, and
instead of also splicing that out of the area array in DOMContentLoaded,
we now make sure to only call registerToolbarNode on areas that are
actually toolbars.

Differential Revision: https://phabricator.services.mozilla.com/D161077

UltraBlame original commit: 9aaa6d1a3568d317344b0afc3f63cd2ce3e944fb