Bug 1798883 - Reserve capacity in Buffer::reserve; r=anba
This fixes a problem where we were hitting simulated OOMs due to a mismatch
between what had been reserved and the actually capacity of the underlying
Vector. It's safe to use up to the capacity of the Vector, but simulated OOMs
are triggered if an attempt is used more space than what has been reserved.
With this change in place, we can use infallible resize operations in the
written method, avoiding problems with simulated OOMs there as well.

Differential Revision: https://phabricator.services.mozilla.com/D161330

UltraBlame original commit: 17d495325e97240e968e5f4aae25611eb1ab4214