Bug 1788850 [wpt PR 35760] - CacheStorage now keeps the fragment in the Request URL but doesn't save it in the Cache key., a=testonly
Automatic update from web-platform-tests
CacheStorage now keeps the fragment in the Request URL but doesn't save it in the Cache key.

The Cache storage now keeps in the Request Metadata, the complete URL (including the fragment). However, when doing the match, the fragment part is removed. This means that these 3 example URLs will all match:

https://example.com/path
https://example.com/path#ref
https://example.com/path#ref2

The latest url put in the cache will be the one saved and returned in the Request.

The specifications (https://w3c.github.io/ServiceWorker/#request-matches-cached-item-algorithm) states that the cached URL should be the request URL and no where is it specified that the fragment should be removed from the cachedURL:
"Let cachedURL be request’s url"

However, it is specified that during matching, the fragment should be excluded:
"If queryURL does not equal cachedURL with the exclude fragment flag set".

----

This CL modified the following:

- `third_party/blink/renderer/core/fetch/request.h`

The `removeFragment` parameter was added to the `CreateFetchAPIRequest(
      bool removeFragment = true)` and by default, for retro-compatibility reasons, the fragment is removed.

Once we are certain that always keeping the fragment doesn't cause any side-effects, we can change this behaviour and just remove the code that was removing the fragment.

- Methods to support manipulating fragments (ref) and their unit tests were also added.

- The cache now saves the url in the request metadata so modifications to the proto file were made.

Bug: 1205111
Change-Id: Ib23fd5a9364bc64159a2859dc9fd9d3512697202
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3859817
Reviewed-by: Charlie Harrison <csharrisonchromium.org>
Commit-Queue: Etienne Noël <etiennenoelchromium.org>
Reviewed-by: Joshua Bell <jsbellchromium.org>
Reviewed-by: Oriol Brufau <obrufauigalia.com>
Reviewed-by: Maks Orlovich <morlovichchromium.org>
Reviewed-by: Finnur Thorarinsson <finnurchromium.org>
Reviewed-by: Yaron Friedman <yfriedmanchromium.org>
Reviewed-by: Mitsuru Oshima <oshimachromium.org>
Cr-Commit-Position: refs/heads/main{#1068711}

--

wpt-commits: 7f9790cd8ab7c20a975a15b1b07e049588a28126
wpt-pr: 35760

UltraBlame original commit: 8bdea73fcf6c8ad86fe42964f0667975c2569cdc