Bug 1800238 - Make popups position: fixed rather than position: absolute. r=TYLin
So that they don't affect the overflow area of the root element. Since
they paint in a different widget over the current one it doesn't make a
lot of sense for them to do so.

They are in the top layer already, so for XUL documents this doesn't
change anything (because there's no root scroll frame and abspos CB
is the same as fixed CB).

For scrollable html documents it does prevent the scroll frame from
accounting for them.

This inverts the destruction order of the default tooltip's
nsMenuPopupFrame and the canvas frame, which caused an interesting leak,
since the tooltip was not getting properly unbound from the
PostDestroyData. nsMenuPopupFrame shouldn't need to manage the default
tooltip anymore, so removing it simplifies the code and also fixes that
leak (which I initially didn't find because I run opt tests by default).

Differential Revision: https://phabricator.services.mozilla.com/D161931

UltraBlame original commit: adae4eadc28dfea0da203624b1c019f47e2051e6