Bug 1792466: Track browser windows that are in the process of being opened so we don't open multiple windows when trying to open many files/urls quickly. r=Gijs
Ideally we would use the window mediator to just find new browser windows that
are in the process of opening but while we can find the windows they just appear
as about:blank with no way to verify that they are browser windows.

This just takes the straightforward approach of forcing code that opens browser
windows to register them with the BrowserWindowTracker and provides a simple
shared API for opening browser windows that does this.

Differential Revision: https://phabricator.services.mozilla.com/D161076

UltraBlame original commit: 36cc09178d76a576e7f5cadc865683148b524881