Bug 1797785 - [devtools] Add tests for Network Observer r=bomsy
Depends on D162193

This adds only very basic tests for the observer, but I would like to discuss about switching the API to a fully event-based solution.
I initially decided against it because I didn't want to create unnecessary changes for DevTools, but in the end, the owner pattern is quite verbose if
we want to use it for testing. Having events instead would make it more flexible, we could test only what we are interested in, with minimal overhead.

If we agree on this, we can land those very simple tests and I can move on to change the API to use events and update the tests, before we add more
tests.

Differential Revision: https://phabricator.services.mozilla.com/D162195

UltraBlame original commit: a7400b45e2e4e6c9191b21e682f42812c7502904