Bug 1799111 - Don't apply auto min sizing when spanning flexible tracks. r=emilio
This is the first part of the bug.  A clarification in the spec states
that a grid item spanning multiple tracks, which include a flexible
track, should not participate in auto min sizing calculation and should
have a width of 0.

https://w3c.github.io/csswg-drafts/css-grid-1/#min-size-auto

For the reftest, the previous overrides are removed as the correct
behaviour is now implemented.

Differential Revision: https://phabricator.services.mozilla.com/D160900

UltraBlame original commit: ec2ad590b1485ab7d895d3d8ba9434739f8dd603