Bug 1793834 - part 8: add preference for number of messages to keep r=handyman
One thing to note is that if the user shrinks this number we won't reclaim any memory - this is because I wanted to avoid doing any more work than necessary when we're logging messages. Let me know if you think this is a problem!

Differential Revision: https://phabricator.services.mozilla.com/D161156

UltraBlame original commit: 9243bbde3d5e663cf93c09a3726065f7077d2583