Bug 1801337 - Fix uninstall on sitepermission AddonCard created on the AOM onInstalled event. r=rpl.
There are 2 folds to this patch.

First, the `SitePermsAddonInstalling` was calling its parent constructor (`SitePermsAddonWrapper`),
without the expected permissions parameter. This would then lead to the `uninstall`
method to not do anything as there were no permissions.
This is fixed by passing the permission for which we want to install the addon.

Second, if multiple addons were installed for the same origin (e.g. "midi" and "midi-sysex")
while about:addons was open, removing the addon would only revoke the "first" permission,
as `SitePermsAddonInstalling` only holds one permission.
To fix this, we define an `SitePermsAddonInstalling#uninstall` method that checks
if we registered a `SitePermsAddonWrapper` instance for this origin, and in such
case, uninstall the addon from this instance instead of the `SitePermsAddonInstalling`,
one, as the `SitePermsAddonWrapper` instance has the whole set of permissions
granted for a given origin.

Test cases are added to ensure we don't regress this.

Differential Revision: https://phabricator.services.mozilla.com/D162431

UltraBlame original commit: aee8df762c947fdaa328dbac2549790755623c20