Bug 1774416 - Allow dynamically fetching uprofiler functions to profile libraries that can't access the Gecko profiler API. r=canaltinova
This is solving a particular problem for us in media land. libxul depends on
libmoz{avcodec,format,util}, but not the opposite. We can't use the profiler
functions (thread registration, markers, etc.) there.

Unfortunately those libraries contain most of the interesting part for a few
codecs, including their own thread pool, so we really want to register those
threads and add markers to their innards so that we can figure out what's going
on and how to use them in the most efficient way.

Another problem is that those libraries are in C, and the markers API is in
C++.

The trick here is that when those libraries are being called into, the functions
in MicroGeckoProfiler.h (that exposes a C API for the profiler) are available in
the address space (it's never too early), so we can fetch them and start using
them, without having libmoz{avcodec,format,util} link to libxul.

This let us write temporary patches over vendored dependencies to diagnose
locally, and/or carry small and easily rebaseable patches to register important
threads that are otherwise invisible to the profiler.

Differential Revision: https://phabricator.services.mozilla.com/D149543

UltraBlame original commit: ba520cced1f1ddd5236da21b9e058decd9395065