Bug 1801768 - Make `mach bootstrap` bootstrap everything that `mach configure --enable-bootstrap` would bootstrap. r=firefox-build-system-reviewers,ahochheiden
--enable-bootstrap doesn't cover everything that `mach bootstrap` deals
with, but the opposite is true, and it's redundant to have to maintain
a list of things to bootstrap when configure already knows what it
wants.

As a side effect, this now makes `mach bootstrap` install wasi-sysroot,
which it wasn't doing before.

Differential Revision: https://phabricator.services.mozilla.com/D162683

UltraBlame original commit: 7ca28fbab11934bd0beb9a4caf904a5a02776de8