Bug 1799028: Move enumerators from Realm to Compartment r=jandem
This makes it possible for us to bake the address of the active iterator list into stub data. If it's per-realm, then we can't assume that all objects reaching a given stub should be added to the same list. With a per-compartment list, we can be sure every object reaching a given stub shares the compartment (even if it's a CCW representing an object from another compartment).

While moving this, I also changed it from being a pointer to a NativeIterator to embedding a NativeIterator directly. This saves an indirection, makes some previous fallible code infallible, and makes everything a little simpler.

Differential Revision: https://phabricator.services.mozilla.com/D162052

UltraBlame original commit: b0dc5f1fd855a998437bb222fd4f090cc1b5190f