Bug 1778510: Improve the shared test infrastructure r=timhuang
This commit moves the test description from the test file
to the shared infrastructure.

It also adds two parameters to the set of shared tests:
extraData and extraPrefs.  extraData is data passed through
to the validation function (mostly) unchanged and unused.
extraPrefs are prefs that applied (and removed) before executing
the test.

Finally, increase the timeout because this is a big test and
it occassionally runs slight over 45 seconds.

Differential Revision: https://phabricator.services.mozilla.com/D156510

UltraBlame original commit: f1472a04734ceaf2df978520a3bb6b65a233194c