Bug 1797073 - Part 3: Collect rule lookup telemetry when getting cookie banner click rule. r=pbz
We collect the rule lookup telemetry when getting cookie banner click
rule because the cookie banner clicking will be called for both
top-level pages and iframes.

This patch adds a parameter to
nsICookieBannerService.getClickRulesForDomain to indicate if the caller
is in top-level context so that we can know and report the data to
corresponding telemetry labels. It also instructs GetRuleForDomain() to
report telemetry if it is called from GetClickRulesForDomain().

Differential Revision: https://phabricator.services.mozilla.com/D162630

UltraBlame original commit: 6b23116830c686746573b4aac3e9d37b2f1d39fe