Bug 1782091 - Create IdentityCredential state map, r=timhuang,pbz
This creates a new service that stores precisely two bits per RP, IDP, and account ID tuple:
has it been used before and is it able to be logged out.

It does so with the additional constraints of respecting private browsing's clear on last session and in-memory storage constraints and being a participant in the Clear Storage Service.

Tests are here. Coverage isn't perfect, but they cover most of the service.

Differential Revision: https://phabricator.services.mozilla.com/D162124

UltraBlame original commit: 1475247d538f7febd720fa76b1032bdd17f8c524