Bug 1803643 - Speed up selector matching with already-lowercase local name selectors. r=dholbert
This makes relatively simple changes so that we check lowercase-ness of
local-name selectors first. If so, we don't need to check whether we're
an HTML element in an HTML document, which requires a fair bit of
pointer-chasing.

Differential Revision: https://phabricator.services.mozilla.com/D163627

UltraBlame original commit: 7a8e8890780f410f86e9b68e4895e0d5f2371d66