Bug 1801307 - Remember last interval to speed up insertion search. r=jrmuizel
This modifies AddInterval in wpf-gpu-raster to try to remember the last interval
it searched for so that it doesn't have to search the entire interval list for
every interval added. This should take us down from O(n^2) to closer to O(n) for
this phase of rasterization.

Differential Revision: https://phabricator.services.mozilla.com/D163795

UltraBlame original commit: 89666af4ddbca2c4e708ff80a6b869d420d45846