Bug 1803538 - [devtools] Add scrollend in debugger event breakpoint panel. r=bomsy.
The event is only supported when `apz.scrollend-event.content.enabled` is set to
true, so we only show it in the debugger when it's supported.

A test is added to ensure the event breakpoint does work as expected.

Differential Revision: https://phabricator.services.mozilla.com/D163910

UltraBlame original commit: 2e7a46447741c83424c68a96d5c9de6deed6246f