Bug 1803508 - [devtools] Fix console stubs generation. r=jdescottes
All calls to `writeStubsToFile` were updated, but not the function signature itself,
which we do in this patch.
We also take this as an opportunity to fix the require calls in the stub file template.

Differential Revision: https://phabricator.services.mozilla.com/D163489

UltraBlame original commit: e6503ac97b27ca5313dc7b16a43e3f4a42e8e5f4