Bug 1803361 - Add pre-run category checks to perf-selector. r=perftest-reviewers,kshampur
There was a bug found regarding how the Pageload (live) category was empty because the negation of live sites !live was inadvertently included. This is expected given the code we have. Instead of changing the behaviour of mach try perf when it comes to handling negation removals in the variants, we'll remove the live-site pageload category and add a restriction to the category definitions. At the same time, the Bytecode-Cached category is removed since it's a variant of some existing tests as well. Note that these removals are done in a later patch with a few other changes.

The tests are also modified to add new tests for this restriction, and change up the tests to ensure they aren't using bad categories.

Differential Revision: https://phabricator.services.mozilla.com/D163836

UltraBlame original commit: f9637dac96c94ab65aa809794ecb7a878b9d7336