Bug 1800483 [wpt PR 36958] - Change LCP startTime to DomHRTime., a=testonly
Automatic update from web-platform-tests
Change LCP startTime to DomHRTime.

Change time conversion for LCP.

This is to make the conversion of LCP start time from TimeTicks to
double in the same way as that of FCP start time does so that the
precision of the two is the same. Currently they are not. The time conversion for LCP happens in WindowPerformance::OnLargestContentfulPaintUpdated(). It converts MonotonicTime to TimeDelta in milliseconds while FCP time is from MonotonicTime to DOMHighResTimeStamp.

This is found during investigation for crbug.com/1345948 where we want to align LCP start time with that of FCP.

See screenshots in the associated bug.

Bug: 1382187
Change-Id: Ibcf88dd61965669e7bf76b905fa2279f80119429
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4011178
Commit-Queue: Hao Liu <haoliukchromium.org>
Reviewed-by: Ian Clelland <iclellandchromium.org>
Reviewed-by: Camille Lamy <clamychromium.org>
Cr-Commit-Position: refs/heads/main{#1077489}

--

wpt-commits: f23c7ed1f813f38be0e5f2236d3bf313b04631b4
wpt-pr: 36958

UltraBlame original commit: 40cb12940389537ccf3fc5257fef9d58b2a87ff2