Bug 1768581 - Part 5: Avoid late initialization of permanent cookie storage. r=necko-reviewers,valentin
The `CookieService` singleton observes "profile-before-change"
(`ShutdownPhase::AppShutdown`) to destroy the cookie storages, but uses
`ClearOnShutdown` in order to release itself.

If we tried to use it the first time after `ShutdownPhase::AppShutdown`,
`InitCookieServices` would instantiate a new DB connection that is never
removed until MozStorage asserts.

As a minimal invasive hack we check the shutdown phase in
`InitCookieStorage` and in case deviate to a `CookiePrivateStorage`.

As a follow up we should probably verify, up to which shutdown phase we
really want to be able to provide the `CookieService` as a whole.

Depends on D162497

Differential Revision: https://phabricator.services.mozilla.com/D163050

UltraBlame original commit: b9c4d8c3dab68adcabada84825119c1e3018bd45