Bug 177175 - cancel subsequent PKCS#11 password prompts if more than one happens concurrently r=jschanck
Unfortunately, since NSS can prompt for PKCS#11 authentication on the main
thread, the prompt in question results in a nested event loop, which means that
the prompt code can be re-entered and another prompt can appear before the
first is dealt with. As long as NSS can run on the main thread, this will
continue to be the case. Recently we've done a lot of work to prevent NSS
running on the main thread, but that work is by no means complete. In the
meantime, we can paper over this situation by cancelling any prompts that come
after one that is still in progress. This will cause some NSS operations to
fail in a way that should be recoverable by simply retrying them (e.g. the user
can refresh the page or close and re-open it).

Differential Revision: https://phabricator.services.mozilla.com/D164280

UltraBlame original commit: 1d04423d3e5d0bf77452d61d90a6dda50a0048c9