Bug 1401592: Add a config option to imitate the old setParameters behavior. r=jib,chutten,webidl,smaug
When the compat mode is used, we log warnings to the JS console instead of
throwing. We also add a bunch of Glean metrics for tracking how often we
see these warnings (or failures) in the field.

Differential Revision: https://phabricator.services.mozilla.com/D161523

UltraBlame original commit: 92f418bbba4a4b7b7b6438b3af47c56837eec551