Bug 1803245 - Add Timeout nsiTimer onto the Document to track active IdentityCredential requests, r=timhuang
This doesn't actually move the timer onto the document.
Instead it is the correct manual juggling of pointers via `.forget()` and `NS_RELEASE()` and some additional testing to make sure it works properly.
This passes tests where we resolve, reject immediately, and reject on timeout and has no leaks in all of those cases.
Putting the timer on the document also required putting the pending promise onto the document with it and that had further wrinkles.
I call that good enough.

Differential Revision: https://phabricator.services.mozilla.com/D164260

UltraBlame original commit: 1b54637116ca1f6f08a9b675607fbca29ce46dcd