Bug 1805262 - Use sentence casing for the All Tabs panel. r=bigiri,flod,fluent-reviewers
As this is a capitalization change, it does not require a Fluent migration nor a
key bump for retranslation.

This patch also adds a test, and adjusts the existing test to be more flexible with
which panels it can inspect.

Differential Revision: https://phabricator.services.mozilla.com/D164480

UltraBlame original commit: 41a17095c7dbc50abd3bf3f4bc5edc8563d357e3