Bug 1789574 - part 1: Make `HTMLEditor` preserve order of inline style elements too r=m_kato
Currently, `HTMLEditor` preserves styled elements with the reversed order in
the loop.
https://searchfox.org/mozilla-central/rev/d7d5c89ee7bc70dec0fd846689ca030f94931393/editor/libeditor/HTMLEditSubActionHandler.cpp#8638-8658

I.e., if `<b><i></i></b>` is restored in the HTML styling mode, it will be
restored as `<i><b>...</b></i>`.  This blocks writing tests for bug 1789574.
Therefore, this should be fixed first.

(As I commented in the method, ideally we should store all inline elements and
restore all of them, but currently we don't need to do it.  It requires redesign
of the style cache, but it seems that we don't know web apps which is broken by
current behavior.  Therefore, I think that we should do it when we meet a
trouble with the behavior.)

Depends on D164519

Differential Revision: https://phabricator.services.mozilla.com/D164520

UltraBlame original commit: 2a9534dbca77d68f9783539a1c7e22f3e6645a98