Bug 1803062 - Don't call FileSystemSyncAccessHandle::Close when the object is already closed; r=dom-storage-reviewers,jesup
The Close method will return a MozPromise. Calling such method if the object is
already closed would uselessly allocate a MozPromise and dispatch a runnable
for the MozPromise resolving. So it's better to call Close only if the object
is not yet closed.

Differential Revision: https://phabricator.services.mozilla.com/D163564

UltraBlame original commit: 0ba70d490377d5d541b757695be730d2bd04251c