Bug 1805904 - Unconditionally enable container properties in UA sheets. r=Oriol
If the UA sheet gets parsed with the pref off, and then the pref gets
flipped on, we don't re-parse the UA sheet (nor we want to).

Enable the property in UA sheets unconditionally so that the test passes
even when the browser was started-up with the container queries pref
off.

Differential Revision: https://phabricator.services.mozilla.com/D164823

UltraBlame original commit: 3d99b5c5db9a93473cd723d41573507d4454614c