Bug 1770447 - Create moz-support-link component. r=hjones,eemeli
This adds the <moz-support-link> custom element as well as stories and
tests for this reusable component. The purpose of this component is to
replace individual implementations of the "Learn more" links present
in Firefox, specifically in about:preferences and about:addons.

See Bug 1801927 for an instance of using this component to refactor
the "Learn more" links in about:preferences#general.

Differential Revision: https://phabricator.services.mozilla.com/D164131

UltraBlame original commit: 29c832430820d7866b969de2c82d193157ea327d