Bug 1596845 - Pass full trr domain into content process r=acreskey,necko-reviewers,kershaw
Previosuly we'd only pass the `TRRService::ProviderKey()` into the content
process. But now we need the full domain for the error page in the content
process, so we now pass the full domain. The ChildDNSService now holds on
to the full domain, but calls into `TRRService` to update the key for
telemetry and returns that when necessary.

Depends on D164348

Differential Revision: https://phabricator.services.mozilla.com/D164642

UltraBlame original commit: 8e1fe024b6802e58de5039f0877f69391c0ae7ec