Bug 1744362 - Part 4: Move mState from Kernel32ExportsSolver to Layout r=handyman
This is a refactoring that makes SharedSection::Reset() do the resolving of Kernel32ExportsSolver. This will allow us in a future patch to let the SharedSection attempt to resolve Kernel32ExportsSolver multiple times, as it will fail if kernel32.dll isn't loaded in the process yet, but we still want to initialize the dynamic blocklist in that case.

Differential Revision: https://phabricator.services.mozilla.com/D164486

UltraBlame original commit: c88902b60d1fe4d52d81f6e7c797a856e3a35d43