Bug 1807235 - Fix test to pass in beta with different default bookmarks. r=daisuke
The test is passing in Nightly but failing in Beta because Beta has multiple
default bookmarks pointing to the same domain. The test always fails when run
alone, but used to pass when run with others because a previous test caused
bookmarked pages frecencies to be set to -1 (recalculate). With the new frecency
handling the frecencies remain valid until recalculated, and example.com is not
over the minimum threshold for being autofilled by the test.
By adding a temporary bookmark to it, we solve the threshold problem.

Differential Revision: https://phabricator.services.mozilla.com/D165544

UltraBlame original commit: ec9efc04322c8fc5b8618d3c6c9e730977a48f41