Bug 1532644 - Make credentials to use "cookie" rather than "{}" as the cookie r=farre,smaug
dynamic-imports-credentials.sub.html,
dynamic-imports-credentials-setTimeout.sub.html and credentials.sub.html
intends to test the cookie behavior based on various conditions and
ORB shouldn't be a factor here. Currently, it uses "{}" as the
cookie and let's update it to something that is not JSON to keep
the intention of the tests.

Differential Revision: https://phabricator.services.mozilla.com/D165413

UltraBlame original commit: 49c1638654d6d172b0cac3ee977a0d37d84b1592