Bug 1807482 - Re-introduce a reduced version of mouse grabs. r=stransky
Some X servers move focus along with the mouse. They do not do that if
the pointer is grabbed tho.

Since we want to rollup popups on focus-out events, and not grabbing the
pointer causes the main window to lose focus, let's try to grab the
pointer again.

This reintroduces similar code to the one removed in bug 1798131, but
with a couple critical differences:

 * It's simpler.

 * We use less-deprecated APIs, though we can't use GdkSeat because we
   still support Gtk < 3.20 in theory.

 * We don't capture enter/leave-notify events in the grab mask. This
   doesn't seem to be needed, and prevents the spurious
   enter/leave-notify events that caused me to remove the pointer grabs
   to begin with.

For now I reintroduced the grab on X11 whole-sale.

If we see back the intermittent failures from bug 1607713 we might want
to detect the environment in a more fine-grained fashion as suggested on
the code comment.

I tweaked test_panel_mouse_coords.xhtml so that the pointer ends up
inside the panel input region, otherwise the motion event goes to the
actual window and the test fails. It's kind of a bug that we dispatch
mousemove events for that region anyways, but...

Also, add a pref for this behavior, and some related logging, since it
may make it easier to debug stuff in the future.

Differential Revision: https://phabricator.services.mozilla.com/D165735

UltraBlame original commit: e33eb418926a9d85dfe9458428c9b4dcd0ff4ff2