Bug 1808389 [wpt PR 37746] - Fragments may become monolithic even if the node isn't., a=testonly
Automatic update from web-platform-tests
Fragments may become monolithic even if the node isn't.

So ask the fragment instead of the LayoutBox. This is an old problem,
and it prevented pre-paint from understanding that a monolithic
fragment would be the containing block for OOF descendants (normally in
block fragmentation, OOFs are direct children of a fragmentainer, but
not if they are contained by a monolithic fragment, like in this case).
However, with CL:4110949 the problem became more severe, since we may
now force off fragmentation at the last OOF fragment (if it's inside a
clipped container), making it monolithic.

out-of-flow-in-multicolumn-112.html is a test for the regression, while
out-of-flow-in-multicolumn-113.html is a test that would fail even
without the aforementioned CL.

Bug: 1404718
Change-Id: I68ebd5218c2bd0e3e895eaf757a90cbc8594a028
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4131528
Reviewed-by: Alison Maher <almahermicrosoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/main{#1088506}

--

wpt-commits: bd8b3c5448c15c15b7eb67cade1a1913ba07b647
wpt-pr: 37746

UltraBlame original commit: 13a274bbd305e97ce9acd640ee76f7efffeca8d9