Bug 1806778 - [devtools] Avoid loading devtools frontend in a distinct loader when opening it from the server codebase. r=devtools-reviewers,jdescottes
The whole frontend opened by this addons actor was loaded in the "devtools loader".
That's because the addons actor was itself loaded in that distinct loader.
The server used by web-ext, trigerred thanks to --start-debugging-server
is loaded in a distinct devtools loader.

So we should be vigilant anytime we load a frontend module from server codebase
to ensure loading them in the shared JSM loader.

Because of that, we ended up having two ObjectFront modules.
One loaded by the web console UI and its browser-loader, using the JSM global.
And another one loaded by the DevToolsClient when building fronts, using the server distinct devtools global.

Differential Revision: https://phabricator.services.mozilla.com/D166100

UltraBlame original commit: 890a59e9766410f9e501a157effaa273015c56a0