Bug 1808982 - Update profiler-get-symbols wasm blob to commit 0373708893e45e8299e58ca692764be448e3457d. r=canaltinova
As usual, the changes to profiler_get_symbols.js were auto-generated by wasm-bindgen
and do not need to be reviewed carefully.

This update comes with the following fixes and improvements:

 - For ELF files without a build ID, the fallback debug ID is now computed correctly. (fixes bug 1805915)
 - If no symbol information is available, and we fall back to a stripped binary,
   the fallback symbols are less misleading. We now show fun_1b3d4f symbols for
   functions with missing symbols, instead of taking the name of the closest
   "exported" symbol. This fixes another long-standing annoyance.
 - Symbols now have a "function_size" attribute, which is needed for the
   upcoming assembly view in order to know how many bytes to disassemble.
 - A new API path has been added: /asm/v1. This is a first implementation
   of an assembly endpoint, to be used in the upcoming assembly view in the
   profiler.

We need to implement one new method in the helper: getCandidatePathsForBinary.
This method is used to find the file which we can get the assembly code bytes from.
However, we end up listing almost the same paths in it as in getCandidatePathsForDebugFile;
the only difference is that we don't list the path to the Windows PDB file,
by not adding an entry for the lib's debugPath.

Differential Revision: https://phabricator.services.mozilla.com/D166227

UltraBlame original commit: b89625b09839b86030302f630c9247535f21a8a4