Bug 1809749 - Use the correct thread for timed out idle runnables from off-main-thread, r=smaug,mccr8
This code was previously incorrectly targeting the timer at the dispatching
thread. This is fortunately not an issue, as idle runnables with timeouts are
always dispatched from the main thread, however it could be an issue for future
idle runnables with timeouts dispatched from off-main-thread.

Differential Revision: https://phabricator.services.mozilla.com/D166610

UltraBlame original commit: 991e6538f2ff896db341fefcaeda9692427e5d29