Bug 1810231 - Remove builtin colorway theme from builtin location when the migrated colorway theme is uninstalled. r=willdurand
This patch also fixes the TV job failure caught by running the new mochitest-browser test case from D166918,
where the test was failing on the second run because installing the colorway theme builtin in the same
profile where is was previously installed by the same test and then left into the builtin location as
disabled and not visible is basically a no-op.

By removing the instance installed in the builtin location we are still preventing the colorway builtin
to be made enabled again when the migrated theme is installed, but the TV job will be able to pass as
expected.

Depends on D166917

Differential Revision: https://phabricator.services.mozilla.com/D167073

UltraBlame original commit: e973db0b1effb85a06c4d98e19c04e5e147880f3