Bug 1799025: Part 6: Get iterator before attaching IC r=jandem
Prior to this patch, we would not attach an IC the first time we saw an object in a for-in loop, because we hadn't created a cached iterator yet. This makes writing testcases a bit awkward, and doesn't seem good for performance.

Depends on D165210

Differential Revision: https://phabricator.services.mozilla.com/D165211

UltraBlame original commit: 3160865a976ac934e04aee33832bac39195b1715