Bug 1810666 - Prevent processes from being re-used after IsSignaledImpendingShutdown r=smaug
Adding the assertion in AssertAlive revealed that we can end up re-using
processes that have been already notified to be shut down. Rather than
adding a flow to revoke the ImpendingShutdown we want to add checks that
avoid this based on mIsSignaledImpendingShutdown.

Note that the revealed cases seemed to happen only after the parent
process had been asked to shutdown. There seems to be a time window
between AppShutdownConfirmed and AppShutdown where we already notified
existing processes that they will shutdown but continue to ask for new
processes from events still in flight (and explicitly enforced to be
processed before we advance the phase). Anticipating the existing checks
to AppShutdownConfirmed however seemed to create some fallout that needs
investigation. We might want to verify if we can add some shutdown check
at a higher level of the stack, like DocShell creation or
ChangeRemoteness execution.

There might be also an opportunity to streamline this further, replacing
the mIsSignaledImpendingShutdown flag by a new LifecycleState (like
MUST_DIE) that captures and freezes our decision to remove this process
and can be subsequently checked.

Differential Revision: https://phabricator.services.mozilla.com/D166992

UltraBlame original commit: 1e06f09456f706c2096e1675fa817b3d6511e54d