Bug 1808613 - wasm: Propagate null inputs given to ref.cast. r=jseward
Dart relies on ref.cast propagating a null input without trapping. I'm
guessing that this was V8's behavior, this commit matches it.

In the future, we'll be changing ref.cast to the latest milestone which
allows specifying this behavior in the instruction. For now, this is
a temporary workaround.

Differential Revision: https://phabricator.services.mozilla.com/D166001

UltraBlame original commit: 7cf9ce2a1f21f86319282418d745ec40b0439ca5