Bug 1804499 - [2/4] Factor out the stall logic in preparation for export  r=glandium
In bug 1794059 it was noted that the IPC shared-memory allocation code
would like to be able to stall-and-retry as well using the same logic.

While it doesn't use VirtualAlloc, the principle is otherwise the same.
Shuffle the relevant code around so that the stall-and-retry logic is
separate from the allocation, in preparation for exporting it.

Differential Revision: https://phabricator.services.mozilla.com/D164106

UltraBlame original commit: 111bc058d5161e1ab1acb0484f34b45a9ba97042