Bug 1811066 - Skip correct test when isolated process is enabled. r=geckoview-reviewers,m_kato
In bug 1706656 I marked PermissionDelegateTest#geolocation() to be
skipped when isolated process is enabled, due to noticing that it was
failing. However, this test is still failing intermittently on
try. The reason is that PermissionDelegateTest#autoplayReject() is
actually the test that causes the crash, and the harness incorrectly
reports this as geolocation() failing.

This patch marks the correct test as being skipped.

Differential Revision: https://phabricator.services.mozilla.com/D167255

UltraBlame original commit: 5b9853b49b0c7ed4880f5cce383224b8e95cc34f