Bug 1805502: Fix invalid test-case in simulcast-answer.html r=jib
This test case was erroneously prepending '~' to a rid in an a=rid attribute,
in addition to the rid in the a=simulcast attribute. '~' has no semantic
meaning in an a=rid attribute, and is in fact syntactically invalid.

Differential Revision: https://phabricator.services.mozilla.com/D165098

UltraBlame original commit: fe8a9f3f90971786a4bb87819d75e43042d53951