Bug 1810655 - Make `HTMLEditor` check readonly flag when its public methods are called r=m_kato
The readonly flag of `HTMLEditor` can be set by chrome script or
`Document.execCommand("contentReadOnly")`.  However, the XPCOM APIs should keep
working to update the editable content.  E.g., if focused editor is a single
line text field, the app may want to change the value even if it's not editable
by the users.

Note that this patch does not fix all cases of all APIs because this is not
important for Firefox even though the status can be created by web apps (but
Firefox does not use XPCOM APIs basically, instead using XUL commands and it
should not work with readonly state since it may be kicked by user
interactions, e.g., executing a menu item or a shortcut key).
Therefore, it's enough to work it in current Thunderbird for now.

Differential Revision: https://phabricator.services.mozilla.com/D167353

UltraBlame original commit: 9d7bac312272bcd326359ead564b4cb913abf33b