Bug 1811806 - [devtools] Fix Browser Console issues when closing it during its initialization. r=devtools-reviewers,jdescottes
Closing during its initialization (easy to reproduce with the multiprocess mode)
ends up disallowing any further reopening of the browser console.
You would have to restart firefox to make it work again.

This was related to _browserConsoleInitializing never been cleared.

There was another one highlighted by the test related to waitAsyncDispatches.
As the wrapper module is loaded via the browser loader,
it uses the document's setTimeout method which is disabled when the console closes
and its document is destroyed. This ended up creating a never resolving promise...

Do not try to run the test on verify as it is quite intensive already.

Differential Revision: https://phabricator.services.mozilla.com/D167527

UltraBlame original commit: 3d23eb63edd36127a36d66259a7851251e0fa4f0