Bug 1809430 - Make some tests account for smaller default window size on macOS. r=botond
Before these patches we were adding the titlebar height even though we
were not rendering it.

This made an apz mochitest with hardcoded heights to fail on macOS.

It also perturbed browser_html_scroll_restoration.js in a way such as
the scroll position after a resize is rounded in a different direction,
but that is harmless.

Depends on D166431

Differential Revision: https://phabricator.services.mozilla.com/D167027

UltraBlame original commit: c996a95c0bf75502d4a07c0a345679f79db12233