Bug 1704500: Test Windows geolocation utility process restart on crash r=gerard-majax
Make sure that the geolocation utility process restarts for georequests that arrive after a crash. This tests process behavior regardless of whether or not the OS is set to allow geolocation (and in automation, it is not).

Depends on D162944

Differential Revision: https://phabricator.services.mozilla.com/D162945

UltraBlame original commit: c3ae64389c26704128be51e2699153fa6c768df0