Bug 1803109 - Discard blocks of data that are too big for two chunks. r=canaltinova
Currently, `ReserveAndPutRaw` allocates a second span even if the data would be too big for the chunk.
Here a second conditional is added to check if the block of data is too big in this scenario and silently discard the data if so.

Differential Revision: https://phabricator.services.mozilla.com/D167167

UltraBlame original commit: adcb31b93a01f60f6f9928db2c3d52c68d424e79