Bug 1689565 - Part 3: Migrate removed tests r=dom-worker-reviewers,smaug
Ported most of the removed tests from D167236, except the payload test because of a misbehavior regarding to icon URL parsing (and thus fails on Chrome if ported as-is).

It's still covered in dom/serviceworkers/test/test_notification_get.html, so skipping it does not decrease the coverage.

Differential Revision: https://phabricator.services.mozilla.com/D167943

UltraBlame original commit: 58172efe03db176c62202e391a434b61179d4ff3