Bug 1704500: Run Windows' ILocation provider in a utility process r=emilio,cmartin
Previously, the ILocation COM object ran in the main process.  Due to a race condition, it would sometimes crash the process so we are moving it to a utility process, where it can crash safely.  The old WindowsLocationProvider class has become a proxy that forwards requests to the child process via the PWindowsLocation parent actor.  The PWindowsLocationChild executes the requests and sends the results back.  Failures in ILocation (like geolocation being off in Windows settings) are sent back as well -- they will cause the WindowsLocationProvider to start the MLS geolocation fallback, as they did before.

Additionally, this patch caches the accuracy setting in order to guarantee the proper accuracy is set when requesting reports.  Previously, we called SetHighAccuracy too early and it was silently not set (bug 1773456).

Depends on D155019

Differential Revision: https://phabricator.services.mozilla.com/D155020

UltraBlame original commit: 77ef9f7cf3cb82c273004af26c8cfa5f13287b8b