Bug 1811342 [wpt PR 38069] - [Web Platform Test] Just-in-time install for paymentrequest event test, a=testonly
Automatic update from web-platform-tests
[Web Platform Test] Just-in-time install for paymentrequest event test

Before this patch, the payment-request-event-manual.https.html
web-platform-test (WPT) was using the PaymentInstruments API to install
a payment handler, attempted to access the `window` object from a
service worker context to determine the test server origin, and did not
specify `details` in the "paymentrequest" event response.

This was suboptimal because the PaymentInstruments API is planned to be
removed, service workers do not have access to the `window` object, and
the `details` field is required for the "paymentrequest" event response.

This patch switches the payment-request-event-manual.https.html to
install a payment handler just-in-time (JIT), removes the dependency on
the test server origin, and adds the `details` field to the
"paymentrequest" event response.

After this patch, the test is passing locally in full Chrome browser and
is not using any APIs that are planned to be removed.

Bug: 1327265
Change-Id: I525acb167d047505553eab6c030ff6a912cfcecf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4178408
Reviewed-by: Stephen McGruer <smcgruerchromium.org>
Commit-Queue: Rouslan Solomakhin <rouslanchromium.org>
Cr-Commit-Position: refs/heads/main{#1094690}

--

wpt-commits: 505b5cbeb696c227fbab73807645f2422e9ad6cd
wpt-pr: 38069

UltraBlame original commit: 9fc07c947b2677d3d090366ddde9bb92592e9069