Bug 1813697 - Honor full CSS-to-device scale for native GTK widgets. r=stransky
KDE uses the GTK text scale factor to implement fractional global
scaling.

While some GTK apps really only use it for text, we interpret it as a
full DPI change (this is configurable via
browser.display.os-zoom-behavior, but that's our default behavior for a
long long time).

We're ignoring the text scale factor right now, but there's no good
reason to it, now that we support fractional scaling. After this patch,
native widgets are consistently scaled  with all the other chrome UI
elements. This includes checkbox menuitems, buttons, etc.

The exception here are scrollbars, which are special and still ignore
the text scale factor here:

  https://searchfox.org/mozilla-central/rev/8e9b4484408154b80d7ede9e1b035819fda48fd2/widget/ScrollbarDrawing.cpp#28-32

But that's fine because scrollbars are the only widget that ignores full
zoom, and that also matches what Windows does with the text scale
factor, so if we want to change it that's probably worth a separate bug.
(I also have some changes to that code in flight for bug 1812868 so I'd
rather not have gratuituous conflicts).

Depends on D168371

Differential Revision: https://phabricator.services.mozilla.com/D168372

UltraBlame original commit: 46f5eda5e110beb420fdd2d03269f7a4ac56e24e