Bug 1809488 - The TRR mode3 error page should not add exception when the resolver is not accessible r=acreskey,Gijs
If the error is caused by the TRR server not being accessible (maybe the
network is down) then adding an exception is unlikely to make the load.
Most likely the subsequent page load will also lead to an error page.

However, after connectivity is restored, the exception would remain.
Currently the ability to remove exceptions is lacking a UI.
In the future we might want to replace the exception button with a button
that would exclude the current tab from TRR instead - that way the
exception is not permanent when the failure is caused by a connection
failure to the TRR server.

Differential Revision: https://phabricator.services.mozilla.com/D168077

UltraBlame original commit: fed641dc5ab5b212d69350cd5aecb41f5b993a6e