Bug 1814202 - Make browser_contextmenu_badiframe.js more reliable. r=NeilDeakin
The test had a race condition in it where contextmenu events had a chance of
not targeting the right document if a subframe hadn't loaded and presented
in time. This patch uses BrowserTestUtils.browserLoaded and requestAnimationFrame
to help ensure that the subframes have both loaded and presented.

Differential Revision: https://phabricator.services.mozilla.com/D168802

UltraBlame original commit: a6ac768943fac74d0abe7434edb803eab4b22c3e