Bug 1814751 - Treat UniquePtr as a normal IPDL type, r=ipc-reviewers,mccr8
Previously it was a property that all types could have, which was
incorrect, as only the UniquePtrType IPDLType could ever be a UniquePtr.
This changes that logic to make it more consistent with other IPDL
Types.

This also removes the check for a UniquePtr type attached to a `using`
statement, as the type can never be an IPDL type in that situation.

Differential Revision: https://phabricator.services.mozilla.com/D168746

UltraBlame original commit: 72bb438124d09d16be3142e933a483cef4983fbe