Bug 1805476 - Stub out an about:translations page with no real functionality; r=nordzilla
This page is hidden behind a pref, but not behind an ifdef. I chose to
do this in order to simplify the potential release of this feature. I
don't foresee issues with bloating the binary during the development,
so this doesn't feel like a risk to me.

Differential Revision: https://phabricator.services.mozilla.com/D164613

UltraBlame original commit: e9674106dd35a19786f6b7ede12f701d32aafd24