Bug 1809064 - Add PFileSystemAccessHandleControl top level protocol; r=dom-storage-reviewers,jari
Sync access handle methods need to run sync loops to block the thread until a
result is available. The main protocol PFileSystemManager and its subprotocol
PFileSystemAccessHandle is bound to the main worker event queue, so incoming
IPC messages from the parent can't be processed until a sync loop is finished.
This becomes a problem when we want to wait for a response from the parent
during the asynchronous closing. Fortunatelly, we can add a new top level
protocol which will be bound to WorkerPrivate::ControlEventTargret, so incoming
IPC messages will be processed even when a sync loop is active.

Differential Revision: https://phabricator.services.mozilla.com/D166343

UltraBlame original commit: b68e1902183357e129a1a53490447fc3cc219f4f