Bug 1815387 - Don't assert that we get a JSOracleParent. r=smaug
This assert is too strong. We might not get an actor, e.g. in the case
of shutting down. And that's OK. It is up to the caller to make sure
that it actually got an actor.

Differential Revision: https://phabricator.services.mozilla.com/D169078

UltraBlame original commit: ff93d26dfdf8445a8739b68ad9cff31d0354d7ac