Bug 1584886 - Remove browser_oneOffContextMenu_setDefault.js disabling for a test suite that is no longer run. r=search-reviewers,daleharvey
This was originally disabled as it was a perma test failure on aarch64. However, those tests have now been disabled in bug 1620460

Therefore remove the restriction line as there's no use for it. If the tests are re-enabled in future, then we can re-assess if necessary.

Differential Revision: https://phabricator.services.mozilla.com/D169433

UltraBlame original commit: 705cfc94311433da90d3741e2b41718bb991a2d9