Bug 1810166 - Ensure date input focus behavior is tested with month-year panel interaction. r=emilio,kcochrane
Adding tests for the `<input type=date>` focus handling after a month-year picker panel is opened from the datepicker, the value is changed, and the panel is dismissed. We want to make sure that the input does not blur until the other input is actually focused (and the datepicker is closed).

Depends on D167698

Differential Revision: https://phabricator.services.mozilla.com/D169393

UltraBlame original commit: 0fb97cb3e32e357e8df363daf762c2c99124b972