Bug 1816765: allow run-task commands to use command-context when given as a list r=taskgraph-reviewers,hneiva
We already support this when it's a string, we ought to when it's a list as well. This is particularly useful when you have a job or task template in a kind, and need to be able to append to the command in concrete jobs, while still using command-context.

Differential Revision: https://phabricator.services.mozilla.com/D169847

UltraBlame original commit: a539fb9b6b8400b134b25fa6d6b6bdc76c436e56