Bug 1784662 - Reduce time spent in browser_ext_omnibox.js test r=willdurand,adw
This is a test-only change. It introduces a new Urlbar pref to make the
existing 3000 ms timeout configurable.

That timeout is currently reached 8 times in the test. That means that
the test will take at least 3 x 8 = 24 seconds to run while basically
doing nothing, and makes the test prone to reach the deadline by which
the test is declared timed-out.

To avoid "Test timed out" errors, the test now reduces the fixed 3000 ms
timeout to 500 ms, so now we wait for 8x500 = 4 seconds instead of 24.

Differential Revision: https://phabricator.services.mozilla.com/D169846

UltraBlame original commit: be309ac9bf47e706fba48695ed1c9f42f828c791