Bug 1817188: Only do structured clone when differential fuzzing r=jandem
The alias set that we use for MFuzzilliHash depends on suppressions in structured cloning that are only turned on when we do differential fuzzing. Changing the alias set based on differential fuzzing runs into problems with add vs addEffectful, so the simplest fix seems to be returning early if the suppressions aren't active.

Differential Revision: https://phabricator.services.mozilla.com/D170141

UltraBlame original commit: 661f273860abcd38fa370df932adb646c675839f