Bug 1810816 - P3 Report response timing for PFetch. r=dom-worker-reviewers,asuth
Response Timing should be reported before the body stream completing.

In the original implementation, the response timing is recorded when OnResponseEnd is called, but it is too late. Response timing could not be ready when response.text() or response.blob() resolves the promise.

This is the third patch for reporting the response timing for PFetch.
This patch adds the needed IPC PFetch/Send/RecvOnReportPerformanceTiming to propagate the response timing from the parent process main thread to the content process worker thread.

Depends on D167937

Differential Revision: https://phabricator.services.mozilla.com/D167938

UltraBlame original commit: e5eacedc370d64a39b6dff5a4a767b37e8af7cd3