Bug 1812156 - Part 2: Require all IPDL using statements to be fully qualified, r=ipc-reviewers,mccr8
This is done by changing the formatting of QualifiedIds in the compiler
to include a leading `::`, anchoring the path in the global namespace.
This means that the generated code will fail to recognize imports by
relative paths, as they will not be found in the correct namespace.

This required no longer using QualifiedId in TypeSpec, as that type is
generally used for local types in IPDL, which are never qualified in a
namespace (only `using` statements actually support being qualified). As
part of this, I also removed the unnecessary wrapper TypeSpec from using
statements, which was never used.

Finally, this also required changing how we handle builtin imports for
the C primitive types, as these types (such as `int`, `bool`, or
`double`) are special keywords, and can't be named in the global
namespace. They are now given a new builtin type name for C builtins,
and defined separately from other using statements.

This change also means that typedefs will now appear in generated code
for types in the global namespace, but this is required to double-check
that the type is actually in the global namespace.

Differential Revision: https://phabricator.services.mozilla.com/D168878

UltraBlame original commit: 2208cd8db3ca17bfc879891f67bdafb842bb367e