Bug 1816108 [wpt PR 38446] - Set the "previous" page name when setting up the fragment builder., a=testonly
Automatic update from web-platform-tests
Set the "previous" page name when setting up the fragment builder.

Every in-flow child, when adding its fragment to the builder, will
propagate its page name to the builder, so that the builder can keep
track of the previous (or current) page name.

The decision whether to break before a child due to a page name change
is made by essentially comparing the page name of the fragment (that
we're considering to add to the builder) with the previous page name
stored in the builder.  Normally this check won't be performed on the
first in-flow child, since there's usually no break opportunity there -
breaks are inserted *between* siblings, not between a parent and its
child.

However, if there's a class C breakpoint before the first in-flow child,
due to a float, we're going to perform this check at the first, and then
it's important that the builder knows the page name. Without this fix,
we'd detect a page name change (from nothing to "foo" - see the test
included), and insert a break before the first in-flow child.

Bug: 1414718
Change-Id: If7d85f7b0e611f5a5d570de586126b0534478c4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4238838
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Cr-Commit-Position: refs/heads/main{#1103954}

--

wpt-commits: 63f502d84afc2fde3c92555a5eb2fcf825734d2f
wpt-pr: 38446

UltraBlame original commit: daf35fbb763c12d7f47ac93128aa2e79a3eff1d0