Bug 1809518 - Replace nothing-cross-origin-corp.txt with a JS file from COEP tests r=farre
nothing-cross-origin-corp.txt can be replaced by any other files,
because it's used a dummy file. This patch replaces it with a different
dummy file that has JS as the content, to avoid ORB blocks the request.
This doesn't change the semantic of the tests.

Depends on D170911

Differential Revision: https://phabricator.services.mozilla.com/D170912

UltraBlame original commit: 29720d7132cf060cebc7afd6f8b0313a935ed55a