Bug 1819279 - [devtools] Switch to a block list of known errors instead of an ignore list r=jdescottes
This patch switches from an ignore list of errors to a block list of the errors
we clearly want to block. We are likely going to regress a lot by missing errors from the ignore list
which shouldn't be blocked as revealed by this BUG. Instead let just block the errors we know about and
let other erros through as most of them are shown with error codes and do not need to be blocked.

Differential Revision: https://phabricator.services.mozilla.com/D171352

UltraBlame original commit: 1440eb683cd134808b075e46a4b3b7287b536efc