Bug 1658986 Part 2: Add region stringification to gfxCriticalError before crashing. r=mstange
We're headed for a crash, so this change makes gfxCriticalError have more
information that will help us understand the crash. Unlike some other
error-checking code in this class, it is not NIGHTLY_BUILD because the
crash signatures in this Bug are reported in Beta and beyond.

Differential Revision: https://phabricator.services.mozilla.com/D171390

UltraBlame original commit: 120ea59f799fab5c12103163facd36a1935784fd