Bug 1816064 - Part 2: Implement the fingerprinting randomization key to cookieJarSettings. r=tjr,pbz
This patch adds a fingerprintingRandomizationKey to the
nsICookieJarSettings. The random key will be generated when loading a
top-level http channel and set to the cookieJarSettings. The key will
propagate to the top-level documant and all sub-documents with the
cookieJarSettings. Accessing the fingerprintingRandomizationKey will
return an error if the fingerprinting randomization is disabled.

Differential Revision: https://phabricator.services.mozilla.com/D170604

UltraBlame original commit: f605a7ddf9502fe74187f430417e9686d10d4b96