Bug 1820189 - Ensure metrics_yamls and pings_yamls contain no duplicate files r=TravisLong,nalexander
Because Firefox Desktop has multiple app ids, but builds to a single binary,
there are cases (like crashes) where the same metrics and pings will be
included in multiple `_metrics` or `_pings` lists.

This is important to allow to happen for `fog-update-bot` to know that all of
those different app ids report the same metrics.

But it's not something we should pass to glean_parser, because even if it
allowed multiples, we don't need multiple copies of these metrics and pings
to be generated.

Depends on D171575

Differential Revision: https://phabricator.services.mozilla.com/D171576

UltraBlame original commit: 06b02b7eb88725950c0740c4b218d3ef841336d6