Bug 1819540 - Apply Mali driver bug workaround on additional GPU. r=gfx-reviewers,lsalzman
In bug 1787520 and bug 1795614 we added a workaround for a driver bug
on Mali G78 and G710 GPUs. It turns out this also affects the G77, now
that some devices have received software updates containing an updated
driver version. This patch applies the workaround to the G77 as well,
as long as the driver version is affected.

Differential Revision: https://phabricator.services.mozilla.com/D171740

UltraBlame original commit: 3bf5fce9130e653bffa77a8ef6119df92b27f6b7