Bug 1818866 [wpt PR 38707] - Skip anonymous blocks in prepaint, a=testonly
Automatic update from web-platform-tests
Skip anonymous blocks in prepaint

When handling OOFs w/ an inline CB in a monolithic ancestor, we would
end up in a case where we were using the wrong paint context, which
resulted in the OOF being painted at an incorrect offset.

We got into this state because we would end up setting the OOF's
container during prepaint to an anonymous parent block of the inline.
Because an OOF cannot be placed in anonymous blocks, we would fail
to find the OOF as a child of that anonymous block.

This would lead us into the same path as fragmented missed OOFs during
prepaint, and we would assume the CB was the fragmentainer. We would
update the context based on this, which is why the OOF was placed
relative to the fragmentainer instead.

To fix this, avoid setting the OOF's CB to an anonymous block during
prepaint to ensure that the OOF is not missed in this case and goes
down the normal (non-fragmented) OOF prepaint path.

Bug: 1412313
Change-Id: I0dfb1a67ddd96404a725cdbf56493de675eb04e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4292046
Commit-Queue: Alison Maher <almahermicrosoft.com>
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/main{#1110381}

--

wpt-commits: b07ee811c0db7d43ae876c21fe9346ce6677f532
wpt-pr: 38707

UltraBlame original commit: 42375c1c113491e6d8ef0fed9648b833c881c31b