Bug 1774569 - Vendor `jsonschema` at `4.10.0` r=ahal
`Poetry` version `1.4` depends on `jsonschema>=4.10.0`. Unfortunately, our current implementation of `./mach vendor python` cannot see that specific dependency, because it specific to the `vendor` site.

`glean-parser` depends on `jsonschema>=3.2`, and when the lockfile is generated, it pulls in `jsonschema==3.2`, which causes an incompatibility with `poetry` in the `vendor` site.

Fortunately the loose dependencies of `glean-parser` and `poetry` are compatible, the lockfile just doesn't know it. As such, if we explicitly specify `jsonschema==4.10.0`, the compatibility in the `vendor` site is resolved.

This is a temporary workaround and will be resolved by either vendoring `poetry` itself, or adding `lockfiles` to every site.

Differential Revision: https://phabricator.services.mozilla.com/D171760

UltraBlame original commit: e027acb00eb5ca53face4fb185006f80e1910bae