Bug 1773549 - [devtools] Add widget for linear easing function. r=devtools-reviewers,jdescottes.
This adds an authoring tool similar to the one we have for cubic bezier.
A swatch is added next to `linear()` functions and clicking on it will show
the widget.
Modifying the line in the widget will update the value of the `linear()` function
in the rule view.

Differential Revision: https://phabricator.services.mozilla.com/D153333

UltraBlame original commit: f62858af25e4e95fac3b50c3a7a94f9c233026b5