Bug 1787008 - Add some dumping of load notifications to a test. r=dholbert
Seems per the logs we're hanging at the beginning of the test which is a
bit surprising. Try to see if we're failing to load any browser in
particular or we're getting an unexpected notification or what not.

Differential Revision: https://phabricator.services.mozilla.com/D172054

UltraBlame original commit: 256a0e3ed82c20eb9eb6d97dba47ee050440bf34