Bug 1821118 - Add a non-zero min-height for the views box. r=Gijs
This prevents the tree header etc from overflowing when the window is
too small to fit that and the details pane together.

The overflow: clip declaration is just an in-depth defense against
overflow, but it isn't really needed because stuff in there should be
scrollable for the most part.

Differential Revision: https://phabricator.services.mozilla.com/D172037

UltraBlame original commit: ea68b31a90570e6e7e17e080a9a45be781b239bf