Bug 1784772 - Forcibly flush scroll position change. r=botond
The lack of the flush is one of the causes of the test timeouts. Without the
flush, the content in question is not overscrollable, thus a swipe gesture
triggers a navigation back to the previous content rather than overscrolling.

Differential Revision: https://phabricator.services.mozilla.com/D172170

UltraBlame original commit: 12999bd7cc248199c1406331c5ca1926a6e2b3ed