Bug 1821763 - Restore center alignment for not-in-content dialog. r=Gijs
So I overlooked in the regressing bug that most of the time
.dialogOverlay is a flexbox (in which case we do want center alignment).

It's only for in-content dialogs where it's a grid (and we want the
unset/stretch behavior).

Since the new rule also applies to window-modal-dialog, remove the
unneeded align="" attribute there.

Differential Revision: https://phabricator.services.mozilla.com/D172328

UltraBlame original commit: 3088c308450dfb8081d1861f25d5e718901c6ad2