Bug 1814727: adjust beetmover tasks to be downstream of attribution tasks for Windows r=releng-reviewers,gbrown
This is mostly just switching the upstream tasks with pull the Windows installers from. The only wrinkle is that we're not attributing the asan-reporter installers (and we shouldn't IMO), so I had to add support for keying upstream tasks on platform in the beetmover manifests. (If we simply listed all three possible upstreams, we ended up pulling _two_ installers for platforms that are attributed...and I don't even know which one would get published, or if both would.)

Differential Revision: https://phabricator.services.mozilla.com/D170246

UltraBlame original commit: 14d94e19b541d28ba1dc1b1b82d4740445fbf360