Bug 1822094 - Add a shared class for deemphasized text in the new Migration Wizard. r=kpatenio,desktop-theme-reviewers,dao
Notably, this also updates the em ratio of the "base" font size and the deemphasized font-size.

In the Figma specification, the base font size is 13px, and the deemphasized font-size is 10px,
which is ~84% of 13px.

Differential Revision: https://phabricator.services.mozilla.com/D172443

UltraBlame original commit: a0f99f252b7d6fd0886f11967b979b59deb8c638