Bug 1820168 - Let the system re-set the current cursor when needed. r=mac-reviewers,bradwerth
This renames nsMacCursorManager to MOZDynamicCursor and makes it implement NSCursor.

We also install a cursor rect which covers the entire ChildView and which refers
to the MOZDynamicCursor singleton.

With the help of the cursor rect, the -[NSCursor set] implementation is now called
at the right times.

Here's a profile which contains the stacks under which -[MOZDynamicCursor set]
is now called: https://share.firefox.dev/3yEXLsC

Differential Revision: https://phabricator.services.mozilla.com/D172484

UltraBlame original commit: 03bc416d339956f046f032ed688cc28e19555505