Bug 1607634 - Part 4a: Wrap non-nullable actor types in NotNull, r=ipc-reviewers,mccr8
The code changes required for this will be in part 4b. The code is designed to
be generic as it will also be used for RefCounted types in a future patch. This
replaces the manual checking which was previously done in struct and union
serializers for non-nullable actor members.

Differential Revision: https://phabricator.services.mozilla.com/D168886

UltraBlame original commit: aebbaa145d2d2d0035f60d3aa6ab1838f20435b4