Bug 1822771 - Don't shrink menuitems / menus. r=dao
An alternative fix would be to remove this min-height here:

  https://searchfox.org/mozilla-central/rev/8ed22fcd56968c95a73a6c82b42f732f01a4bdae/browser/themes/shared/customizableui/panelUI-shared.css#968

Which makes min-height auto come into account and constrain the items.

But squishing menus is unlikely to be what you want (if only because
they're always in a scrollable box). So this also works.

Differential Revision: https://phabricator.services.mozilla.com/D172802

UltraBlame original commit: 0c6905a6718fd78cf45473a6e0810c54d0115064