Bug 1822918 - Fetch the weather suggestion after wake and coming online. r=daisuke
This listens for some observer service notifications so we can re-fetch the
weather suggestion after wake and coming online, without having to wait for the
timer to fire. Please see the bug for background.

I changed the interval (repeating timer) to a timeout (non-repeating timer)
since the timer needs to be recreated on wake if the wake happens before the
start of the next fetch period.

Differential Revision: https://phabricator.services.mozilla.com/D172844

UltraBlame original commit: 2144589764eb6f81c4f58785f01bbc6ed712ef47