Bug 1823159 - Avoid racily initializing the PermissionManager twice during startup, r=timhuang,baku
It appears that the issue is caused by a change in behaviour from bug
1799470 which made the profile available earlier during startup. If the
permission manager is started between the time when the profile becomes
available and profile-do-change, it would've previously initialized
without storage, and initialized storage after profile-do-change, but
after the changes it may initialize twice instead.

This patch changes the profile-do-change handler to check if we already
have a profile, and avoid initializing the DB twice in that situation.

Differential Revision: https://phabricator.services.mozilla.com/D173062

UltraBlame original commit: 9c40ff363e56c6377daa85e64c930ea80617b1f4