Bug 1823622 - Part 3: Lower the threshold at which we calculate mark rate telemetry r=sfink
We don't calculate this when marking is quick to exclude unreliable results
when the denominator of the calculation is very small. However we can mark in
less than 1 millisecond for small heaps with parallel marking, so this reduces
the threshold to 1 microsecond.

Depends on D173128

Differential Revision: https://phabricator.services.mozilla.com/D173129

UltraBlame original commit: 2d6a608863af31027707cc2c92db1fb3d4cbc50b