Bug 1816165 - [hazards] Use common getCallees() from analyzeRoots.js and remove the gcEdges mechanism. r=jonco
Now that both the callgraph generation and the hazard detection phases use the same source of callee information, the gcEdges mechanism to transmit data between them is no longer needed. Which is good, because the whole analysis is basically a map/reduce operation:

 - map: extract callee information from all functions
 - reduce: merge into a global callgraph
 - map: gather hazards from all functions
 - reduce: merge into a single output

The gcEdges output caused the first map step to produce two different outputs, which was a bit of a nuisance for the driver. (Now it can handle multiple outputs and no longer needs to. Oh well.)

Differential Revision: https://phabricator.services.mozilla.com/D170150

UltraBlame original commit: 5da1a824ea4163fb02193cbdc68f3923e9a833c0