Bug 1823751: Make macOS pasteboard use NSString for urls, and don't convert to NSData. r=mac-reviewers,spohl
This change makes PasteboardDictFromTransferable use NSString to store the
file url in the NSDictionary, and then when converting to the native
pasteboard, check whether we have an URL and use the setString method
in that case. This complies with the macOS requirement that NSString
content is used for the kUTTypeFileURL key.

There are further opportunities for improving usage of NSPasteboard,
specifically that all URLs should be written using writeContents, that we
shouldn't be using the kUTTypeFileURL key as an NSString, but those issues
are not addressed in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D173359

UltraBlame original commit: baa3ac5e605e9c61f67100012ff4cc4915e9c9e1