Bug 1823811 - Count this.x assignments inside functions and use to estimate property count r=jandem
This is of course a strict over-estimate, as `this.x = 1; this.x = 2` would
scan as two properties not one, but it's most important to get the correct
size class of object allocated, and this is low investiment to do so.

We also include the count of class fields to handle large ES6 classes.

On an allocation micro-benchmark, this produces a speedup of 2x on a microbenchmark modeled
on the matrix-react benchmark, and a speedup of about 60% on a class allocation benchmark.

  $ DISABLE_SIZE_ESTIMATE=1 ./mach run t.js
   0:00.72 /home/matthew/unified/obj-opt-shell-nodebug-x86_64-pc-linux-gnu/dist/bin/js t.js
  Function  10.98023681640625
  Class:    11.85999755859375

  $ DISABLE_SIZE_ESTIMATE=0 ./mach run t.js
  Function  5.34127197265625
  Class:    7.23751220703125

(The environment variable isn't included in this patch)

Differential Revision: https://phabricator.services.mozilla.com/D173576

UltraBlame original commit: a1f717fd667775ca2222db464b4a448e1bf846fb