Bug 1815790 - Replace intgemm by gemmology r=yury
gemmology is a rewrite of intgemm based on xsimd (that we already
vendor), with a focus on the API we actually use.

It also supports sse2 and has a decent implementation for arm32 and
aarch64.

This patch integrates sse2, ssse3 and avx2 support.

Differential Revision: https://phabricator.services.mozilla.com/D171265

UltraBlame original commit: 6cc1c7fef689246819ff760cccd2c7ff7d7ee12f