Bug 1821133 [wpt PR 38892] - VT: Be resilient if there is no documentElement()., a=testonly
Automatic update from web-platform-tests
VT: Be resilient if there is no documentElement().

This patch updates all of the places where we access documentElement()
to make sure it either is there (via DCHECKs) or gate access
conditionally.

This is to address the referenced clusterfuzz. I can't repro this
locally, but it is clearly an access into a documentElement() when it's
null.

Since we can handle the null situations gracefully in all cases, this
is just an improvement to the robustness of the code.

R=khushalsagarchromium.org

Bug: 1413903
Change-Id: I1b81e36a2b88291e364407241c3b0fd0c34de252
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4317782
Reviewed-by: Khushal Sagar <khushalsagarchromium.org>
Commit-Queue: Khushal Sagar <khushalsagarchromium.org>
Cr-Commit-Position: refs/heads/main{#1114882}

--

wpt-commits: f87ddce734451b714fe46eaec46d8118901162f6
wpt-pr: 38892

UltraBlame original commit: 049c1a1c5470603d17712cb93155429d60a09533