Bug 1821044 [wpt PR 38877] - Simplify and fix named pages layout., a=testonly
Automatic update from web-platform-tests
Simplify and fix named pages layout. (#38877)

If an element had a page name specified, we'd use that as the start page
name. But we should rather let descendants get a chance to set set a
page name first, and only if no name is propagated from descendants, we
should use that of the element itself.

<div id="first" style="page:a">a</div>
<div id="second" style="page:a">
  <div id="inner" style="page:b">b</div>
</div>

Because of this propagation failure, the start page name of #second
would be "a", so we wouldn't insert a break before it (since we're
already on a page with that name). And breaking before #inner isn't
possible, since there's no valid class A break point before it (such
break points only exist *between* siblings). In other words, we depend
on the page name to be propagated upwards until we find a valid break
point, for this to work correctly.

When propagating page names from children, just use that value, not the
page name of the container fragment itself, let alone the page name set
somewhere in the ancestry. Use the specified page name when finishing
layout, but only if no descendants have provided one.

Although the spec talks about start and end page names [1], we really
don't need to keep track of that during layout. Once we get a different
page name, we need to break anyway. Simplify everything. Rather than
having a start and end page value in NGLayoutResult, and the page name
in NGPhysicalBoxFragment, just keep the latter, and use it for
everything.

[1] https://drafts.csswg.org/css-page-3/#using-named-pages

Bug: 1421421
Change-Id: I70306e6949639208942f547c7bce22b51b3c78b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4316229
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/main{#1114578}

Co-authored-by: Morten Stenshorne <mstenshochromium.org>
--

wpt-commits: 14b08e9b23f521bd6410395c13d69a5257d0662e
wpt-pr: 38877

UltraBlame original commit: 3eb7b466669fabfd009ee6a2dd93ab8c39e53603