Bug 1825603 - Add more logs for RemoteMediaDataDecoder r=media-playback-reviewers,alwu
The `VideoDecoder` interface, associated with a `MediaDataDecoder`, is
going to be introduced in Bug 1749045, which can be run on either the
main thread or the worker thread, in the content process.

While the `VideoDecoder` that `RemoteMediaDataDecoder` is associated
with can run on the worker thread, the thread where
`RemoteMediaDataDecoder` runs on, hosted by `RemoteDecoderManagerChild`,
is created and released on the main thread only. Without extra
carefulness, it's easy to have some racing issues in the code with these
complex thread arrangements. It's even harder to figure out the causes
of the issues without clear logs. Adding more RDD logs can be helpful to
debug, especially when rr-recording couldn't capture some
timing-critical intermittent racing issues.

Differential Revision: https://phabricator.services.mozilla.com/D174147

UltraBlame original commit: b3c55744883fba84166fcd1819bd357db351ac74