Bug 1781941 - [devtools] Wait for original sources when mapping frames. r=devtools-reviewers,bomsy
With the new location object having direct references to source object,
we now have to wait for sources objects to be registered in the reducers
before trying to instantiate the related frame locations.

Also fix a typo in waitForSourceToBeRegisteredInStore.

Last but not least, we no longer need to wait for the source related to
the selected frame as mapFrames will do it.
Now we have a strong guaranty that frame.location will be set
with a valid source object. That for both generated *and* original sources.

Differential Revision: https://phabricator.services.mozilla.com/D173838

UltraBlame original commit: 1c50d1ff30f7c5986c861e5d3625c53000e13f06