Bug 1826743 - Remove shadowing variables from the apz.scrollthumb.calc=false branch. r=hiro,dlrobertson
The |translation| variable was redeclared in the branch, shadowing the
variable in the function's top-level scope, but it's the latter that
was actually used to transform the thumb.

The patch also removes a shadowing |asyncScroll| variable, though that
one was harmless.

Differential Revision: https://phabricator.services.mozilla.com/D174936

UltraBlame original commit: b89011e45f7f5086230b87679898504ad845874f