Bug 1827214 - Avoid unnecessary (and slow) jitdump mmap on macOS. r=dpalmeiro
On Linux, this mmap is needed so that the perf.data file contains MMAP2
records with the filename.
On macOS, samply hooks the `open` (and `fopen`) function to observe the
jitdump path, and no mmap is needed.

Depends on D175046

Differential Revision: https://phabricator.services.mozilla.com/D175047

UltraBlame original commit: 2969fa3d6880c1e59395dfbcef961580fbd00d81