Bug 1818364 - Run performance tests on Chromium-as-Release (Linux). r=perftest-reviewers,sparky
This patch lays a foundation for running chromium-as-release binaries with our raptor-browsertime tests.
To start with, this just deals with the Linux platform and minimal changes will be required in patches in the near future for Win/MacOS platforms.
To begin with, we only test on priority benchmarks at the moment (e.g. sp/sp3).

Differential Revision: https://phabricator.services.mozilla.com/D173757

UltraBlame original commit: 765c496e216c2dfd5ac18e726d506681ae3a37a4