Bug 1823392 [wpt PR 39082] - [scope] Introduce RelationType::kScopeActivation, a=testonly
Automatic update from web-platform-tests
[scope] Introduce RelationType::kScopeActivation

We currently match :scope incorrectly: in CheckPseudoScope, we consider
the selector as matching if there is *any* activation root that matches.
This gives incorrect behavior for e.g. ':scope > :scope', because
we can match :scope against different activation roots within the same
selector. This is not correct. Instead, the element chosen to represent
:scope must remain stable across a given selector evaluation, and then
we have to try the whole selector N times for N different roots,
in principle.

In order to avoid (re)trying the whole selector N times for N
activation roots, we prepend a special relation kScopeActivation
to any compound that contains :scope and/or &. When we encounter
kScopeActivation during selector matching, we try each activation root
(in order) as the :scope element, until we find a match. This way
we avoid retrying the *whole* selector N number of times: we only
retry from the point of the first compound which contains :scope/&.
E.g. for '.a > :scope > .b > .c', the selector really looks like
the following internally: '.a > :true$:scope > .b > .c' (using $ to
represent kScopeActivation). This means we can evaluate the '.b > .c'
part normally, without doing any work related to scope.

The kScopeActivation relation is always attached to the :true
pseudo-class, and is inserted parse-time as follows:

 1. Before the implicit :scope that is prepended with a descendant
   combinator if the selector is not ":scope-containing".
 2. Before the implicit :scope that is prepended to relative selectors.
 3a. Before compounds which contain :scope specified by the author.
 3b. Before compounds which contain '&' specified by the author.

(3b) happens because having '&' in a selector makes it
":scope-containing", and is therefore not guaranteed to otherwise
contain :scope. Additionally, '&' behaves like :scope when there is
no parent selector, which is the case for scope rules which have
no <scope-start> selector.

This CL breaks "scope (:host)". Another solution must be found for
dealing with this (Issue 1418304).

[1] https://drafts.csswg.org/css-cascade-6/#in-scope

Fixed: 1418014
Bug: 1418304
Change-Id: If46aca89c3ebe83b62bbd1f7c9ffbaa5b5692e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4345796
Reviewed-by: Rune Lillesveen <futharkchromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruudchromium.org>
Cr-Commit-Position: refs/heads/main{#1119334}

--

wpt-commits: 3e81feb63b43e75ee58ecff553b7acf37aef2ab1
wpt-pr: 39082

UltraBlame original commit: 39f5995e0c92faa325593605846e5730170fc712