Bug 1827621 - Add a mininum heap size threashold at which to use parallel marking r=sfink
Benchmark results show that using parallel marking is not always a win for
small heaps, so add a threshold for its use.

This is set to 10MB, which will result in using parallel marking ~70% of the
time.

This parameter is set to zero for jit-tests and shell js tests so we get test
coverage where heap sizes are expected to be small (and maximum performance is
not a goal).

Differential Revision: https://phabricator.services.mozilla.com/D175249

UltraBlame original commit: 127346707e747e34fa883f7348e1d1bf2bfb33bb