Bug 1827918 - Part 1: Don't call TenuredCell consturctor but rely on derived constructor r=sfink
Currently we call both the TenuredCell consturctor and the derived constructor
for tenured allocations. While this is not illegal, it's strange and we should
rely on the derived constructor.

The patch changes the tenured allocation paths to use void* pointers to
unconstructed memory.

I don't know why the change to cell header initialization was necessary. It
seems the TenuredCell constructor initialized the header to zero for some
reason, but I don't see why it should have.

Differential Revision: https://phabricator.services.mozilla.com/D175372

UltraBlame original commit: e8cfac05da4abd82d0a364abe284c00903e03acb