Bug 1824806 [wpt PR 39223] - Attempt to move past monolithic content when printing., a=testonly
Automatic update from web-platform-tests
Attempt to move past monolithic content when printing. (#39223)

Let monolithic content take up additional space on pages they overflow
into, if this prevents content overlap (that wouldn't occur it we
weren't paginated).

When printing, when we detect that something overflows the page, we'll
store this overflow in the break token, so that we can move past it on
subsequent pages. This kind of works, because, in our implementation,
pages are stacked in the block direction, so that the block-start offset
of the next page is the same as the block-end offset of the preceding
page.

The current solution here only aims to get the basic things right-ish:
inside regular block containers, inside floats, and inside out-of-flow
positioned elements. In addition, we need to make sure that repeated
table headers and footers don't get messed up. We have to repeat those
on every page, even if there's monolithic content in the way. In
addition to actually (probably) making sense visually, our
implementation requires this: once we have decided to repeat a table
section, we need to be consistent about it, or the fragment repeated
will get upset.

For layout systems like flex, grid and tables, we don't attempt to do
anything special, for now. This shouldn't cause major problems, but it
does mean that content in a flow parallel to the one that has monolithic
content (e.g. inside sibling table cells) will be affected (pushed down)
by the monolithic content, even if there is no need to do so.

Content next to floats with monolithic content inside also doesn't
behave correctly. To fix this, we would need to translate and copy the
exclusion space from one fragmentainer to another. It should be a rather
straight-forward thing to do, but I'd like to see a use-case for it
first.

Firefox appears to do something very similar, to prevent monolithic
overflow from overlapping with subsequent content, with limitations
comparable to what this CL does.

Bug: 1402540
Change-Id: If2bfec4737ea9ee7169036144f590a913123b6be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4370688
Reviewed-by: Alison Maher <almahermicrosoft.com>
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/main{#1123282}

Co-authored-by: Morten Stenshorne <mstenshochromium.org>
--

wpt-commits: 155ae72eff461d2dfa02467f7af76c73660add39
wpt-pr: 39223

UltraBlame original commit: ab5135108051ba03d3d5b6c8a74c0c6a5b167be7