Bug 1827830 - Simplify AutoSetNewObjectMetadata more. r=jonco
The problem in the bug is that in the browser, GC can call back into JS code,
violating the assertion checking that we have only one `AutoSetNewObjectMetadata`
active on the stack.

In debug builds, we now use a counter for the number of active `AutoSetNewObjectMetadata`
instances on the stack.

This patch simplifies the object metadata code more. The `NewObjectMetadataState`
variant was adding a lot of complexity. It's simpler to remove it and use the
`numActiveAutoSetNewObjectMetadata_` field for similar assertions.

Differential Revision: https://phabricator.services.mozilla.com/D175456

UltraBlame original commit: f36db2d7462595158f8e2b6c0cc0e22adae2e0a2