Bug 1809941 - [devtools] Add line break after parens when content is long. r=devtools-reviewers,bomsy.
When dealing with a `(` token, we check the next tokens to find the matching
closing paren. In the same time, we compute the length of all the tokens inside
the paren.
If we reach a given length (60 at the moment) without finding the closing paren,
we consider it "long" and will treat it (and the closing paren) as line delimiter.
On top of that, we expand line delimiters to `||` and `&&` when there not in a
"short" paren, so we can split up long lines.

Differential Revision: https://phabricator.services.mozilla.com/D175477

UltraBlame original commit: 62af6f051232687464725688d0dbc9adbbdee4ff