Bug 1825709 - Account for min/max bsize in nsTextControlFrame::ComputeBaseline. r=dshin
This isn't needed for nsTextControlFrame because its ComputeAutoSize
implementation doesn't return an unconstrained line-height for inputs,
so we never end up in the UNCONSTRAINEDSIZE case, but it's needed for
date/time inputs.

Use GetLineHeight while at it, since it's the inflated line-height which
is what we want, and may be cached so we can avoid computing it.

Maybe in the future we can make date/time inputs just use
nsTextControlFrame, which would prevent this from happening in the
future.

Depends on D175745

Differential Revision: https://phabricator.services.mozilla.com/D175746

UltraBlame original commit: 4d22afc76172c03bdde41882bfaafd1f91e3ab50