Bug 1827440 - Stop sending a redundant actor reference in Send__delete__, r=ipc-reviewers,mccr8
Previously we would both send it as the routing target for the actor as
well as as an implicit actor parameter. This patch keeps the same API
from the caller's point of view, but avoids sending the second reference
to the actor in the Send__delete__ message, which also avoids potential
issues with the second reference being null under fuzzing.

Differential Revision: https://phabricator.services.mozilla.com/D175545

UltraBlame original commit: 8dc8887ffdf574ad92e35148904ffffadeb6c7a1