Bug 1828396 - Check for partially initialized objects in GC compartment checks r=jandem
Ideally we'd make the object's state consistent if we fail to allocate dynamic
slots, but that's inconvient here because it could require allocating a shape
to match the reduced slot span.

Instead, check for partially initialized objects in the only place we encounter
them which is in the GC's compartment checking.

Differential Revision: https://phabricator.services.mozilla.com/D175848

UltraBlame original commit: 5beb3159e6ff559b3bab933a1523585e46e7c6fc