Bug 1826681: fix(webgpu): propagate `label` to new shader modules r=webgpu-reviewers,nical
We don't set the [`GPUObjectBase.label`] of the `shaderModule` we create in our implementation of [`GPUDevice.createShaderModule`], but we should. Fix it! You can see this reflected in the error diagnostics currently returned by Naga in the [`GPUCompilationMessage.message`] field.

[`GPUObjectBase.label`]: https://www.w3.org/TR/webgpu/#dom-gpuobjectbase-label
[`GPUDevice.createShaderModule`]: https://www.w3.org/TR/webgpu/#dom-gpudevice-createshadermodule
[`GPUCompilationMessage.message`]: https://www.w3.org/TR/webgpu/#dom-gpucompilationmessage-message

Differential Revision: https://phabricator.services.mozilla.com/D175405

UltraBlame original commit: 8aca8c275bebbd8316cd88adddac09dcf474c187