Bug 1829493 - [wayland] Don't keep firing vsync at idle rate once idle. r=rmader
Once we detect we're idle, remove the idle timer (until the frame callback
fires again, which would start the timer).

That makes vsync not fire at 1Hz on occluded windows (it makes vsync not fire
at all for those). Matching the windows behavior too.

We had this so that we didn't leak in bug 1786247, but the right fix is
bug 1828587.

Differential Revision: https://phabricator.services.mozilla.com/D176282

UltraBlame original commit: d43bc8b819219400e31b7b4ac04570f9205dffb5