Bug 1828966 - Ensure Pin to Toolbar context menu item shows the correct state when opening the context menu. r=mconley
I could only reproduce on macOS but it is indeed a valid bug. I noticed
that moving the `await` code at the bottom of the function fixed it but
that didn't sound too great.

I decided to propose this patch instead, and I made sure that we
wouldn't hit a similar problem with the "disabled" state of the remove
button. Apparently, that still works as expected (on macOS).

I am not quite sure how to cover that with a test, though, because we
have test coverage for the attributes and the HTML markup is correctly
updated. For example, I could see the correct `checked` attribute in the
HTML (in the devtools) with the wrong context menu item state rendered,
sigh. This likely means the race is at a lower level (cocoa?).

Differential Revision: https://phabricator.services.mozilla.com/D176416

UltraBlame original commit: 123d6c05b43fbf76d56f9326a177a1851610133d