Bug 1274334 - Enable STL wrappers on macos. r=firefox-build-system-reviewers,ahochheiden
Practically speaking, this doesn't make a huge difference, because
libmozglue contains operator new/operator delete overrides. This will
enable inlining opportunities, though.

It is worth noting that there's equivalent code in
js/src/old-configure.in but it's left as is because it is actually no-op
at the moment. This will be fixed in bug 1829049.

The check for __EXCEPTIONS is replaced with __cpp_exceptions because the
former is defined for any type of exception, including ObjC exceptions,
while the latter is defined for C++ exceptions only.

Differential Revision: https://phabricator.services.mozilla.com/D175976

UltraBlame original commit: bf43611b4c0e3b53b11630c575dc2418f47207f8