Bug 1724336: Add some more URLPreloader path sanity checks. r=mccr8
Most of the startup assertions we're seeing are from when we decode multiple
paths that are empty strings. That doesn't seem likely to happen very often
from on disk corruption, so presumably it must be happening before/when we
encode the cache file?

I don't have much hope that this will turn up the cause, but it seems worth a
try, and it definitely shouldn't hurt.

Differential Revision: https://phabricator.services.mozilla.com/D176403

UltraBlame original commit: b3b31200864236d6e5bf16746b574e474238bedb