Bug 1829733 - Remove extraneous block margin around migration wizard checkboxes. r=niklas
This also updates the iterator that was - unchecking and hiding all of
the resource checkboxes during the start of the routine that runs
when the browser / profile selector changes. This original iterator
was accidentally including the "Select All" checkbox from Variant 2,
and setting the "hidden" attribute on it. That attribute has no effect,
but it _does_ cause the label:not([hidden]) selector in the migration
wizard to not select it, which cause the label to not have the right
vertical alignment.

This patch updates the iterator to just select and hide the checkboxes
that have resources associated with them.

Differential Revision: https://phabricator.services.mozilla.com/D176342

UltraBlame original commit: f39eb2f35098fa841b918c434a1c71f76a021131