Bug 1830383 - Implement "Thanks for your feedback" message in Suggest results. r=dao,fluent-reviewers,flod
This builds on D174994 and implements the "thanks for your feedback" message
shown inside a Suggest result when the user gives feedback on it without
dismissing it. This applies to weather suggestions, addons, Pocket, etc. The
implementation is very similar to the implementation of row/group labels.

References:

* [Figma](https://www.figma.com/file/Hdi0oHB7trRcncyVAKZypO/accuweather-explorations?node-id=2421%3A62540&t=svOk7TxQv4V7Y9L4-1) (see "A11y review - user feedback" in sidebar)
* [Clickable prototype](https://www.figma.com/proto/Hdi0oHB7trRcncyVAKZypO/accuweather-explorations?page-id=2192%3A42825&node-id=2394-52468&viewport=246%2C526%2C0.12&scaling=min-zoom&starting-point-node-id=2394%3A52468&show-proto-sidebar=1) (see "Revised 4/3" in sidebar)
* [Content design](https://docs.google.com/document/d/1Mgt_oAIEDz_sF-YBqqUGtfKHQCAtRzEQohpfqk-2X8U/edit?usp=sharing)

Depends on D174994

Differential Revision: https://phabricator.services.mozilla.com/D175729

UltraBlame original commit: 3e03829064fb058b0dcc7191aab45877a7711a00