Bug 1831099 - Remove lock requirement for ImageContainer::mNotifyCompositeListener to avoid deadlock. r=jnicol
In bug 1827024 we added a lock to protect mNotifyCompositeListener but
this turned out to be overzealous and unnecessary as we currently only
ever access the member on the ImageBridge thread. We can just assert
instead to ensure we don't regress on this. Removing the lock
requirement solves a deadlock.

Differential Revision: https://phabricator.services.mozilla.com/D177033

UltraBlame original commit: 3b37c0ddd1eab1e80914678078934a2b529f9467