Bug 1831981 - Share code for expanding all keyword and shorthands when iterating transition-property list. r=emilio
The only difference is the usage of `nsCSSProps::IsEnabled()` for all
keyword when handling the cancellation of existing transitions.

However, this is a kind of preference check so it shouldn't break the
consistency. And in fact we don't have to cancel the no longer used
transitions if we specifies `all` keyword. Besides, I will update this part
again to remove the for loop, in a later patch.

This patch doesn't change behaviors.

Differential Revision: https://phabricator.services.mozilla.com/D175533

UltraBlame original commit: bcbb9e50a0f6662e2f613538baf2a50abeb36fb6