Bug 1830919 - [devtools] Fold addWebConsoleCommands and WebConsoleActor._getWebConsoleCommands into WebConsoleManager.getWebConsoleCommands. r=devtools-reviewers,jdescottes
We were processing the commands in two distinct layers in how we were
going from command objects passed to WebConsoleCommandsManager.register
and the precise object passed to the Debugger APIs (bindings/helpers.sandbox).

This allows to see that setter weren't used and commands can only be:
- a function ('clear()')
- an object with a function as 'set' attribute ('$0')

Also renamed 'sandbox' to 'bindings' everywhere so it is clearer it is the same
object in the whole codebase.
And return only bindings and helperResults as 'owner' object is rather an internal
object of commands.

Differential Revision: https://phabricator.services.mozilla.com/D177143

UltraBlame original commit: 814b6dc3a2aa33d8c3a872d607603c6c74f10a33