Bug 1830274 - Change origins alternative frecency. r=daisuke
Change the origins alternative frecency to be the sum of frecencies for all the
contained pages visited in the last 90 days. This is similar to the default
algorithm, but introduces a stronger recency factor.
This change is not covered by tests, because the alternative algorithm may
change often and dinamically, so we'll add tests when/if it graduates.

The moz_meta key changes to not just track a version number, but an object that
can also track additional variables (the number of days to cut off in this case).
Any change to any of the variables is like changing the algorithm version,
causing a recalculation at the next application startup.

Differential Revision: https://phabricator.services.mozilla.com/D177043

UltraBlame original commit: 88b584b11d1f5f2f55b652594b3e4d3fade80f6d