Bug 1831394 - omit the retries left counter from the invalid PIN prompt when it is large. r=keeler,fluent-reviewers,flod
The PIN for a user's security key will be temporarily blocked after three
failed PIN entry attempts, but the "retries left" counter in the invalid PIN
prompt is the (typically larger) number of attempts left before the PIN is
permanently blocked. This patch makes it so that the retries left counter is
only shown when it is equal to 1, 2, or 3.

Differential Revision: https://phabricator.services.mozilla.com/D177294

UltraBlame original commit: 76aeed342459f29532f973c48fad38855eedeb0f