Bug 1831824 - libwebrtc: In PacketSequencer set timestamps of rtx padding packets when there is no last packet. r=webrtc-reviewers,dbaker
Prior to this patch timestamps are not adjusted when they are 0, and they are 0
when the packet sequencer has not yet seen a media packet. Code and comments in
PacketSequencer and RTPSender::GeneratePadding make it clear that rtx padding
packets are allowed prior to seeing a media packet, and therefore the 0
timestamp case has to be handled.

For rtx the padding packets do not need to have the same timestamp as any media
packets, like plain padding packets do -- because they can only be part of a media
frame, so a media packet has to be known.

With this patch both rtp timestamps and capture timestamps are set to current
time when sequencing rtx padding packets without having seen a media packet.

This fixes a DCHECK failure in TransmissionOffset::Write.

Differential Revision: https://phabricator.services.mozilla.com/D177306

UltraBlame original commit: b0339fd77c82c4c54e5aacaeef66d739b1643827