Bug 1796603 - Reuse existing conditions when all values are subsets of the existing value, r=Sasha
If we have to update some of the metadata for a test (e.g. because one
subtest got different results), for the subtests that didn't change we
want to avoid creating new conditions where possible.

Previously we reused existing conditions in the case that all the new
values for run info matching that condition were the same, but didn't
handle the case that the values are different, but subsets of the
existing value.

This adds additional handling for the subset case, so that when we
have a condition like [TIMEOUT, OK] and various runs give a mix of
those existing values, we just reuse the existing conditional rather
than creating more specific ones.

Differential Revision: https://phabricator.services.mozilla.com/D176003

UltraBlame original commit: 40cf739f51d96b78db70b37ce1620eb60b0ddad6