Bug 1833354 - Fix chroma subsampling in DMABufSurface. r=media-playback-reviewers,padenot,stransky
In the ImportPRIMESurfaceDescriptor method we set the sizes of the
buffers for each plane.  This was done by successively halving the
buffer width and height for each successive plane.  This works for
2-planar formats with 4:2:0 chroma sub-sampling (e.g. NV12) but not for
3-planar 4:2:0 formats (e.g. YUV420).  Update this method to work with
all 4:2:0 formats regardless of number of planes.

Differential Revision: https://phabricator.services.mozilla.com/D178193

UltraBlame original commit: 0f37afc972afa030f95941d62ad79af321979c46