Bug 1817997 - Update TrackBufferManager to use the new TimeUnit. r=alwu
This one tricky bit is when appending frames to the internal buffer. We want to
do all computations in the time base of the container, that's more precise. When
the offset is set by js (as a 64-bits double, converted to integer
microseconds), we convert it to the time base of the track, and accumulate this
way, to have minimal error.

Differential Revision: https://phabricator.services.mozilla.com/D173312

UltraBlame original commit: 412e0a9ec4e8bb99ca4e1eb310083528f2c89449