Bug 1833688. Simplify ComputePagesPerSheetAndPageSizeTransform. r=dholbert
Prior to this change, it wasn't at all clear that the single page-per-sheet
calculations are actually very simple. This makes it a lot clearer. It also
changes the transform code to avoid unnecessary mixing of Pre* and Post*
operations, which was making the code more difficult to reason about.

Separating out this change also makes the changes that will be introduced in
bug 1833239 a lot easier to understand.

Differential Revision: https://phabricator.services.mozilla.com/D178351

UltraBlame original commit: 9031d628fca940c0ac0e951a1302e17c15a2af21