Bug 1833682. Fix n-pages-per-sheet regression, to ensure we properly scale pages beyond the first sheet. r=dholbert
When splitting out the code for bug 1830900 I missed this rather important line.
That bug changed the code to store certain values on a per-sheet basis, but
forgot to include this line to calculate them for all sheets (as opposed to
just the first sheet).

Differential Revision: https://phabricator.services.mozilla.com/D178347

UltraBlame original commit: bf580270836192ec987e4bea1af5cf337c7bcf94