Bug 1830795, part 4 - Implement lifetime checking for content list cache entries. r=nika
Add a new field to nsContentList that should be true if and only if
that list is in one of the hashtable caches.

The entry destructor release assert double checks that any list it contains
knew it was in an entry. This should be superfluous as long as
SetContentList is the only place that sets HashEntry::mContentList to a non-null
value, and nsContentList::mInHashtable to true.

The first SetContentList release assert ensures that the entry starts out empty,
because if it wasn't we'd need to tell the current list that it is being
removed from the cache entry.

The second SetContentList release assert ensures that the list is never in more
than one entry at the same time, which allows us to make mInHashtable a boolean
and not an integer.

The release asserts in the remove-from-hashtable methods ensure that the list
isn't in a hashtable afterwards. These are done as two separate checks to try to
help with debugging if we see these asserts get hit. These methods are called
from the destructor, and other places related to tearing down a content list.

Differential Revision: https://phabricator.services.mozilla.com/D177845

UltraBlame original commit: f1d8476e13e60f7b96074b8a5f9f56e8e2602972