Bug 1834101 - part2 : discard seeking request from the old MDSM. r=jolin
The old MDSM is going to be destroyed and it would reject the seeking
promise, which would result in aborting [1] the current seeking promise
for the media element.

Therefore, we should cancel that seeking request because we would
perform another new seek after recreating the MDSM, which would be the
one actually performing seeking.

[1]
https://searchfox.org/mozilla-central/rev/e409c755f313f6befb2c00f260bf873ccc191924/dom/media/MediaDecoder.cpp#949

Depends on D178556

Differential Revision: https://phabricator.services.mozilla.com/D178557

UltraBlame original commit: 1fe0f07396265808c51d99d47bc692d6ea1a8659