Bug 1833854 - Part 3: Allow min and max nursery sizes to be set regardless of current values r=sfink
Currently we prevent setting a new minimum value that's greater than the
current maximum value and vice versa.

This is problematic because the caller may not know the current values. For
example, they may want to set both min and max to values that are valid with
respect to each other but not with resepect to existing values. (This is caused
by the fact that the API only allows setting one value at a time but that's a
larger issue.)

Instead, handle this the way we do for all other parmeters. When setting a new
value that invalidates an invariant with repsect to another parmeter, adjust
that other parameter to maintain the invariant.

Differential Revision: https://phabricator.services.mozilla.com/D178527

UltraBlame original commit: ce5f0d5ba79c8183f261811f226a5aba0e58ac06