Bug 1829582 [wpt PR 39658] - [css-flex] Disable fixed-table's infinite size for intrinsic flex sizing, a=testonly
Automatic update from web-platform-tests
[css-flex] Disable fixed-table's infinite size for intrinsic flex sizing

A fixed-table's intrinsic max-content is ~infinite when it has a
%-width. That's been around for a while and ships in all browsers.

But it is problematic for the new flex intrinsic sizing algorithm
because now the infinity percolates up to the flex container's
min-content size[1]. We obviously don't want a min-content size of
infinity and gmail relies on the old behavior.

So this is a hacky patch that makes fixed-tables max-content size be the
same as auto-tables when an ancestor flex table is calculating its
intrinsic sizes. It will probably be web compatible, but unsure.

This affects both the min and max intrinsic size of the flex container.

[1] New flex min-content algorithm accounts for item's flex base sizes.
A %-width fixed-table's flex base size is its max-content size, which,
as mentioned above, is ~infinity.

Change-Id: I7010fa73df8e88a4812b23780e5d58a3f63da730
Bug: 240765
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4420422
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: David Grogan <dgroganchromium.org>
Cr-Commit-Position: refs/heads/main{#1135691}

--

wpt-commits: ce59e3b3eeb39099ff2a1f1328e7ecff28a39120
wpt-pr: 39658

UltraBlame original commit: d4d8e34d1139415832f0b332e17519bfc86958b1