Bug 1830459 [wpt PR 39744] - [balance-text] Add a test for `text-indent`, a=testonly
Automatic update from web-platform-tests
[balance-text] Add a test for `text-indent`

This patch adds a test of `text-wrap: balance` for a block
with `text-indent`.

Note the balancing algorithm is not defined in the spec; it's
UA-defined. But the text in this test can be fully balanced
that it's reasonable to expect any algorithms to pass.

This test hits a `DCHECK` which happens when one of lines
overflows. This `DCHECK` is no longer valid, it was checking
applicable block but now it checks overflow, which is valid,
so removed.

Bug: 1251079
Change-Id: I581ec2323ef9a87082fb6d3b4cc788458152bb79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4483399
Auto-Submit: Koji Ishii <kojiichromium.org>
Reviewed-by: Kent Tamura <tkentchromium.org>
Commit-Queue: Kent Tamura <tkentchromium.org>
Cr-Commit-Position: refs/heads/main{#1136957}

--

wpt-commits: 2e69cbe2507d59e4417ee699c7760eaf828dba13
wpt-pr: 39744

UltraBlame original commit: 1ed8e487dd7e4c0b299eead18a925babf4ae6b34