Bug 1793557 - Convert Extension.jsm strings to Fluent. r=mkmelin,geckoview-reviewers,robwu,flod,owlish
This changes the arguments of `ExtensionData.formatPermissionStrings()`.
The second `bundle` arg is dropped, and a `localization` option is added.
Call sites in m-c are updated, but this will also need a matching update for Thunderbird.
A few Thunderbird test cases will also need to be updated,
as they currently point to a non-existing localization file paths
"messenger/addons.ftl" and "messenger/addonPermissions.ftl".

As discussed at the addon workweek,
the Fluent l10n keys for extension permissions match the pattern:

    webext-perms-description-{name}

where `{name}` is the permission's sanitized name.
A fluent-lint exception is added for the capitalization of these generated names.
To allow for message updates and subsequent l10n identifier updates,
a `PERMISSION_L10N_ID_OVERRIDES` map is provided.

Because Fluent localization keys are not enumerable
and attempting to format a missing key is an error,
the `PERMISSIONS_WITH_MESSAGE` set must be kept in sync with message updates.

Differential Revision: https://phabricator.services.mozilla.com/D158663

UltraBlame original commit: 50f952ef9d20328e760305b4ce71cd1f28ecb5e5