Bug 1817997 - Fix some test. r=alwu
Two tests are adjusted to round differently, this doesn't have real
consequences.

For the third test the result should really be 0 (the segments are of the same
duration), but it was a very small number (about 1us in seconds) before because
of rounding.

Differential Revision: https://phabricator.services.mozilla.com/D176036

UltraBlame original commit: 286d9df49c11ce12cffbdd2a9db3b3d154be19f0