Bug 1407461 Part 2: Allow hardware acceleration for Windows drivers with mismatched DLL and registry versions in all channels. r=gfx-reviewers,jrmuizel
The presence of "DriverVersionMismatch" does not seem to be an outlying
indicator in crash reports for Intel 965 Express hardware. This patch
allows hardware WebRender in all channels for that hardware, and we'll
deal with individual driver bugs as they appear.

Differential Revision: https://phabricator.services.mozilla.com/D179147

UltraBlame original commit: f4c42d4a4b295853bf07994819a354d476d5833e