Bug 1826946 - Delay frecency calculation for history batch inserts. r=daisuke
When inserting multiple URIs in history, it makes sense to delay expensive
frecency calculations to a later time. We only set recalc_frecency and the
PlacesFrecencyRecalculator will take care of it.
We keep immediately calculating frecency only when a single url is inserted,
that is the case when the user is normally browsing the Web.

Differential Revision: https://phabricator.services.mozilla.com/D179263

UltraBlame original commit: b34dc8f7157a5c572efbf7d16dabd1f04df768d8