Bug 1833980 [wpt PR 40086] - [wptrunner] Wait for TestRunnerManager threads to exit on interrupt, a=testonly
Automatic update from web-platform-tests
[wptrunner] Wait for TestRunnerManager threads to exit on interrupt (#40086)

When wptrunner is interrupted, the manager group sets its stop flag but
doesn't wait for the runner manager threads to acknowledge and join.
Besides possible resource leakage, this can cause the threads to
continue logging events after the main thread calls `shutdown()` on the
`web-platform-tests` logger, which is not allowed.

In this change, the main thread waits for the threads to join with a
timeout on an interrupt. The timeout needs to be long enough for the
threads to cycle through their event loops one more time, but not so
long as to appear unresponsive.
--

wpt-commits: 4428a332888632faa2f59d0319043e9acbd9afad
wpt-pr: 40086

UltraBlame original commit: 32bc40a983d1fdadcca5bf10c9d34c5b238c62af