Bug 1723195: Allow intervening generic accessibles between table rows and cells, r=Jamie
This commit ensures that we get the proper non-generic parent when creating a
grid cell accessible. Before this patch, we wouldn't see this grid cell as a
grid cell accessible. This patch also adds a test for this case, along with a
test for a similar case that inserts a generic accessible between the table
itself and a rowgroup.

Differential Revision: https://phabricator.services.mozilla.com/D180393

UltraBlame original commit: cb6adbea9cddb2f353604ee6d763690176ab1087