Bug 1829054 clarify the Start() to Stop() lifetime of AudioSinkWrapper::mEndedPromise r=padenot
Move Ensure() to Start() to clarify that muting and SetPlaying() should not
create a different promise for clients to watch.

Move ResolveIfExists() from Shutdown() to Stop() to make the behavior when
muted consistent with behavior when unmuted and with DecodedStream and
VideoSink.

Shutdown() asserts that Stop() has been called.

Depends on D180652

Differential Revision: https://phabricator.services.mozilla.com/D180653

UltraBlame original commit: 2fbfefb8c6cfddd0278b53353298837afd1f7d70