Bug 1829054 use a separate promise to track AudioSink and AudioSinkWrapper ended r=padenot
In a future patch this will allow the AudioSinkWrapper to remain unended when
the AudioSink ends due to the device becoming unavailable, and to reopen
another AudioSink when the device becomes available again.

Re-enqueuing unplayed audio even when not required is favored over
the complexity of an otherwise unnecessary parameter.

Differential Revision: https://phabricator.services.mozilla.com/D180658

UltraBlame original commit: 203bf275e99e99882638a6b3357c0a6a2f97bffc