Bug 1837305 - Part 3: Use PositionOrAuto for "at <position>" in circle and ellipse. r=emilio
The omitted postion has special meaning for offset-path. It defaults to using
offset-position as the circle/ellipse center if the explicit center is
not given. Just like what we do for ray(), here we also use
PositionOrAuto for it.

We omit this component if it is not given when serializing offet-path,
and preserve the value even if it specifies the default value, for the
same reason.

For shape-outside and clip-path, we don't change their behaviors for now.
This bug is for offset-path and we will use it in the following patches.
We should fix them in Bug 1837340.

Differential Revision: https://phabricator.services.mozilla.com/D179626

UltraBlame original commit: bc7a93717f749e25fb241b5316e6c29d21aacd08