Bug 1802193 Part 3: Make a PiP test work with asynchronous fullscreen transitions. r=pip-reviewers,mhowell
This test is checking the state of PiP UI elements in the video window,
and how they change in response to fullscreen. The logic in the test waits
on the fullscreenchange events it receives from the parent process, and
the PiP UI changes depending on the fullscreenchange events it receives
from the content process. The changes in this patch ensure that the PiP UI
elements have time to receive their fullscreenchange events from the
refresh driver. This was already in place for fullscreen exits, presumably
because it was causing problems when transitioning from test-to-test. This
change adds that same pause to fullscreen enters, which gives time for the
UI elements to receive their event and get updated.

Prior to this change, this test was using the pattern of defining async
functions which await other Promises. But the last Promise in the chain
was never awaited. This changes the test to use the pattern of returning
the Promises, which works becauses the calling functions are awaiting
the results of each step in the chain.

It also adds more logging to the test.

Depends on D180884

Differential Revision: https://phabricator.services.mozilla.com/D181123

UltraBlame original commit: 3146a51a44d128118634a7605794b8aeea232d83