Bug 1838731 - Fix scrollbar kind computation for scrollcorner. r=mstange
The scrollcorner isn't a child of a scrollbar so we don't find it.
Instead, check for the scrollable node. This info is only used on macOS,
this fixes the rendering of rtl scrollcorners with non-overlay
scrollbars, e.g.:

  <div style="border: 1px solid; width: 100%; height: 100%; overflow: scroll; direction: rtl"></div>

Or so.

Differential Revision: https://phabricator.services.mozilla.com/D181259

UltraBlame original commit: 57dfc72d508c3d7ac1e157efd4939c22f08d47fd