Bug 1836824 Part 1: Make HTMLCanvasElement re-register its frame refresh driver when a context is updated. r=jgilbert
Since HTMLCanvasElement has a lazily-created context, and the
RequestedFrameRefreshObserver will fail to watch the state when the
context is null, this patch adds a hook to try to re-register when
UpdateContext is called. The superclass calls UpdateContext after the
context is initially created.

Differential Revision: https://phabricator.services.mozilla.com/D181149

UltraBlame original commit: 102bde84aa26492bbc9ced897303364152f908b1