Bug 1820216 - Run Custom chromium-as-release performance tests on OSX. r=perftest-reviewers,sparky
This patch extends the chromium-as-release perftest configuration to be able to also run on the macOSX platform using a macOSX build.
With this, we now run CaR on all of our x64 platforms.

Also, the mac variant as added to the `mach try perf` selector

Differential Revision: https://phabricator.services.mozilla.com/D180825

UltraBlame original commit: 1f03620644bd5273dfdfe64e0da9d9766361e4ed