Bug 1839062 - Check whether nursery is enabled before requesting minor GC during JIT BigInt allocation r=jandem
I previously added an assertion to Nursery::requestMinorGC that the nursery is
enabled and the fuzzers have found another case where it isn't true. When
nursery big int allocation fails in JIT code we request a minor GC, but the
nursery could be disabled so we have to add this check.

Differential Revision: https://phabricator.services.mozilla.com/D181362

UltraBlame original commit: 41581deb338a0f524c02f374390e58c71f6656ec