Bug 1802340 - Add a shim for tsn.ca to address login issue with dFPI. r=anti-tracking-reviewers,webcompat-reviewers,twisniewski,pbz
Adding a shim that calls privilege StorageAccessAPI when tsn.ca using
the bellmedia.ca to perform the login.

We update the matching rule for the content script of crave.ca because it
also uses the bellmedia to do the login. With the updated matching rule,
we will only load the relevant shim script depending on the query string
of the login page.

Differential Revision: https://phabricator.services.mozilla.com/D180438

UltraBlame original commit: 5ade2b37847325019878bde832493c342759d7eb