Bug 1839534 - Return early from FontPrefs::Init() when we're late in shutdown. r=jfkthame
If we're initializing FontPrefs this late, we're probably in the middle
of shutting down the XPCOM component manager, and the preference service
will not be available.

Technically there's a later phase, CCPostLastCycleCollection, before we
actually do the component shutdown, but both of these phases are so late
(for instance, XPCOM threads no longer exist) that we should really not be
starting up any font stuff.

The goal of this change is to avoid a very common warning in debug builds.

Differential Revision: https://phabricator.services.mozilla.com/D181586

UltraBlame original commit: bdd30255916a63580ad763a599075077749c1a17