Bug 1719855 - Use CreateSingleTouchData rather than using a ctor of SingleTouchData for the first touch-move event. r=botond
The ctor doesn't set mScreenPoint so that we fail to detect bailing out from
in-slop state state, thus during processing the touch-move event we drop the
event [1].

The reason why the gtest has passed is that the corresponding touch block in
question is for a long-tap event instead of touch-start/move/end events it
never becomes in-slop state. In a subsequent change, we differentiate the
touch block for a long-tap from the block for touch-start/move/end events.

[1] https://searchfox.org/mozilla-central/rev/504bd0a7ee3110ecb2b31f89b9ce54faf76e228b/gfx/layers/apz/src/InputQueue.cpp#190-193

Depends on D176475

Differential Revision: https://phabricator.services.mozilla.com/D176476

UltraBlame original commit: f161eecb0f2cefa57b9e6e37666f03f6e67f6969