Bug 1838214 - Include the log level in mach human/terminal log output r=ahochheiden,ahal
This includes the log level as a single character (with a space, adding 2 total characters to log
lines). It is colored when there is a terminal present. The info log level has no added characters
(it being the most common). There is an additional argument to the log formatters to enable showing
the log level, and it defaults to the same value as the argument to show the timestamps.

Differential Revision: https://phabricator.services.mozilla.com/D180829

UltraBlame original commit: 3be43da6c87b64ad5d7f6e0ddce6678062e7c65d