Bug 1840018 - Re-enable the tests of composition event debug break points which were disabled in bug 1784847 r=bomsy,devtools-reviewers
The crash bug is now fixed in bug 1690827.  Therefore, we can re-enable the
tests.

Unfortunately, I observed bug 1785014 in the tryserver with running a lot of
TV tasks.  But it seems that the crash rate is acceptable.  If it's not let's
back this out again.

Differential Revision: https://phabricator.services.mozilla.com/D181868

UltraBlame original commit: db6d82935c97ebb7d8c55a64956977cb204d3b99