Bug 1829068 use a TaskQueue for background stream initialization r=padenot
The runnables will need to be run serially in a future patch when the sink will
not necessarily be the same for each runnable.

This will also allow us to use InvokeAsync() and Then() in a future patch for
easier promise management.

Depends on D181971

Differential Revision: https://phabricator.services.mozilla.com/D181972

UltraBlame original commit: 690eb6585422ed5e53aee4f1cf3beed49f7798e9