Bug 1839265 - Make the `animationend` event listener do not assume that it's always fired for `<dir>` r=m_kato
According to MDN, `animationend` event is never fired if the event target (`<dir>` in this case)
is removed.  If it's slow environment, `<dir>` element may be removed before the listener receives
the event for it.  Therefore, it should not assume that it always comes.

Differential Revision: https://phabricator.services.mozilla.com/D181961

UltraBlame original commit: 8cc85121d9b7c49b9def06d8556709ce75354933