Bug 1840116 - [devtools] Fix modifying selector of child rule via CSSOM. r=devtools-reviewers,ochameau.
When updating a selector in a constructed stylesheet,
we looked through the stylesheet cssRules to properly
insert the new rule at the right position.
But for nested rules, we'd never find them, and the
modification of the selector wouldn't happen.

Differential Revision: https://phabricator.services.mozilla.com/D181956

UltraBlame original commit: b332adf20b98690987c6f2a7ef5021ee832dfb52