Bug 1834815 - Enable full+cross LTO by default for shippable Linux and Windows build r=glandium
This has an important impact on compilation time, so we limit that to
shippable builds and under MOZ_AUTOMATION.

The benefit on speedometer time is not significant on macosx and android builds, so
we skip that platform.

Builds tend to stall on arm platforms, so we limit the target to
x86_64.

Differential Revision: https://phabricator.services.mozilla.com/D179912

UltraBlame original commit: 502f57bb0ebbce52ba15616a4aa213f2478f8e1f