Bug 1837272 - patch 1 - Ensure the 8-bit version of nsLineBreaker::AppendText clears the word-continuation flag on encountering space. r=emilio
The issue here was that the 8-bit version of AppendText failed to reset the flag on space.
Where text-transform:capitalize is applied to a multi-word chunk of text, this wasn't an issue
because AppendText delegates to the more comprehensive 16-bit version whenever capitalization
is requested, so only that version cares about the word-continuation flag.

However, in this case the whitespace *between* the spans does not itself have the capitalization
property set, and so it doesn't go via the 16-bit function, and we failed to clear the flag.

Differential Revision: https://phabricator.services.mozilla.com/D183251

UltraBlame original commit: 9733569bfd85647324f2d2df9a5f833f2760e04d