Bug 1839598 - part 1: move WasmStructObject::createStruct out of WasmGcObject.h.  r=rhunt.
Currently there is, in wasm/WasmGcObject.h, the complete implementation of
WasmStructObject::createStruct.  This is in the .h file so that it can be
inlined into its caller, Instance::structNew.  This ad-hoc-ery is a hindrance
for the patches that follow, so this first patch pushes the implementation of
WasmStructObject::createStruct back into WasmGcObject.cpp.

With this patch applied, the header now contains just the declaration for

  template <bool ZeroFields> static WasmStructObject* createStruct(..)

and the .cpp file contains the implementation, in the standard way.

Differential Revision: https://phabricator.services.mozilla.com/D182147

UltraBlame original commit: 2c942bc1cb671201fc124342703fe51e4caf5eba