Bug 1842007 - Part 7: Remove MozPromise::FromDomPromise r=xpcom-reviewers,necko-reviewers,nika,jesup
Nobody uses it and currently not super useful as the possible return value is limited to primitives. It could be extended to support DOM objects, but I don't know there's actual use cases for that as in many cases the promise is resolved with non-DOM values.

Differential Revision: https://phabricator.services.mozilla.com/D183026

UltraBlame original commit: 9b4ba013fc3171cb68e03372f642519a7f67157c