Bug 1805514 - Part 2: Compare with structured comparison rather than string comparison. r=nrishel
We have gotten a great deal of mileage comparing Windows toast XML as
strings.  However, we intend to start including JSON in the
`arguments` blocks round-tripping through the notification server, and
the encoding of JSON first into strings and then into XML is extremely
painful to read.

This ugly commit rewrites the XML to be simpler, parsing the launch
and action arguments into data structures.  The simplified XML is
compared as a string while the arguments are compared as data
structures.  In the future, we might parse entirely into a data
structure for comparison.

Differential Revision: https://phabricator.services.mozilla.com/D182313

UltraBlame original commit: 6397f085a6b0c780d7d3e0a38e8cdc534b30cd9a