Bug 1762225 - Introduce BackgroundContextOwner r=rpl
Before this patch, ext-background.js assumed that it fully controlled
the creation and termination of the background context. This assumption
is too optimistic, because the background can be unloaded externally.

This patch introduces the BackgroundContextOwner class, which owns the
background's ExtensionPageContextParent or BackgroundWorkerContextParent
instance associated with an actual background. A BackgroundContextOwner
is instantiated at most once per extension and is responsible for
tracking the active background context (to expose externally).

In follow-up patches, relevant logic will gradually be moved to
BackgroundContextOwner to improve the overall reliability of background
lifetime tracking.

Differential Revision: https://phabricator.services.mozilla.com/D183214

UltraBlame original commit: 432063ff43c75c16b03630b03da39d2579808684