Bug 1838080: Work around a race in ChannelSendFrameTransformerDelegate. r=pehrsons,webrtc-reviewers
This variable can be null when a ChannelSendFrameTransformerDelegate is in use,
because that does an async dispatch to the encoder queue in the handling for
transformed frames. If this is unset while that dispatch is in flight, we
nullptr crash.

Differential Revision: https://phabricator.services.mozilla.com/D180735

UltraBlame original commit: 56555ecee7f36ae73abff1cbbd06807c2b65fc19