Bug 1842496 - only set `mNextSample` when the next sample is available. r=media-playback-reviewers,padenot
The next sample seems not always be available after seek so it
doesn't make sense to set a nullptr to `mNextSample`. But it's still not
clear to me why we won't be able to get a sample if the seek time is
already in the buffered range [1].

We should be able to solve the crash by not storing the nullptr in
`mNexSample` which would ensure we always call `mManager->GetSample` to
check whether thereis available sample.

[1] https://searchfox.org/mozilla-central/rev/4537bb2abc8eb49506bbad27b89ab364ac3062d3/dom/media/mediasource/MediaSourceDemuxer.cpp#380-392

Differential Revision: https://phabricator.services.mozilla.com/D184050

UltraBlame original commit: c28c9108aae146f4608dbac9d8c43c272a64cd9d