Bug 1841266 - part 1: cache required array element size in TypeDefData.  r=rhunt.
This simple patch caches the element size for array types in struct TypeDefData,
so that the allocator doesn't need to chase from TypeDefData to TypeDef to
find the size. Hence it removes a dependent load in the allocation path, but
more importantly removes the need for allocation of arrays to look at the
TypeDef at all.

This patch is very similar to a slightly earlier one which did the same thing
for struct sizes.

Differential Revision: https://phabricator.services.mozilla.com/D183246

UltraBlame original commit: 31801f9bf3c459d253620130ebb7ca539beae7b9