Bug 1845352 - Remove "quit-application-granted" call from AddonTestUtils.promiseShutdownManager r=rpl
The "quit-application-granted" notification was originally part of
AddonTestUtils.promiseShutdownManager because XPIProvider.jsm used to
subscribe to that notification to determine whether to shut down.

A refactor in bug 1461145 replaced the notification with the use of
`AsyncShutdown.quitApplicationGranted`, which is already covered by
`MockAsyncShutdown.quitApplicationGranted`. To simulate add-on manager
shutdown, it is therefore unnecessary to trigger the notification.

Worse, the dispatch of the notification causes other parts of Firefox to
commence shutdown, with unexpected side effects. For example, the linked
bug describes unexpected extension process termination in the middle of
a xpcshell test:

- Due to bug 1845778, it was possible for 2 concurrent extension
  processes to exist. And when the extension document hosted in the
  second extension process unloaded, that process would be shut down
  because `dom.ipc.keepProcessesAlive.extension=1` in tests (not 2).

- This in its turn triggers the "ipc:content-shutdown" notification that
  gets translated to "extension-process-crash". The implementation here
  assumes that there is only one extension process, so the notification
  causes ext-backgroundPage.js to *sometimes* interpret the termination
  as a process crash (since bug 1762225).

- On Android, the process exit is detected as a process crash, because
  "intentional" extension process shutdowns are sometimes flagged as
  abnormal shutdowns. This could happen because ContentParent.cpp has
  Android-specific logic to eagerly kill content processes to minimize
  the chance of having a dead process being around. In contrast, the
  non-Android implementation is more patient (bug 1668376).

Differential Revision: https://phabricator.services.mozilla.com/D184754

UltraBlame original commit: 10ecf0a64753a115a0e570e4ce8c900a669528ad